### 2023-06-14 #5557 @editorialbot
[PRE REVIEW]: MacroModelling.jl: A Julia package for developing and solving dynamic stochastic general equilibrium models
**Submitting author:** @thorek1 (<a href="http://orcid.org/0000-0002-0068-1809">Thore Kockerols</a>)
**Repository:** https://github.com/thorek1/MacroModelling.jl
**Branch with paper.md** (empty if default branch): joss
**Version:** v.0.1.23
