+++ test/test_higher_moments.jl
@@ -2,7 +2,26 @@ using MacroModelling
 import IterativeSolvers as ℐ
 using LinearAlgebra, LinearMaps
 
+function intersecting_elements(arr1::Union{Symbol,Vector}, arr2::Union{Symbol,Vector})
+    common = []
+    
+    if arr1 isa Symbol
+        arr1 = [arr1]
+    end
+    if arr2 isa Symbol
+        arr2 = [arr2]
+    else
+        arr2 = copy(arr2)
+    end
 
+    for i in arr1
+        if i in arr2
+            push!(common, i)
+            deleteat!(arr2, findfirst(==(i), arr2))  # remove the matched element from arr2
+        end
+    end
+    return common
+end
 
 function position_in_symmetric_matrix(position::Int, length::Int)
     # Check if the vector length is a triangular number
@@ -85,13 +104,46 @@ end
 # vec(AA)
 # position_in_full_vector(1,6)
 
-function upper_triangle(mat::AbstractMatrix{T}) where T
+function upper_triangle(mat::AbstractArray{T,2}) where T
     @assert size(mat, 1) == size(mat, 2) "The input matrix must be square"
 
     upper_elems = T[]
     for i in 1:size(mat, 1)
         for j in i:size(mat, 2)
             push!(upper_elems, mat[i, j])
+        end
+    end
+    return upper_elems
+end
+
+
+function upper_triangle(mat::AbstractArray{T,3}; alt::Bool = false, triple::Bool = false) where T
+    @assert size(mat, 1) == size(mat, 2) "The input matrix must be square"
+
+    upper_elems = T[]
+    if alt
+        for i in 1:size(mat, 1)
+            for j in 1:size(mat, 2)
+                for k in j:size(mat, 3)
+                    push!(upper_elems, mat[i, j, k])
+                end
+            end
+        end
+    elseif triple
+        for i in 1:size(mat, 1)
+            for j in i:size(mat, 2)
+                for k in j:size(mat, 3)
+                    push!(upper_elems, mat[i, j, k])
+                end
+            end
+        end
+    else
+        for j in 1:size(mat, 2)
+            for k in j:size(mat, 3)
+                for i in 1:size(mat, 1)
+                    push!(upper_elems, mat[i, j, k])
+                end
+            end
         end
     end
     return upper_elems
@@ -712,6 +764,577 @@ for i=1:nx^2
 end
 
 P = kron(I(nx),U)
+
+
+using Statistics, LinearAlgebra, StatsBase
+
+# check distributional properties by simulating
+shocks = randn(2,100000)
+sim = get_irf(m, shocks = shocks, periods = 0, levels = true, algorithm = :pruned_second_order, initial_state = collect(get_SS(m, derivatives=false)))
+
+[mean(i) for i in eachrow(sim[:,:,1])]
+[Statistics.var(i) for i in eachrow(sim[:,:,1])]
[…]
