### 2023-06-18 152de44
managed 3rd order variance
