+++ test/test_higher_moments.jl
@@ -1105,16 +1105,34 @@ for (i¹,s¹) in enumerate(inputs)
 
             if any([k ∈ m.timings.exo && v == 1 for (k,v) in combo_cnts])
                 continue
-            elseif all([k ∈ m.timings.exo && v == 2 for (k,v) in combo_cnts]) && all([k ∈ m.timings.exo && v == 1 for (k,v) in intrsct_cnts]) && length(intrsct_cnts) > 0
+            elseif all([k ∈ m.timings.exo && v == 2 for (k,v) in combo_cnts]) && 
+                all([k ∈ m.timings.exo && v == 1 for (k,v) in intrsct_cnts]) && 
+                length(intrsct_cnts) > 0
+
                 Γ₂[i¹,i²] = 1
-            elseif all([k ∈ m.timings.exo && v == 4 for (k,v) in combo_cnts]) && all([k ∈ m.timings.exo && v == 2 for (k,v) in intrsct_cnts])
+
+            elseif all([k ∈ m.timings.exo && v == 4 for (k,v) in combo_cnts]) && 
+                all([k ∈ m.timings.exo && v == 2 for (k,v) in intrsct_cnts])
+
                 Γ₂[i¹,i²] = 2
-            elseif length(setdiff(keys(combo_cnts),m.timings.exo)) == 0 && length(intrsct_cnts) > 0 && all([intrsct_cnts[i] > 0 for i in collect(intersect(keys(combo_cnts),keys(intrsct_cnts)))]) && any([combo_cnts[i] == 4 for i in collect(intersect(keys(combo_cnts),keys(intrsct_cnts)))])
+
+            elseif length(setdiff(keys(combo_cnts),m.timings.exo)) == 0 && 
+                length(intrsct_cnts) > 0 && 
+                all([intrsct_cnts[i] > 0 for i in collect(intersect(keys(combo_cnts),keys(intrsct_cnts)))]) && 
+                any([combo_cnts[i] == 4 for i in collect(intersect(keys(combo_cnts),keys(intrsct_cnts)))])
+
                 Γ₂[i¹,i²] = 3
-            elseif all([k ∈ m.timings.exo && v == 6 for (k,v) in combo_cnts]) && all([k ∈ m.timings.exo && v == 3 for (k,v) in intrsct_cnts])
+
+            elseif all([k ∈ m.timings.exo && v == 6 for (k,v) in combo_cnts]) && 
+                all([k ∈ m.timings.exo && v == 3 for (k,v) in intrsct_cnts])
+
                 Γ₂[i¹,i²] = 15
-            # elseif i¹ > n_shocks && i¹ <= n_shocks + nx * nu && i² > n_shocks && i² <= n_shocks + nx * nu && 
-            elseif length(filter(((j,u),) -> j ∈ m.timings.exo, combo_cnts)) > 0 && sum(values(filter(((j,u),) -> !(j ∈ m.timings.exo), combo_cnts))) == 2 && all([i[2] == 1 for i in keys(filter(((j,u),) -> !(j ∈ m.timings.exo), combo_cnts))]) && all([k ∈ m.timings.exo && v >= 1 for (k,v) in  filter(((j,u),) -> j ∈ m.timings.exo, intrsct_cnts)])
+                
+            elseif length(filter(((j,u),) -> j ∈ m.timings.exo, combo_cnts)) > 0 && 
+                sum(values(filter(((j,u),) -> !(j ∈ m.timings.exo), combo_cnts))) == 2 && 
+                all([i[2] == 1 for i in keys(filter(((j,u),) -> !(j ∈ m.timings.exo), combo_cnts))]) && 
[…]
