+++ test/test_higher_moments.jl
@@ -196,9 +196,9 @@ function duplication(p)
 
     j = Int.(vec(a))
 
-    m = Int(p*(p+1)/2)
+    mm = Int(p*(p+1)/2)
     
-    DP = zeros(p*p,m)
+    DP = zeros(p*p,mm)
 
     for r in 1:size(DP,1)
         DP[r, j[r]] = 1
@@ -208,6 +208,53 @@ function duplication(p)
 
     return DP, DPinv
 end
+
+
+function triplication(p)
+    TP = zeros(Int,p^3, Int(p*(p+1)*(p+2)/6))
+    # TPinv = zeros(Int(p*(p+1)*(p+2)/6), Int(p*(p+1)*(p+2)/6))
+
+    for k=1:p
+        for j=k:p
+            for i=j:p
+                idx = unique([[i, j, k],
+                            [i, k, j],
+                            [j, k, i],
+                            [j, i, k],
+                            [k, j, i],
+                            [k, i, j]])
+                for r in idx        
+                    ii = r[1]
+                    jj = r[2]
+                    kk = r[3]
+
+                    n = ii + (jj-1)*p + (kk-1)*p^2
+                    mm = Int(i+(j-1)*p + 1/2*(k-1)*p^2 - 1/2*j*(j-1) + 1/6*k*(k-1)*(k-2) - 1/2*(k-1)^2*p)
+                    
+                    TP[n,mm] = 1
+
+                    # if i==j && j==k
+                    #     TPinv[m,n] = 1
+                    # elseif i>j && j==k
+                    #     TPinv[m,n] = 1/3
+                    # elseif i==j && j>k
+                    #     TPinv[m,n] = 1/3
+                    # elseif i>j && j>k
+                    #     TPinv[m,n] = 1/6
+                    # end
+
+                end
+                n=n+1
+            end
+        end
+    end
+
+    return TP
+end
+# triplication(2)
+# end
+
+
 
 # translate_mod_file("/Users/thorekockerols/Downloads/ReplicationDSGEHOS-main/AnSchorfheide_Gaussian.mod")
 # include("/Users/thorekockerols/Downloads/ReplicationDSGEHOS-main/AnSchorfheide_Gaussian.jl")
@@ -809,8 +856,6 @@ diag(reshape(C3y0,5,5,5))
 
 # transition to third order pruned solution
 
-
-
 varobs = [:YGR, :INFL, :INT]
 T = m.timings
 states = m.timings.past_not_future_and_mixed
@@ -1029,39 +1074,79 @@ Ey   = ybar + C*Ez+d; # recall y = yss + C*z + d
 # size of input vector
 n_entries = Int(nu + nu*(nu+1)/2 + nx*nu)
 nz = 3*nx + 2*nx^2 +nx^3
-nxi = nu+nu^2+2*nx*nu+2*nx*nu+3*nx^2*nu+3*nu^2*nx+nu^3
+nxi = nu + nu^2 + 2*nx*nu + 2*nx*nu + 3*nx^2*nu + 3*nu^2*nx + nu^3
 nu2 = nu*(nu+1)/2 |> Int
 nx2 = nx*(nx+1)/2 |> Int
 nu3 = nu2*(nu+2)/3 |> Int
 nximin = nu + nu2 + 2*nu*nx + nu*nx2 + nu2*nx + nu3
[…]
