+++ test/test_higher_moments.jl
@@ -1476,10 +1476,56 @@ s¹ = inputs[3]
 s² = inputs[4]
 s³ = inputs[4]
 
-# no
-s¹ = inputs[28]
-s² = inputs[28]
+# 1
+s¹ = inputs[4]
+s² = inputs[2]
+s³ = inputs[1]
+# 8
+s¹ = inputs[3]
+s² = inputs[3]
 s³ = inputs[3]
+# 3
+s¹ = inputs[1]
+s² = inputs[4]
+s³ = inputs[27]
+# 12
+s¹ = inputs[5]
+s² = inputs[2]
+s³ = inputs[29]
+
+s¹ = inputs[5]
+s² = inputs[27]
+s³ = inputs[29]
+# no
+s¹ = inputs[5]
+s² = inputs[26]
+s³ = inputs[26]
+
+s¹ = inputs[27]
+s² = inputs[27]
+s³ = inputs[27]
+
+# 6
+s¹ = inputs[3]
+s² = inputs[28]
+s³ = inputs[28]
+
+s¹ = inputs[5]
+s² = inputs[26]
+s³ = inputs[28]
+
+s¹ = inputs[3]
+s² = inputs[28]
+s³ = inputs[28]
+
+# 9
+s¹ = inputs[4]
+s² = inputs[26]
+s³ = inputs[29]
+
+s¹ = inputs[4]
+s² = inputs[27]
+s³ = inputs[28]
 
 
 combo = [(s¹ isa Symbol ? [s¹] : s¹)... , (s² isa Symbol ? [s²] : s²)..., (s³ isa Symbol ? [s³] : s³)...]
@@ -1502,23 +1548,13 @@ push!(S, s³ isa Symbol ? [s³] : s³)
 
 s = Dict([(element,v) => count(==(element),i) for (v,i) in enumerate(S) for element in unique(i)])
 
-(length(filter(((j,u),) -> u == 2, s)) == 2  || length(filter(((j,u),) -> (u == 3 && j[2] == 1) || (u == 2 && j[2] >= 1) || (u == 1 && j[2] >= 2), intrsct_cnts_type)) > 0) && 
-(length(s) <= 4 || sum(values(s)) % 2 == 0) && 
-!all(values(s) .== 2) &&
-length(filter(((j,u),) -> j ∈ m.timings.exo && u == 4, combo_cnts)) > 0 &&
-sum([i[2] for i in s]) % 2 == 0
 
-# length(setdiff(Set([i[1] for i in keys(intrsct_cnts)]), keys(filter(((j,u),) -> j ∈ m.timings.exo && u == 4, combo_cnts)))) == 0 &&
-# length([i[1] for i in keys(intrsct_cnts)]) > 1
-# keys(combo_cnts)
-
-length(filter(((j,u),) -> u ∉ [1,2], s)) == 0 &&
-(length(filter(((j,u),) -> u == 2, s)) == 2 || length(filter(((j,u),) -> (u == 3 && j[2] == 1) || (u == 2 && j[2] >= 1) || (u == 1 && j[2] >= 2), intrsct_cnts_type)) > 0) && 
-length(filter(((j,u),) -> j ∈ m.timings.exo && u == 4, combo_cnts)) > 0
-
-
-
-max(length(filter(((j,u),) -> u == 2, s)), length(intrsct_cnts_type))
+(length(filter(((j,u),) -> u == 2, s)) == 2  || length(filter(((j,u),) -> (u == 3 && j[2] == 1) || (u == 2 && j[2] >= 1) || (u == 1 && j[2] >= 2), intrsct_cnts_type)) > 0 || length(filter(((j,u),) -> (u == 1 && j[2] == 1), intrsct_cnts_type)) >= 2) && 
[…]
