### 2023-06-19 4e03cb0
got the numbers (from shocks)
