+++ test/test_higher_moments.jl
@@ -1421,124 +1421,84 @@ C2y0 = C * C2z0 * C' + DFxi * Γ₂ * DFxi'
 
 Γ₃ = zeros(nximin, nximin, nximin);
 
-#no
-s¹ = inputs[26]
+# covars
+s¹ = inputs[1]
+s² = inputs[3]
+s³ = inputs[10]
+
+s¹ = inputs[2]
 s² = inputs[5]
-s³ = inputs[1]
+s³ = inputs[11]
 
-s¹ = inputs[4]
-s² = inputs[2]
-s³ = inputs[1]
-
-s¹ = inputs[6]
-s² = inputs[3]
-s³ = inputs[1]
-
-s¹ = inputs[26]
-s² = inputs[3]
-s³ = inputs[1]
-
-s¹ = inputs[29]
-s² = inputs[4]
-s³ = inputs[1]
+s¹ = inputs[28]
+s² = inputs[5]
+s³ = inputs[10]
 
 s¹ = inputs[27]
-s² = inputs[4]
-s³ = inputs[1]
-
+s² = inputs[5]
+s³ = inputs[11]
 
 s¹ = inputs[1]
-s² = inputs[1]
-s³ = inputs[1]
-
-s¹ = inputs[3]
-s² = inputs[3]
-s³ = inputs[1]
-
-s¹ = inputs[3]
-s² = inputs[5]
-s³ = inputs[5]
-
-#yes
-s¹ = inputs[28]
-s² = inputs[5]
-s³ = inputs[1]
-
-s¹ = inputs[28]
-s² = inputs[3]
-s³ = inputs[1]
-
-s¹ = inputs[3]
-s² = inputs[1]
-s³ = inputs[1]
-
-s¹ = inputs[3]
 s² = inputs[4]
+s³ = inputs[11]
+
+s¹ = inputs[1]
+s² = inputs[14]
+s³ = inputs[3]
+
+s¹ = inputs[1]
+s² = inputs[16]
+s³ = inputs[3]
+
+s¹ = inputs[1]
+s² = inputs[15]
 s³ = inputs[4]
 
-# 1
-s¹ = inputs[4]
-s² = inputs[2]
-s³ = inputs[1]
-# 8
-s¹ = inputs[3]
-s² = inputs[3]
-s³ = inputs[3]
-# 3
 s¹ = inputs[1]
-s² = inputs[4]
-s³ = inputs[27]
-# 12
-s¹ = inputs[5]
-s² = inputs[2]
-s³ = inputs[29]
+s² = inputs[17]
+s³ = inputs[4]
 
-s¹ = inputs[5]
-s² = inputs[27]
-s³ = inputs[29]
+s¹ = inputs[1]
+s² = inputs[25]
+s³ = inputs[6]
 # no
-s¹ = inputs[5]
-s² = inputs[26]
-s³ = inputs[26]
+s¹ = inputs[1]
+s² = inputs[11]
+s³ = inputs[4]
 
-s¹ = inputs[27]
-s² = inputs[27]
-s³ = inputs[27]
+s¹ = inputs[1]
+s² = inputs[5]
+s³ = inputs[10]
 
-# 6
-s¹ = inputs[3]
-s² = inputs[28]
-s³ = inputs[28]
 
-s¹ = inputs[5]
-s² = inputs[26]
-s³ = inputs[28]
+s¹ = inputs[26]
+s² = inputs[5]
+s³ = inputs[10]
 
-s¹ = inputs[3]
-s² = inputs[28]
-s³ = inputs[28]
+s¹ = inputs[1]
+s² = inputs[16]
+s³ = inputs[1]
 
-# 9
-s¹ = inputs[4]
-s² = inputs[26]
[…]
