### 2023-06-21 747c305
redid shocks with new logic
