+++ test/test_higher_moments.jl
@@ -1149,7 +1149,7 @@ Fxi[row13_u_u_u,col7_u_u_u] = TPu
 Γ₂ = spzeros(nximin, nximin)
 
 
-u_u = reshape([[i, k] for k in m.timings.exo for i in m.timings.exo],nu,nu)
+u_u = reshape([[(i, i == k ? 1 : 0), (k, i == k ? 1 : 0)] for k in m.timings.exo for i in m.timings.exo],nu,nu)
 
 xf_u = reshape([[i, (k,1)] for k in m.timings.past_not_future_and_mixed for i in m.timings.exo],nx,nu)
 # u_xf = reshape([(i, (k,1)) for i in m.timings.exo for k in m.timings.past_not_future_and_mixed],nu,nx)
@@ -1498,34 +1498,56 @@ s³ = inputs[25]
 
 s¹ = inputs[1]
 s² = inputs[5]
-s³ = inputs[16]
+s³ = inputs[26]
 
-combo = [(s¹ isa Symbol ? [s¹] : s¹)... , (s² isa Symbol ? [s²] : s²)..., (s³ isa Symbol ? [s³] : s³)...]
-combo_cnts = Dict([element => count(==(element),combo) for element in unique(combo)])
+s¹ = inputs[1]
+s² = inputs[3]
+s³ = inputs[26]
 
-intrsct = []
-push!(intrsct, intersecting_elements(s¹,s²))
-push!(intrsct, intersecting_elements(s¹,s³))
-push!(intrsct, intersecting_elements(s²,s³))
+s¹ = inputs[4]
+s² = inputs[28]
+s³ = inputs[29]
 
-intrsct_unique_cnts = Dict([(element,v) => count(==(element),i) for (v,i) in enumerate(intrsct) for element in unique(i)])
-intrsct_cnts_raw = [element => count(==(element),i) for (v,i) in enumerate(intrsct) for element in unique(i)]
-intrsct_cnts = Dict(intrsct_cnts_raw)
-intrsct_cnts_type = Dict([(element[1],element[2]) => count(==(element),intrsct_cnts_raw) for element in unique(intrsct_cnts_raw)])
+ϵ²_terms = filter(u -> all([(v isa Symbol ? 0 : v[1]) ∈ m.timings.exo for v in u]) && u[1][2] == 1,[(s¹ isa Symbol ? [s¹] : s¹), (s² isa Symbol ? [s²] : s²), (s³ isa Symbol ? [s³] : s³)])
 
-S = []
-push!(S, s¹ isa Symbol ? [s¹] : s¹)
-push!(S, s² isa Symbol ? [s²] : s²)
-push!(S, s³ isa Symbol ? [s³] : s³)
+terms = filter(u -> !(all([(v isa Symbol ? 0 : v[1]) ∈ m.timings.exo for v in u]) && u[1][2] == 1),[(s¹ isa Symbol ? [s¹] : s¹), (s² isa Symbol ? [s²] : s²), (s³ isa Symbol ? [s³] : s³)])
 
-s = Dict([(element,v) => count(==(element),i) for (v,i) in enumerate(S) for element in unique(i)])
+if length(ϵ²_terms) > 0
+    ϵ²_combo = [c isa Symbol ? c : c[1] ∈ m.timings.exo ? c[1] : c for c in reduce(vcat,ϵ²_terms)]
+    ϵ²_combo_cnts = Dict([element => count(==(element),ϵ²_combo) for element in unique(ϵ²_combo)])
+else
+    ϵ²_combo = []
+    ϵ²_combo_cnts = []
+end
 
[…]
