+++ test/test_higher_moments.jl
@@ -1421,92 +1421,10 @@ C2y0 = C * C2z0 * C' + DFxi * Γ₂ * DFxi'
 
 Γ₃ = zeros(nximin, nximin, nximin);
 
-# covars
-s¹ = inputs[1]
-s² = inputs[3]
-s³ = inputs[10]
 
-s¹ = inputs[2]
-s² = inputs[5]
-s³ = inputs[11]
-
-s¹ = inputs[28]
-s² = inputs[5]
-s³ = inputs[10]
-
-s¹ = inputs[27]
-s² = inputs[5]
-s³ = inputs[11]
-
-s¹ = inputs[1]
-s² = inputs[4]
-s³ = inputs[11]
-
-s¹ = inputs[1]
-s² = inputs[14]
-s³ = inputs[3]
-
-s¹ = inputs[1]
-s² = inputs[16]
-s³ = inputs[3]
-
-s¹ = inputs[1]
-s² = inputs[15]
-s³ = inputs[4]
-
-s¹ = inputs[1]
-s² = inputs[17]
-s³ = inputs[4]
-
-s¹ = inputs[1]
-s² = inputs[25]
-s³ = inputs[6]
-# no
-s¹ = inputs[1]
-s² = inputs[11]
-s³ = inputs[4]
-
-s¹ = inputs[1]
-s² = inputs[5]
-s³ = inputs[10]
-
-
-s¹ = inputs[26]
-s² = inputs[5]
-s³ = inputs[10]
-
-s¹ = inputs[1]
-s² = inputs[16]
-s³ = inputs[1]
-
-s¹ = inputs[1]
-s² = inputs[14]
-s³ = inputs[5]
-
-
-# s¹ = inputs[1]
-# s² = inputs[14]
-# s³ = inputs[3]
-
-s¹ = inputs[1]
-s² = inputs[17]
-s³ = inputs[4]
-
-s¹ = inputs[1]
-s² = inputs[6]
-s³ = inputs[25]
-
-s¹ = inputs[1]
-s² = inputs[5]
-s³ = inputs[26]
-
-s¹ = inputs[1]
-s² = inputs[3]
-s³ = inputs[16]
-
-s¹ = inputs[1]
-s² = inputs[6]
-s³ = inputs[20]
+s¹ = inputs[6]
+s² = inputs[26]
+s³ = inputs[23]
 
 ϵ²_terms = filter(u -> all([(v isa Symbol ? 0 : v[1]) ∈ m.timings.exo for v in u]) && u[1][2] == 1,[(s¹ isa Symbol ? [s¹] : s¹), (s² isa Symbol ? [s²] : s²), (s³ isa Symbol ? [s³] : s³)])
 
@@ -1536,16 +1454,11 @@ intersect_ϵ²_cnts = Dict([element => count(==(element),intersect_ϵ²) for ele
 
 # values(combined_combo_cnts)
 
-length(filter(((j,u),) -> j ∈ m.timings.exo && u == 4, combined_combo_cnts)) == 1 &&
-length(filter(((j,u),) -> j ∈ m.timings.exo && u % 2 != 0, combined_combo_cnts)) == 0 &&
-length(filter(((j,u),) -> j ∈ m.timings.exo && u == 2, ϵ²_combo_cnts)) == 1
 
 
+Γ₃[6,16,23]#[17,4]
+gamma3xi[6,16,23]#[17,4]
 
-Γ₃[:,:,1]#[16,5]
-gamma3xi[:,:,1]#[17,4]
-
-C2z0[1,1]*2
 
 Γ₃ = zeros(nximin, nximin, nximin);
 
@@ -1596,7 +1509,7 @@ for (i¹,s¹) in enumerate(inputs)
 
                     Γ₃[i¹,i²,i³] = 1
 
[…]
