### 2023-06-23 1b03490
put everything in a function
