+++ test/test_higher_moments_functions.jl
@@ -1,0 +1,1556 @@
+using MacroModelling
+import IterativeSolvers as ℐ
+using LinearAlgebra, LinearMaps
+
+function intersecting_elements(arr1::Union{Symbol,Vector}, arr2::Union{Symbol,Vector})
+    common = []
+    
+    if arr1 isa Symbol
+        arr1 = [arr1]
+    end
+    if arr2 isa Symbol
+        arr2 = [arr2]
+    else
+        arr2 = copy(arr2)
+    end
+
+    for i in arr1
+        if i in arr2
+            push!(common, i)
+            deleteat!(arr2, findfirst(==(i), arr2))  # remove the matched element from arr2
+        end
+    end
+    return common
+end
+
+function upper_triangle(mat::AbstractArray{T,2}) where T
+    @assert size(mat, 1) == size(mat, 2) "The input matrix must be square"
+
+    upper_elems = T[]
+    for i in 1:size(mat, 1)
+        for j in i:size(mat, 2)
+            push!(upper_elems, mat[i, j])
+        end
+    end
+    return upper_elems
+end
+
+
+function upper_triangle(mat::AbstractArray{T,3}; alt::Bool = false, triple::Bool = false) where T
+    @assert size(mat, 1) == size(mat, 2) "The input matrix must be square"
+
+    upper_elems = T[]
+    if alt
+        for i in 1:size(mat, 1)
+            for j in 1:size(mat, 2)
+                for k in j:size(mat, 3)
+                    push!(upper_elems, mat[i, j, k])
+                end
+            end
+        end
+    elseif triple
+        for i in 1:size(mat, 1)
+            for j in i:size(mat, 2)
+                for k in j:size(mat, 3)
+                    push!(upper_elems, mat[i, j, k])
+                end
+            end
+        end
+    else
+        for j in 1:size(mat, 2)
+            for k in j:size(mat, 3)
+                for i in 1:size(mat, 1)
+                    push!(upper_elems, mat[i, j, k])
+                end
+            end
+        end
+    end
+    return upper_elems
+end
+
+function duplication(p)
+    a = sparse(tril(ones(p,p)))
+
+    j = 1
+
+    for k in 1:p
+        for i in 1:p
+            if a[i,k]== 1
+                a[i,k] = j
+                j +=1
+            end
+        end
+    end
+
+    a = a + transpose(tril(a,-1))
+
+    j = Int.(vec(a))
+
+    mm = Int(p*(p+1)/2)
+    
+    DP = zeros(p*p,mm)
+
+    for r in 1:size(DP,1)
+        DP[r, j[r]] = 1
+    end
+
+    DPinv = (DP'*DP)\DP'
+
+    return DP, DPinv
+end
+
+
+function triplication(p)
+    TP = zeros(Int,p^3, Int(p*(p+1)*(p+2)/6))
+    # TPinv = zeros(Int(p*(p+1)*(p+2)/6), Int(p*(p+1)*(p+2)/6))
+
+    for k=1:p
+        for j=k:p
+            for i=j:p
+                idx = unique([[i, j, k],
+                            [i, k, j],
+                            [j, k, i],
+                            [j, i, k],
+                            [k, j, i],
[…]
