### 2023-06-24 #5557 @jbytecode
👋👋👋 Dear @SichengHe and @Athene-ai  👋👋👋

JOSS publishes articles about open source research software. The submission I'd like you to review is titled: "**MacroModelling.jl: A Julia package for developing and solving dynamic stochastic general equilibrium models**". You can find more information at the top of this Github issue (https://github.com/openjournals/joss-reviews/issues/5557).

Thank you in advance!
