### 2023-06-25 #5557 @Athene-ai
I am available

On Saturday, 24 June 2023, Mehmet Hakan Satman ***@***.***>
wrote:

> 👋👋👋 Dear @SichengHe <https://github.com/SichengHe> and @Athene-ai
> <https://github.com/Athene-ai> 👋👋👋
>
> Would you be willing to assist in reviewing this submission for *JOSS* (*Journal
> of Open Source Software*)?
>
> JOSS publishes articles about open source research software. The
> submission I'd like you to review is titled: "*MacroModelling.jl: A Julia
> package for developing and solving dynamic stochastic general equilibrium
> models*". You can find more information at the top of this Github issue (
> #5557 <https://github.com/openjournals/joss-reviews/issues/5557>).
>
> The review process at JOSS is unique: it takes place in a GitHub issue, is
> open, and author-reviewer-editor conversations are encouraged. If you have
> any questions please let me know.
>
> This is the pre-review issue. After setting at least 2 reviewers we will
> start the review process in a separate thread. In that thread, there will
> be 23 check items for each single reviewer.
>
> Thank you in advance!
>
