### 2023-06-25 60123ac
works for 3 vars and 3 shocks
