+++ test/test_higher_moments_functions.jl
@@ -147,7 +147,7 @@ end
 # translate_mod_file("/Users/thorekockerols/Downloads/ReplicationDSGEHOS-main/AnSchorfheide_Gaussian.mod")
 # include("/Users/thorekockerols/Downloads/ReplicationDSGEHOS-main/AnSchorfheide_Gaussian.jl")
 
-include("AnSchorfheide_Gaussian.jl")
+include("AnSchorfheide_Gaussian3.jl")
 m = AnSchorfheide_Gaussian
 
 
@@ -209,7 +209,7 @@ ybar = sol[1][indexin(intersect(model_order,m.var),m.var)]
 ## First-order moments, ie expectation of variables
 IminA = I - A
 Ez   = IminA \ c
-Ey   = ybar + C * Ez + d; # recall y = yss + C*z + d
+Ey   = ybar + C * Ez + d # recall y = yss + C*z + d
 
 
 ## Compute Zero-Lag Cumulants of innovations, states and controls
@@ -605,7 +605,8 @@ function build_Γ₃(m, C2z0::AbstractMatrix, Ey::Vector)
         shock_indices = [1:nu + nu2..., nximin - nu3 + 1:nximin...]
     end
     
-    Γ₃ = zeros(nximin, nximin, nximin);
+    # Γ₃ = zeros(nximin, nximin, nximin);
+    Γ₃ = spzeros(nximin^2, nximin);
 
     for (i¹,s¹) in enumerate(inputs)
         for (i²,s²) in enumerate(inputs)
@@ -645,46 +646,46 @@ function build_Γ₃(m, C2z0::AbstractMatrix, Ey::Vector)
                         length(filter(((j,u),) -> j ∈ m.timings.exo && u == 2, ϵ²_combo_cnts)) == 1 &&
                         length(intersect_ϵ²_cnts) > 0
 
-                        Γ₃[i¹,i²,i³] = 2
+                        Γ₃[(i¹-1)*nximin+i²,i³] = 2
 
                     elseif all([k ∈ m.timings.exo && v == 2 for (k,v) in combined_combo_cnts]) &&
                         length(ϵ²_combo_cnts) == 0
 
-                        Γ₃[i¹,i²,i³] = 1
+                        Γ₃[(i¹-1)*nximin+i²,i³] = 1
 
                     elseif i¹ == i² && i¹ == i³ && all(values(combined_combo_cnts) .== 6)
 
-                        Γ₃[i¹,i²,i³] = 8
+                        Γ₃[(i¹-1)*nximin+i²,i³] = 8
 
                     elseif length(filter(((j,u),) -> j ∈ m.timings.exo && u == 4, combined_combo_cnts)) == 1 &&
                         length(filter(((j,u),) -> j ∈ m.timings.exo && u == 2, combined_combo_cnts)) == 1  &&
                         length(ϵ²_combo_cnts) == 0
 
-                        Γ₃[i¹,i²,i³] = 3
+                        Γ₃[(i¹-1)*nximin+i²,i³] = 3
 
                     elseif any(values(combined_combo_cnts) .== 6) && 
                         length(intersect_ϵ²_cnts) > 0 && 
                         !(i¹ == i² && i¹ == i³)
 
-                        Γ₃[i¹,i²,i³] = 12 # Variance of ϵ²
+                        Γ₃[(i¹-1)*nximin+i²,i³] = 12 # Variance of ϵ²
 
[…]
