### 2023-06-25 867814d
cleaned up functional script, no ordering
