### 2023-06-27 #5557 @jbytecode
@daviddewhurst - Thank you for accepting our invitation. Since the whole review process is interactive, there is not such thing as a round. While we are seeking the 2nd reviewer, you can take a look at the [instructions](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html). After we find a second reviewer, the review will start in a separate thread.
