### 2023-06-27 #5557 @gdalle
Hi!
I'll be honest, I have no clue what a dynamic stochastic general equilibrium model is, but I can certainly evaluate the development craftsmanship and AD aspects
