### 2023-06-27 #5598 @Athene-ai
Yes I have

On Tuesday, 27 June 2023, Mehmet Hakan Satman ***@***.***>
wrote:

> @Athene-ai <https://github.com/Athene-ai> - You couldn't have finished
> the entire review in 5 minutes, right?
>
