### 2023-06-27 #5598 @thorek1
Re: currently failing tests

Indeed they can fail on windows machines if the test instance doesn't have enough memory allocated. The latest release failed only on the windows 1.8 instance, but not windows 1.9. This doesn't happen on Linux at all so far.

I tried deleting objects and have the garbage collector free up RAM in the test script but that's not sufficient on windows. Suggestions welcome
