+++ Project.toml
@@ -32,11 +32,15 @@ Subscripts = "2b7f82d5-8785-4f63-971e-f18ddbeb808e"
 SymPy = "24249f21-da20-56a4-8eb1-6a02cf4ae2e6"
 Symbolics = "0c5d862f-8b57-4792-8d23-62f2024744c7"
 
+[weakdeps]
+StatsPlots = "f3b207a7-027a-5e70-b257-86293d7955fd"
+Turing = "fce5fe82-541a-59a6-adf8-730c64b5f9a0"
+
 [compat]
 AxisKeys = "^0.2"
 BlockTriangularForm = "^0.1"
 ChainRulesCore = "^1"
-ComponentArrays = "^0.13"
+ComponentArrays = "^0.13, 0.14"
 DataStructures = "^0.18"
 DocStringExtensions = "^0.8, 0.9"
 DynarePreprocessor_jll = "^6.2"
@@ -78,7 +82,3 @@ Zygote = "e88e6eb3-aa80-5325-afca-941959d7151f"
 
 [targets]
 test = ["Aqua", "CSV", "DataFrames", "DynamicPPL", "MCMCChains", "LineSearches", "Optim", "Test", "Turing", "FiniteDifferences", "Zygote", "Plots", "StatsPlots"]
-
-[weakdeps]
-StatsPlots = "f3b207a7-027a-5e70-b257-86293d7955fd"
-Turing = "fce5fe82-541a-59a6-adf8-730c64b5f9a0"
