### 2023-07-16 thorek1/MacroModelling.jl#33 @gdalle
Installation fails
Hi there, congrats on the package!
I'm reviewing your JOSS submission, so expect more issues from me as I make my way through the checklist :)

https://github.com/openjournals/joss-reviews/issues/5598

When I try to install MacroModelling.jl, some Python compatibility issues pop up.
I strongly suggest you use [PythonCall.jl](https://github.com/cjdoris/PythonCall.jl) + [CondaPkg.jl](https://github.com/cjdoris/CondaPkg.jl) to manage your Python dependencies.
As the following bug suggests, PyCall is less reliable and reproducible.

```julia
(@v1.9) pkg> activate --temp
  Activating new project at `/tmp/jl_ZNvdrc`

(jl_ZNvdrc) pkg> add MacroModelling
...
Precompiling project...
  ✗ MacroModelling
  0 dependencies successfully precompiled in 2 seconds. 196 already precompiled.
  1 dependency errored. To see a full report either run `import Pkg; Pkg.precompile()` or load the package

julia> using MacroModelling
[ Info: Precompiling MacroModelling [687ffad2-3618-405e-ac50-e0f7b9c75e44]
ERROR: LoadError: InitError: Failed to import required Python module sympy.

For automated sympy installation, try configuring PyCall to use the Conda.jl package's Python "Miniconda" distribution within Julia. Relaunch Julia and run:
    ENV["PYTHON"]=""
    Pkg.build("PyCall")
before trying again.

Note that Conda.jl can use custom conda environment. Read more about `CONDA_JL_HOME` configuration in Conda.jl documentation:
    https://github.com/JuliaPy/Conda.jl

To install sympy using /home/guillaume/mambaforge/bin/conda, you can run the following command from your system shell:
    /home/guillaume/mambaforge/bin/conda install sympy

The pyimport exception was: PyError (PyImport_ImportModule

The Python package sympy could not be imported by pyimport. Usually this means
that you did not install sympy in the Python version being used by PyCall.

PyCall is currently configured to use the Python version at:

/home/guillaume/mambaforge/bin/python3

and you should use whatever mechanism you usually use (apt-get, pip, conda,
etcetera) to install the Python package containing the sympy module.

One alternative is to re-configure PyCall to use a different Python
version on your system: set ENV["PYTHON"] to the path/name of the python
executable you want to use, run Pkg.build("PyCall"), and re-launch Julia.

Another alternative is to configure PyCall to use a Julia-specific Python
distribution via the Conda.jl package (which installs a private Anaconda
Python distribution), which has the advantage that packages can be installed
and kept up-to-date via Julia.  As explained in the PyCall documentation,
set ENV["PYTHON"]="", run Pkg.build("PyCall"), and re-launch Julia. Then,
To install the sympy module, you can use `pyimport_conda("sympy", PKG)`,
where PKG is the Anaconda package that contains the module sympy,
or alternatively you can use the Conda package directly (via
`using Conda` followed by `Conda.add` etcetera).

) <class 'ModuleNotFoundError'>
ModuleNotFoundError("No module named 'sympy'")
```
