### 2023-07-16 thorek1/MacroModelling.jl#34 @gdalle
Memory resets in the tests?
https://github.com/thorek1/MacroModelling.jl/blob/f4e752e931b490f01882e9ae8046d8c6b57eaff8/test/runtests.jl#L31

Why is this needed after every test set?

https://github.com/openjournals/joss-reviews/issues/5598
