### 2023-07-16 thorek1/MacroModelling.jl#35 @gdalle
CI tests failing
Apparently it's because of a lack of memory

https://github.com/openjournals/joss-reviews/issues/5598#issuecomment-1610198924

I think #34 is a sign that there are unwanted memory leaks. Curious about the answer there.
