### 2023-07-16 thorek1/MacroModelling.jl#36 @gdalle
Summary is too technical
I'm not an economist, so I have no clue what a DSGE model is. Adding one or two sentences to the paper explaining what they to and outlining their relevance in your field would go a long way

https://github.com/openjournals/joss-reviews/issues/5598
