### 2023-07-16 thorek1/MacroModelling.jl#37 @gdalle
State of the field is missing
The paper claims that the software
> stands out for its ability to calculate sensitivities of model moments, its automatic variable declaration, and effective steady state solver.

However, a detailed comparison with related packages is only provided in the README. I think this comparison also belongs in the docs, and a synthetic version in the main paper.

When I'm lazy, I often copy `README.md` into `docs/src/index.md` so that I don't have to write the same content twice. The only thing that gets lost in the way are the edit links, but that can be fixed.

https://github.com/openjournals/joss-reviews/issues/5598
