### 2023-07-16 thorek1/MacroModelling.jl#38 @gdalle
Performance claims
I'm not sure if
> The package stands out for its [...] _effective_ steady state solver.

is a performance claim, but if it is, it needs to be backed up with benchmarks against competing packages.

If you're looking for inspiration, I recently did cross-language benchmarks (Julia/Python) for [HiddenMarkovModels.jl](https://github.com/gdalle/HiddenMarkovModels.jl).
They are compatible with [PkgBenchmark.jl](https://github.com/JuliaCI/PkgBenchmark.jl) and are also automatically added to the docs for each new release I tag.

https://github.com/openjournals/joss-reviews/issues/5598
