### 2023-07-16 #5598 @gdalle
Finished my first round of review, see the issues above. For now I'm especially waiting on https://github.com/thorek1/MacroModelling.jl/issues/33 to be fixed so that I can evaluate reproducibility.
