### 2023-07-16 thorek1/MacroModelling.jl#34 @thorek1
indeed tests on the windows machine running 1.8 fail with out of memory error messages. the highlighted code was an attempt to overcome this issue be deleting the large model objects and triggering the garbage collector. my current best understanding is that the out of memory errors persist because the compiled functions are kept in memory. a lot of functions are compiled in case of higher order Taylor approximations of the model. in order to have the CI runs go through without errors, I can either drop the 1.8 windows run or avoid doing 3rd order for 1.8. windows. all other CI runs are unaffected
