### 2023-07-17 thorek1/MacroModelling.jl#34 @gdalle
Have you checked if the variable m actually contains anything after a test set? From what I can tell, the macro introduces a local scope anyway (https://docs.julialang.org/en/v1/stdlib/Test/#Working-with-Test-Sets) so my guess is m is not assigned
