### 2023-07-17 thorek1/MacroModelling.jl#34 @gdalle
Do you have any support for your intuition that memory runs out because of the size of the compiled code? That sounds very surprising to me
