### 2023-07-17 thorek1/MacroModelling.jl#34 @thorek1
I know that when i setup a model everything related to the model is kept in a struct with the name of the model. I also know that he compiles the functions and i guess they are kept in memory. i will go through the julia docs to find a way to give me a breakdown of the objects occupying memory inside a julia session...
