### 2023-07-17 thorek1/MacroModelling.jl#34 @gdalle
Honestly I don't have an issue with removing the tests on Windows, and I will tick the CI box in the review as long as it runs on Ubuntu. I just find it very puzzling and I'd like to fix it together instead of sweeping it under the rug ^^
