### 2023-07-17 thorek1/MacroModelling.jl#35 @gdalle
@thorek1 did you see this?

https://github.com/JuliaLang/julia/issues/39396#issuecomment-768456099
