### 2023-07-19 856f29e
remove windows 1.8 CI
