### 2023-07-19 thorek1/MacroModelling.jl#35 @thorek1
I removed windows + Julia 1.8 from CI runs so now it passes. But I will dig deeper what was the cause there.
