### 2023-07-20 thorek1/MacroModelling.jl#33 @thorek1
this one is about sympy (python package) not being there. essentially you need to make sure julia talks to python (SymPy.jl relies on PyCall.jl and Conda.jl) and has sympy installed in the python environment linked to julia. In a previous version I tried to automatise this with CondaPkg but it had no effect in my case. I will try again with the latest version.
