### 2023-07-20 thorek1/MacroModelling.jl#33 @gdalle
> essentially you need to make sure julia talks to python (SymPy.jl relies on PyCall.jl and Conda.jl) and has sympy installed in the python environment linked to julia.

I figured as much, but I tried installing it the naive way by just following the instructions in the README, which didn't mention anything about setting up Python :)

> In a previous version I tried to automatise this with CondaPkg but it had no effect in my case. I will try again with the latest version.

The main advantage of CondaPkg is that it dispenses the user from having to fight with Python themselves. I've used it successfully several times, happy to help if needed
