### 2023-07-21 d05e1b8
add sympy dependency through condapkg
