+++ CondaPkg.toml
@@ -1,0 +1,3 @@
+
+[deps]
+sympy = ""
