+++ Project.toml
@@ -1,13 +1,14 @@
 name = "MacroModelling"
 uuid = "687ffad2-3618-405e-ac50-e0f7b9c75e44"
 authors = ["Thore Kockerols <mail@thorekockerols.com>"]
-version = "0.1.23"
+version = "0.1.24"
 
 [deps]
 AxisKeys = "94b1ba4f-4ee9-5380-92f1-94cde586c3c5"
 BlockTriangularForm = "adeb47b7-70bf-415a-bb24-c358563e873a"
 ChainRulesCore = "d360d2e6-b24c-11e9-a2a3-2a2ae2dbcce4"
 ComponentArrays = "b0b7db55-cfe3-40fc-9ded-d10e2dbeff66"
+CondaPkg = "992eb4ea-22a4-4c89-a5bb-47a3300528ab"
 DataStructures = "864edb3b-99cc-5e75-8d2d-829cb0a9cfe8"
 DocStringExtensions = "ffbed154-4ef7-542d-bbb7-c09d3a79fcae"
 DynarePreprocessor_jll = "23afba7c-24e5-5ee2-bc2c-b42e07f0492a"
@@ -32,11 +33,16 @@ Subscripts = "2b7f82d5-8785-4f63-971e-f18ddbeb808e"
 SymPy = "24249f21-da20-56a4-8eb1-6a02cf4ae2e6"
 Symbolics = "0c5d862f-8b57-4792-8d23-62f2024744c7"
 
+[weakdeps]
+StatsPlots = "f3b207a7-027a-5e70-b257-86293d7955fd"
+Turing = "fce5fe82-541a-59a6-adf8-730c64b5f9a0"
+
 [compat]
 AxisKeys = "^0.2"
 BlockTriangularForm = "^0.1"
 ChainRulesCore = "^1"
 ComponentArrays = "^0.13"
+CondaPkg = "^0.2"
 DataStructures = "^0.18"
 DocStringExtensions = "^0.8, 0.9"
 DynarePreprocessor_jll = "^6.2"
@@ -78,7 +84,3 @@ Zygote = "e88e6eb3-aa80-5325-afca-941959d7151f"
 
 [targets]
 test = ["Aqua", "CSV", "DataFrames", "DynamicPPL", "MCMCChains", "LineSearches", "Optim", "Test", "Turing", "FiniteDifferences", "Zygote", "Plots", "StatsPlots"]
-
-[weakdeps]
-StatsPlots = "f3b207a7-027a-5e70-b257-86293d7955fd"
-Turing = "fce5fe82-541a-59a6-adf8-730c64b5f9a0"
