+++ v0.1.24/search_index.js
@@ -1,0 +1,3 @@
+var documenterSearchIndex = {"docs":
+[{"location":"unfinished_docs/how_to/#Use-calibration-equations","page":"-","title":"Use calibration equations","text":"","category":"section"},{"location":"unfinished_docs/how_to/","page":"-","title":"-","text":"Next we need to add the parameters of the model. The macro @parameters <name of the model> takes care of this:","category":"page"},{"location":"unfinished_docs/how_to/","page":"-","title":"-","text":"@parameters RBC begin\n    std_z = 0.01\n    ρ = 0.2\n    δ = 0.02\n    α = 0.5\n    β = 0.95\nend","category":"page"},{"location":"unfinished_docs/how_to/","page":"-","title":"-","text":"No need for line endings. If you want to define a parameter as a function of another parameter you can do this:","category":"page"},{"location":"unfinished_docs/how_to/","page":"-","title":"-","text":"@parameters RBC begin\n    std_z = 0.01\n    ρ = 0.2\n    δ = 0.02\n    α = 0.5\n    beta1 = 1\n    beta2 = .95\n    β | β = beta2/beta1\nend","category":"page"},{"location":"unfinished_docs/how_to/","page":"-","title":"-","text":"Note that the parser takes parameters assigned to a numerical value first and then solves for the parameters defined by relationships: β | .... This means also the following will work:","category":"page"},{"location":"unfinished_docs/how_to/","page":"-","title":"-","text":"@parameters RBC begin\n    β | β = beta2/beta1\n    std_z = 0.01\n    ρ = 0.2\n    δ = 0.02\n    α = 0.5\n    beta1 = 1\n    beta2 = .95\nend","category":"page"},{"location":"unfinished_docs/how_to/","page":"-","title":"-","text":"More interestingly one can use (non-stochastic) steady state values in the relationships:","category":"page"},{"location":"unfinished_docs/how_to/","page":"-","title":"-","text":"@parameters RBC begin\n    β = .95\n    std_z = 0.01\n    ρ = 0.2\n    δ = 0.02\n    α | k[ss] / (4 * q[ss]) = 1.5\nend","category":"page"},{"location":"unfinished_docs/how_to/#Higher-order-perturbation-solutions","page":"-","title":"Higher order perturbation solutions","text":"","category":"section"},{"location":"unfinished_docs/how_to/#How-to-estimate-a-model","page":"-","title":"How to estimate a model","text":"","category":"section"},{"location":"unfinished_docs/how_to/#Interactive-plotting","page":"-","title":"Interactive plotting","text":"","category":"section"},{"location":"call_index/#Index","page":"Index","title":"Index","text":"","category":"section"},{"location":"call_index/","page":"Index","title":"Index","text":"","category":"page"},{"location":"api/","page":"API","title":"API","text":"Modules = [MacroModelling]\nOrder   = [:function, :macro]","category":"page"},{"location":"api/#MacroModelling.Beta-Union{Tuple{T}, NTuple{4, T}} where T<:Real","page":"API","title":"MacroModelling.Beta","text":"Beta(μ, σ, lower_bound, upper_bound; μσ)\n\n\nConvenience wrapper for the truncated Beta distribution.\n\nIf μσ = true then μ and σ are translated to the parameters of the distribution. Otherwise μ and σ represent the parameters of the distribution.\n\nArguments\n\nμ [Type: Real]: mean or first parameter of the distribution, \nσ [Type: Real]: standard deviation or first parameter of the distribution\nlower_bound [Type: Real]: truncation lower bound of the distribution\nupper_bound [Type: Real]: truncation upper bound of the distribution\n\nKeyword Arguments\n\nμσ [Type: Bool]: switch whether μ and σ represent the moments of th
[…]
