### 2023-07-21 thorek1/MacroModelling.jl#33 @thorek1
I added CondaPkg and sympy as a python dependency. it worked on the Julia linux docker image out of the box. thx for the hint.
