### 2023-07-22 thorek1/MacroModelling.jl#33 @thorek1
This seems trickier than i thought. SymPy.jl relies on PyCall->Conda
I understand this setup does not guarantee automatic installation of the sympy python package but using PythonCall->CondaPkg would. let me check whether importing it using pythoncall ensures the same functionality as SymPy.jl
