### 2023-07-22 9704d21
first try: segfaulting
