+++ src/MacroModelling.jl
@@ -5,8 +5,21 @@ import DocStringExtensions: FIELDS, SIGNATURES, TYPEDEF, TYPEDSIGNATURES, TYPEDF
 # import StatsFuns: normcdf
 using PrecompileTools
 import SpecialFunctions: erfcinv, erfc
-import SymPy: @vars, solve, subs, free_symbols
-import SymPy
+
+import CondaPkg
+import PythonCall
+SymPy = PythonCall.pyimport("sympy")
+
+import Base: log, exp
+log(x::PythonCall.Py) = SymPy.log(x)
+exp(x::PythonCall.Py) = SymPy.exp(x)
+
+function translate_exponentiation_from_python_to_julia(x)
+    replace(x,"**" => "^")
+end
+# import SymPy: SymPy.@vars, solve, subs, free_symbols
+# import SymPy
+
 import Symbolics
 import ForwardDiff as ℱ 
 # import Zygote
@@ -141,10 +154,13 @@ end
 
 function simplify(ex::Expr)
     ex_ss = convert_to_ss_equation(ex)
+    # println(ex_ss)
+    for exx in get_symbols(ex_ss)
+        # println(:($exx = SymPy.Symbol($(string(exx)), real = true, finite = true)))
+	    eval(:($exx = SymPy.Symbol($(string(exx)), real = true, finite = true)))
+    end
 
-	eval(:(@vars $(get_symbols(ex_ss)...) real = true finite = true ))
-
-	parsed = ex_ss |> eval |> string |> Meta.parse
+	parsed = ex_ss |> eval |> string |> translate_exponentiation_from_python_to_julia |> Meta.parse
 
     postwalk(x -> x isa Expr ? x.args[1] == :conjugate ? x.args[2] : x : x, parsed)
 end
@@ -423,13 +439,24 @@ function create_symbols_eqs!(𝓂::ℳ)
         end
     end
 
-    expr =  quote
-                @vars $(symbols_pos...)  real = true finite = true positive = true
-                @vars $(symbols_neg...)  real = true finite = true negative = true 
-                @vars $(symbols_none...) real = true finite = true 
-            end
+    # expr =  quote
+    for pos in symbols_pos
+        eval(:($pos = SymPy.Symbol($(string(pos)), real = true, finite = true, positive = true)))
+    end
+    
+    for neg in symbols_neg
+        eval(:($neg = SymPy.Symbol($(string(neg)), real = true, finite = true, negative = true)))
+    end
 
-    eval(expr)
+    for none in symbols_none
+        eval(:($none = SymPy.Symbol($(string(none)), real = true, finite = true)))
+    end
+            #     SymPy.symbols($(string(symbols_pos...),  real = true, finite = true, positive = true)
+            #     SymPy.symbols($(string(symbols_neg...)  ,real = true, finite = true, negative = true)
+            #     SymPy.symbols($(string(symbols_none...)) ,real = true, finite = true)
+            # end
+
+    # eval(expr)
 
     symbolics(map(x->eval(:($x)),𝓂.ss_aux_equations),
                 map(x->eval(:($x)),𝓂.dyn_equations),
@@ -506,17 +533,19 @@ function remove_redundant_SS_vars!(𝓂::ℳ, Symbolics::symbolics)
 
     for i in redundant_idx
         for var_to_solve in redundant_vars[i]
-            soll = try solve(ss_equations[i],var_to_solve)
+            # println(ss_equations[i])
+            # println(var_to_solve)
+            soll = try SymPy.solve(ss_equations[i],var_to_solve)
             catch
             end
             
             if isnothing(soll)
[…]
