+++ src/structures.jl
@@ -94,58 +94,58 @@ struct timings
 end
 
 struct symbolics
-    ss_equations::Vector{SymPy.Sym}
-    dyn_equations::Vector{SymPy.Sym}
-    # dyn_equations_future::Vector{SymPy.Sym}
+    ss_equations::Vector{PythonCall.Py}
+    dyn_equations::Vector{PythonCall.Py}
+    # dyn_equations_future::Vector{PythonCall.Py}
     
-    # dyn_shift_var_present_list::Vector{Set{SymPy.Sym}}
-    # dyn_shift_var_past_list::Vector{Set{SymPy.Sym}}
-    # dyn_shift_var_future_list::Vector{Set{SymPy.Sym}}
+    # dyn_shift_var_present_list::Vector{Set{PythonCall.Py}}
+    # dyn_shift_var_past_list::Vector{Set{PythonCall.Py}}
+    # dyn_shift_var_future_list::Vector{Set{PythonCall.Py}}
 
-    # dyn_shift2_var_past_list::Vector{Set{SymPy.Sym}}
+    # dyn_shift2_var_past_list::Vector{Set{PythonCall.Py}}
 
-    dyn_var_present_list::Vector{Set{SymPy.Sym}}
-    dyn_var_past_list::Vector{Set{SymPy.Sym}}
-    dyn_var_future_list::Vector{Set{SymPy.Sym}}
-    # dyn_ss_list::Vector{Set{SymPy.Sym}}
-    dyn_exo_list::Vector{Set{SymPy.Sym}}
+    dyn_var_present_list::Vector{Set{PythonCall.Py}}
+    dyn_var_past_list::Vector{Set{PythonCall.Py}}
+    dyn_var_future_list::Vector{Set{PythonCall.Py}}
+    # dyn_ss_list::Vector{Set{PythonCall.Py}}
+    dyn_exo_list::Vector{Set{PythonCall.Py}}
 
-    # dyn_exo_future_list::Vector{Set{SymPy.Sym}}
-    # dyn_exo_present_list::Vector{Set{SymPy.Sym}}
-    # dyn_exo_past_list::Vector{Set{SymPy.Sym}} 
+    # dyn_exo_future_list::Vector{Set{PythonCall.Py}}
+    # dyn_exo_present_list::Vector{Set{PythonCall.Py}}
+    # dyn_exo_past_list::Vector{Set{PythonCall.Py}} 
 
-    dyn_future_list::Vector{Set{SymPy.Sym}}
-    dyn_present_list::Vector{Set{SymPy.Sym}}
-    dyn_past_list::Vector{Set{SymPy.Sym}} 
+    dyn_future_list::Vector{Set{PythonCall.Py}}
+    dyn_present_list::Vector{Set{PythonCall.Py}}
+    dyn_past_list::Vector{Set{PythonCall.Py}} 
 
-    var_present_list::Vector{Set{SymPy.Sym}}
-    var_past_list::Vector{Set{SymPy.Sym}}
-    var_future_list::Vector{Set{SymPy.Sym}}
-    ss_list::Vector{Set{SymPy.Sym}}
-    var_list::Vector{Set{SymPy.Sym}}
-    # dynamic_variables_list::Vector{Set{SymPy.Sym}}
-    # dynamic_variables_future_list::Vector{Set{SymPy.Sym}}
+    var_present_list::Vector{Set{PythonCall.Py}}
+    var_past_list::Vector{Set{PythonCall.Py}}
+    var_future_list::Vector{Set{PythonCall.Py}}
+    ss_list::Vector{Set{PythonCall.Py}}
+    var_list::Vector{Set{PythonCall.Py}}
+    # dynamic_variables_list::Vector{Set{PythonCall.Py}}
[…]
