+++ Project.toml
@@ -21,6 +21,7 @@ LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
 LinearMaps = "7a12625a-238d-50fd-b39a-03d52299707e"
 MacroTools = "1914dd2f-81c6-5fcd-8719-6d5c9610ff09"
 PrecompileTools = "aea7be01-6a6a-4083-8856-8a6e6704d82a"
+PythonCall = "6099a3de-0909-46bc-b1f4-468b9a2dfc0d"
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
 RecursiveFactorization = "f2c3362d-daeb-58d1-803e-2bc74f2840b4"
 Reexport = "189a3867-3050-52da-a836-e630ba90ab69"
@@ -30,7 +31,6 @@ SparseArrays = "2f01184e-e22b-5df5-ae63-d93ebab69eaf"
 SpecialFunctions = "276daf66-3868-5448-9aa4-cd146d93841b"
 SpeedMapping = "f1835b91-879b-4a3f-a438-e4baacf14412"
 Subscripts = "2b7f82d5-8785-4f63-971e-f18ddbeb808e"
-SymPy = "24249f21-da20-56a4-8eb1-6a02cf4ae2e6"
 Symbolics = "0c5d862f-8b57-4792-8d23-62f2024744c7"
 
 [weakdeps]
@@ -54,6 +54,7 @@ LaTeXStrings = "^1"
 LinearMaps = "^3"
 MacroTools = "^0.5"
 PrecompileTools = "^1"
+PythonCall = "^0.9"
 RecursiveFactorization = "^0.2"
 Reexport = "^1"
 Requires = "^1"
@@ -62,7 +63,6 @@ SpecialFunctions = "^2"
 SpeedMapping = "^0.3"
 StatsPlots = "^0.15"
 Subscripts = "^0.1"
-SymPy = "^1"
 Symbolics = "^5"
 Turing = "^0.25"
 julia = "1.8"
