### 2023-07-22 6c75174
second try with SymPyCall works
