+++ src/MacroModelling.jl
@@ -5,8 +5,9 @@ import DocStringExtensions: FIELDS, SIGNATURES, TYPEDEF, TYPEDSIGNATURES, TYPEDF
 # import StatsFuns: normcdf
 using PrecompileTools
 import SpecialFunctions: erfcinv, erfc
-import SymPy: @vars, solve, subs, free_symbols
-import SymPy
+import PythonCall
+import SymPyCall: symbols, solve, subs, free_symbols
+import SymPyCall
 import Symbolics
 import ForwardDiff as ℱ 
 # import Zygote
@@ -142,7 +143,9 @@ end
 function simplify(ex::Expr)
     ex_ss = convert_to_ss_equation(ex)
 
-	eval(:(@vars $(get_symbols(ex_ss)...) real = true finite = true ))
+    for x in get_symbols(ex_ss)
+	    eval(:($x = symbols($(string(x)), real = true, finite = true)))
+    end
 
 	parsed = ex_ss |> eval |> string |> Meta.parse
 
@@ -423,13 +426,22 @@ function create_symbols_eqs!(𝓂::ℳ)
         end
     end
 
-    expr =  quote
-                @vars $(symbols_pos...)  real = true finite = true positive = true
-                @vars $(symbols_neg...)  real = true finite = true negative = true 
-                @vars $(symbols_none...) real = true finite = true 
-            end
+    for pos in symbols_pos
+        eval(:($pos = symbols($(string(pos)), real = true, finite = true, positive = true)))
+    end
+    for neg in symbols_neg
+        eval(:($neg = symbols($(string(neg)), real = true, finite = true, negative = true)))
+    end
+    for none in symbols_none
+        eval(:($none = symbols($(string(none)), real = true, finite = true)))
+    end
+    # expr =  quote
+    #             @vars $(symbols_pos...)  real = true finite = true positive = true
+    #             @vars $(symbols_neg...)  real = true finite = true negative = true 
+    #             @vars $(symbols_none...) real = true finite = true 
+    #         end
 
-    eval(expr)
+    # eval(expr)
 
     symbolics(map(x->eval(:($x)),𝓂.ss_aux_equations),
                 map(x->eval(:($x)),𝓂.dyn_equations),
@@ -514,9 +526,9 @@ function remove_redundant_SS_vars!(𝓂::ℳ, Symbolics::symbolics)
                 continue
             end
             
-            if length(soll) == 0 || soll == SymPy.Sym[0] # take out variable if it is redundant from that euation only
+            if length(soll) == 0 || soll == SymPyCall.Sym[0] # take out variable if it is redundant from that euation only
                 push!(Symbolics.var_redundant_list[i],var_to_solve)
-                ss_equations[i] = ss_equations[i].subs(var_to_solve,1).replace(SymPy.Sym(ℯ),exp(1)) # replace euler constant as it is not translated to julia properly
+                ss_equations[i] = ss_equations[i].subs(var_to_solve,1).replace(SymPyCall.Sym(ℯ),exp(1)) # replace euler constant as it is not translated to julia properly
             end
 
         end
@@ -563,7 +575,7 @@ function solve_steady_state!(𝓂::ℳ, symbolic_SS, Symbolics::symbolics; verbo
     
     n = n_blocks
 
-    ss_equations = vcat(Symbolics.ss_equations,Symbolics.calibration_equations) .|> SymPy.Sym
+    ss_equations = vcat(Symbolics.ss_equations,Symbolics.calibration_equations) .|> SymPyCall.Sym
     # println(ss_equations)
 
     SS_solve_func = []
[…]
