+++ src/SymPyCall/gen_methods.jl
@@ -1,0 +1,197 @@
+## XXX We should be hooking to the pyconvert methods XXX
+## we need Py -> Sym (asSymbolic)
+## we need Sym -> Py unSym
+asSymbolic(x::Sym) = x
+function asSymbolic(x::Py)
+
+    cls = x.__class__
+
+    if pyconvert(Bool, cls == pybuiltins.list)
+        return asSymbolic.(x)
+    elseif pyconvert(Bool, cls == pybuiltins.dict)
+        return Dict(asSymbolic(k) => asSymbolic(v) for (k,v) ∈ pyconvert(PyDict, x))
+    elseif pyconvert(Bool, cls == pybuiltins.tuple)
+        return Tuple(asSymbolic(xᵢ) for xᵢ ∈ x)
+    elseif pyconvert(Bool, cls == pybuiltins.set)
+        return Set(asSymbolic(xᵢ) for xᵢ ∈ x)
+    # elseif pyconvert(Bool, cls == pybuiltins.list)
+    end
+
+
+    Bool(pygetattr(x, "is_FiniteSet", false)) && return Set(asSymbolic(xᵢ) for xᵢ ∈ x)
+
+    if Bool(pygetattr(x, "is_Matrix", false))
+        sz = pyconvert(Tuple, x.shape)
+        !isa(sz[1], Real) && return Sym(x) # MatrixSymbol special case
+        return [asSymbolic(x.__getitem__((i-1, j-1))) for i ∈ 1:sz[1], j ∈ 1:sz[2]]
+    end
+    Sym(x)
+end
+asSymbolic(x::String) = x
+asSymbolic(x) = (@show x; Sym(pyconvert(Py, x)))
+↑(args...; kwargs...) = asSymbolic(args...; kwargs...)
+
+
+# used to pass arguments down into calls
+unSym(x) = x
+unSym(x::Sym) = getfield(x, :py)
+unSym(x::Tuple) = unSym.(x)
+unSym(x::Vector) = Tuple(unSym.(x)) # <-- is this an issue?
+unSym(x::Matrix) = sympy.py.Matrix(Tuple(↓(mᵢ for mᵢ ∈ eachrow(x)))) # mutable dense matrix
+unSym(x::Dict) = convert(PyDict,Dict(unSym(k) => unSym(v) for (k,v) ∈ x)) # AbstractDict?
+unSym(x::Set) = sympy.py.Set(SymPyCall.unSym(collect(x)))
+unSym(x::Irrational{:π}) = unSym(sympy.pi)
+unSymkwargs(kw) = (k=>unSym(v) for (k,v) ∈ kw)
+↓(args...; kwargs...) = unSym(args...; kwargs...)
+
+#PythonCall.Py(x::Sym) = ↓(x)
+PythonCall.Py(x::NTuple{N,T}) where {N, T <: SymbolicObject} = ↓(x)
+PythonCall.Py(x::Vector{T}) where {T <: SymbolicObject} = unSym.(x)
+PythonCall.Py(x::Matrix{T}) where {T <: SymbolicObject} = unSym.(x)
+PythonCall.Py(x::Set{T}) where {T <: SymbolicObject} = unSym(x)
+
+
+
+## --------------------------------------------------
+
+# generic method pattern: (pymodule, pymethod, juliamodule, juliamethod)
+# these don't need to be exported
+# in SymPy we use getmembers to generate this; not sure what is better
+#const
+generic_methods = (
+    (:sympy, :cos, :Base, :cos),
+    (:sympy, :sin, :Base, :sin),
+    (:sympy, :tan, :Base, :tan),
+    (:sympy, :sec, :Base, :sec),
+    (:sympy, :csc, :Base, :csc),
[…]
