+++ .github/workflows/ci.yml
@@ -18,7 +18,7 @@ jobs:
           # - 'nightly' # fails for zygote
         os:
           - ubuntu-latest
-          # - macOS-latest #slow
+          - macOS-latest #slow
           - windows-latest #slow
         arch:
           - x64
