### 2023-07-22 902cb19
take out symbolic SS test
