### 2023-07-22 thorek1/MacroModelling.jl#33 @thorek1
I got it to work with PythonCall as a backend and tested it in a separate branch. The issue is I don't get PythonCall to work with PrecompileTools. Do you have any experience with this?

It would be a petty sacrificing some speed for convenience.

On the PythonCall website it says during precompilation you **must not** interact with python...
