+++ .github/workflows/ci.yml
@@ -29,6 +29,9 @@ jobs:
           - version: '1.8'
             os: ubuntu-latest
             arch: x64
+          - version: '^1.10.0-0'
+            os: ubuntu-latest
+            arch: x64
     steps:
       - uses: actions/checkout@v3
       - uses: julia-actions/setup-julia@latest
