+++ src/MacroModelling.jl
@@ -889,6 +889,9 @@ function solve_steady_state!(𝓂::ℳ, symbolic_SS, Symbolics::symbolics; verbo
                 # push!(SS_solve_func,:(closest_solution = 𝓂.NSSS_solver_cache[findmin([sum(abs2,pars[end] - params_flt) for pars in 𝓂.NSSS_solver_cache])[2]]))
                 # push!(SS_solve_func,:(inits = [transformer(max.(lbs,min.(ubs, closest_solution[$(n_block)] ))),closest_solution[end]]))
                 push!(SS_solve_func,:(inits = max.(lbs,min.(ubs, closest_solution[$(n_block)]))))
+
+                # push!(SS_solve_func,:(println([$(calib_pars_input...),$(other_vars_input...)])))
+
                 push!(SS_solve_func,:(block_solver_RD = block_solver_AD([$(calib_pars_input...),$(other_vars_input...)],
                                                                         $(n_block), 
                                                                         𝓂.ss_solve_blocks[$(n_block)], 
@@ -1009,9 +1012,11 @@ function solve_steady_state!(𝓂::ℳ, symbolic_SS, Symbolics::symbolics; verbo
     end
 
 
-    solve_exp = :(function solve_SS(parameters::Vector{Real}, 𝓂::ℳ, 
-    # fail_fast_solvers_only::Bool, 
-    verbose::Bool)
+    solve_exp = :(function solve_SS(parameters::Vector{Real}, 
+                                    𝓂::ℳ, 
+                                    # fail_fast_solvers_only::Bool, 
+                                    verbose::Bool)
+
                     params_flt = typeof(parameters) == Vector{Float64} ? parameters : ℱ.value.(parameters)
                     current_best = sum(abs2,𝓂.NSSS_solver_cache[end][end] - params_flt)
                     closest_solution_init = 𝓂.NSSS_solver_cache[end]
@@ -1058,6 +1063,352 @@ function solve_steady_state!(𝓂::ℳ, symbolic_SS, Symbolics::symbolics; verbo
 
     return nothing
 end
+
+
+
+
+function solve_steady_state!(𝓂::ℳ; verbose::Bool = false)
+    unknowns = union(𝓂.vars_in_ss_equations, 𝓂.calibration_equations_parameters)
+
+    @assert length(unknowns) <= length(𝓂.ss_aux_equations) + length(𝓂.calibration_equations) "Unable to solve steady state. More unknowns than equations."
+
+    incidence_matrix = fill(0,length(unknowns),length(unknowns))
+
+    eq_list = vcat(union.(union.(𝓂.var_list_aux_SS,
+                                        𝓂.ss_list_aux_SS),
+                            𝓂.par_list_aux_SS),
+                    union.(𝓂.ss_calib_list,
+                            𝓂.par_calib_list))
+
+    for i in 1:length(unknowns)
+        for k in 1:length(unknowns)
+            incidence_matrix[i,k] = collect(unknowns)[i] ∈ collect(eq_list)[k]
+        end
+    end
+
+    Q, P, R, nmatch, n_blocks = BlockTriangularForm.order(sparse(incidence_matrix))
+    R̂ = []
+    for i in 1:n_blocks
+        [push!(R̂, n_blocks - i + 1) for ii in R[i]:R[i+1] - 1]
+    end
+    push!(R̂,1)
+
+    vars = hcat(P, R̂)'
+    eqs = hcat(Q, R̂)'
+
+    # @assert all(eqs[1,:] .> 0) "Could not solve system of steady state and calibration equations for: " * repr([collect(Symbol.(unknowns))[vars[1,eqs[1,:] .< 0]]...]) # repr([vcat(𝓂.ss_equations,𝓂.calibration_equations)[-eqs[1,eqs[1,:].<0]]...])
[…]
