### 2023-07-23 c95bedb
add aqua tests back in
