### 2023-07-23 7da2496
clean up SymPyCall
