+++ src/SymPyCall/assumptions.jl
@@ -1,378 +1,0 @@
-## Assumptions
-## http://docs.sympy.org/0.7.2/modules/assumptions/index.html
-
-"""
-     refine
-
-Simplify an expression using assumptions; [refine](https://docs.sympy.org/dev/modules/assumptions/refine.html).
-"""
-refine(ex, assumpts...) = sympy.refine(ex, assumpts...)
-
-"""
-    ask(query)
-
-Returns `true`, `false`, or `nothing`; [ask](https://docs.sympy.org/dev/modules/assumptions/ask.html)
-
-Example:
-
-```jldoctest ask
-julia> using SymPyCall
-
-julia> @syms x::integer y::integer
-(x, y)
-
-julia> ask(𝑄.integer(x*y), And(𝑄.integer(x), 𝑄.integer(y)))
-ERROR: UndefVarError: And not defined
-Stacktrace:
- [1] top-level scope
-   @ none:1
-
-julia> ## really slow isprime:
-       filter(x -> ask(𝑄.prime(x)), 1:10)
-4-element Vector{Int64}:
- 2
- 3
- 5
- 7
-```
-
-"""
-function ask(x::Sym, args...)
-    val = sympy.ask(x, args...)
-    Bool(Py(val) == pybuiltins.None) && return nothing
-    Bool(Py(val) == pybuiltins.True) && return true
-    Bool(Py(val) == pybuiltins.False) && return false
-    throw(DomainError("Argument not none, true, or false"))
-end
-
-ask(x::Bool, args...) = x
-ask(x::Nothing, args...) = x
-
-
-
-
-# ## should we support & and | for (sym,sym) pairs? Not sure
-# ## dependso on what happens to | (x==0) and ex == x usage for ==
-# ## for now, we can combine terms logically with And, Or, Not...
-# ## these are in logic module
-
-
-
-# ## We make a module Q to hold the assumptions
-# ## this follows this page http://docs.sympy.org/0.7.5/_modules/sympy/assumptions/ask.html
-# """
-#     𝑄 [\itQ]
-#     SymPy.Q
-
-# Documentation for the `SymPy.Q` module, exported as `𝑄`.
-
-# SymPy allows for
-# [assumptions](https://docs.sympy.org/latest/modules/assumptions/index.html)
-# on variables. These may be placed on free sympols at construction.
-
-# For example, the following creates a real value variable `x` and a postive, real variable `y`:
-
-# ```jldoctest 𝑄
-# julia> using SymPy
-
-# julia> @syms x::real
-# (x,)
-
-# julia> @syms y::(real,positive)
-# (y,)
-# ```
-
-# The `Q` module exposes a means to *q*uery the assumptions on a
-# variable. For example,
-
-# ```jldoctest 𝑄
-# julia> ask(𝑄.positive(y))  # true
-# true
-
-# julia> ask(𝑄.negative(y))  # false
-# false
-
-# julia> ask(SymPyCall.Q.positive(x))  # `nothing`
-
-# julia> ask(SymPyCall..Q.positive(x^2)) # `nothing` -- might be 0
-
-# julia> ask(SymPyCall.Q.positive(1 + x^2)) # true  -- must be postive now.
-# true
-# ```
-
-# The ask function uses tri-state logic, returning one of 3 values:
-# `true`; `false`; or `nothing`, when the query is indeterminate.
-
-# The construction of predicates is done through `Q` methods. These can
-# be combined logically. For example, this will be `true`:
-
-# ```jldoctest 𝑄
-# julia> ask(𝑄.positive(y) & 𝑄.negative(-x^2 - 1))
-
-# ```
-
-# The above use `&` as an infix operation for the binary operator
[…]
