+++ Project.toml
@@ -1,7 +1,7 @@
 name = "MacroModelling"
 uuid = "687ffad2-3618-405e-ac50-e0f7b9c75e44"
 authors = ["Thore Kockerols <mail@thorekockerols.com>"]
-version = "0.1.24"
+version = "0.1.25"
 
 [deps]
 AxisKeys = "94b1ba4f-4ee9-5380-92f1-94cde586c3c5"
