### 2023-07-23 thorek1/MacroModelling.jl#33 @thorek1
and I got that fixed now as well. I'm not relying on python during recompilation after all.

have a look at the latest version. it should run out of the box. it does in the CI runs at least.

and thanks for the suggestion to use PythonCall that should make it easier for many users to get this up and running
