### 2023-07-23 thorek1/MacroModelling.jl#33 @gdalle
Awesome! I'll check it out and finish my review next week
