### 2023-07-23 thorek1/MacroModelling.jl#36 @thorek1
I added a few lines on DSEG models. The first two paragraphs of `paper.md` in the `joss` branch now read:

`MacroModelling.jl is a Julia [@Bezanson:2012] package for developing and solving dynamic stochastic general equilibrium (DSGE) models. _These kind of models are typicaly used to decsribe the behaviour of a macroeconomy and are particularly suited for counterfactual analysis (economic policy evaluation) and exploring/quantifying specific mechanisms (academic research). These models are difficult to work with because they consist of a nonlinear system of equations describing a stochastic control problem._

The goal of this package is to reduce coding time and speed up model development by providing functions for creating, calibrating, simulating, and estimating discrete-time DSGE models. The package includes several pre-defined models from prominent economic papers, providing an immediate starting point for policy makers, economic researchers, and students.`
