### 2023-07-23 a466e89
add bib and other packages to docs
