### 2023-07-23 thorek1/MacroModelling.jl#38 @thorek1
usually the user derives the steady state conditions by hand or hopes that a numerical solver finds it automatically (which is far from being guaranteed)

what MacroModelling does different is that it used symbolic manipulation to reduce the steady state problem into smaller separate subproblems and then it uses effective algorithms (as in they find it, not they find it faster than other algorithms) to find the steady state. the algorithm uses a range of tricks, from search space transformation, automatic domain restrictions, restarts with different initial values, warm starts using previous solutions. these are not new in isolation but in this case they work very well in combination. by that I mean that they solve all 14 models implemented with MacroModelling out of the box without any starting point guess

The solver is a paper on it's own I think. Anyway, you think some of this info should be included?

@benchmarking: that's a whole other dimension I don't focus on here yet. It will become relevant at later stages of package development
