+++ src/MacroModelling.jl
@@ -15,9 +15,10 @@ import LinearAlgebra as ℒ
 import ComponentArrays as 𝒞
 import BlockTriangularForm
 import Subscripts: super, sub
-import IterativeSolvers as ℐ
+# import IterativeSolvers as ℐ
+import Krylov, LinearOperators
 import DataStructures: CircularBuffer
-using LinearMaps
+# using LinearMaps
 using ImplicitDifferentiation
 import SpeedMapping: speedmapping
 # import NLboxsolve: nlboxsolve
@@ -215,7 +216,7 @@ function levenberg_marquardt(f::Function,
     lower_bounds::Array{T,1}, 
     upper_bounds::Array{T,1}; 
     xtol::T = eps(), 
-    ftol::T = 1e-10, 
+    ftol::T = eps(), 
     iterations::S = 250, 
     ϕ̄::T    =       8.0,
     ϕ̂::T    =       0.904,
@@ -1066,7 +1067,7 @@ block_solver_AD(parameters_and_solved_vars::Vector{<: Real},
     guess::Vector{Float64}, 
     lbs::Vector{Float64}, 
     ubs::Vector{Float64};
-    tol::AbstractFloat = eps(Float64),
+    tol::AbstractFloat = eps(),
     # timeout = 120,
     starting_points::Vector{Float64} = [0.897, 1.2, .9, .75, 1.5, -.5, 2.0, .25],
     # fail_fast_solvers_only = true,
@@ -1212,7 +1213,7 @@ end
 
 
 function second_order_stochastic_steady_state_iterative_solution(𝐒₁𝐒₂::AbstractArray{Float64}, 𝓂::ℳ, pruning::Bool;
-    tol::AbstractFloat = 1e-10)
+    tol::AbstractFloat = eps())
     (; 𝐒₁, 𝐒₂) = 𝐒₁𝐒₂
 
     state = zeros(𝓂.timings.nVars)
@@ -1319,7 +1320,7 @@ function calculate_second_order_stochastic_steady_state(parameters::Vector{M}, �
     
     ∇₂ = calculate_hessian(parameters, SS_and_pars, 𝓂)
     
-    𝐒₂ = calculate_second_order_solution(∇₁, ∇₂, 𝐒₁; T = 𝓂.timings)
+    𝐒₂ = calculate_second_order_solution(∇₁, ∇₂, 𝐒₁, 𝓂.solution.perturbation.second_order_auxilliary_matrices; T = 𝓂.timings)
 
     𝐒₁ = [𝐒₁[:,1:𝓂.timings.nPast_not_future_and_mixed] zeros(𝓂.timings.nVars) 𝐒₁[:,𝓂.timings.nPast_not_future_and_mixed+1:end]]
 
@@ -1343,7 +1344,7 @@ end
 
 
 function third_order_stochastic_steady_state_iterative_solution(𝐒₁𝐒₂𝐒₃::AbstractArray{Float64}, 𝓂::ℳ, pruning::Bool;
-    tol::AbstractFloat = 1e-10)
+    tol::AbstractFloat = eps())
     (; 𝐒₁, 𝐒₂, 𝐒₃) = 𝐒₁𝐒₂𝐒₃
 
     state = zeros(𝓂.timings.nVars)
@@ -1442,7 +1443,7 @@ function third_order_stochastic_steady_state_iterative_solution(𝐒₁𝐒₂�
 end
 
 
[…]

[…]
