### 2023-07-25 7f95833
rm compats for weakdeps; no check by compathelper
