+++ test/runtests.jl
@@ -11,8 +11,8 @@ include("functionality_tests.jl")
 @testset verbose = true "Code quality (Aqua.jl)" begin
     # Aqua.test_all(MacroModelling)
     @testset "Compare Project.toml and test/Project.toml" Aqua.test_project_extras(MacroModelling)
-    # @testset "Project.toml formatting" Aqua.test_project_toml_formatting(MacroModelling)
-    @testset "Stale dependencies" Aqua.test_stale_deps(MacroModelling)# fix this to handle CondaPkg not being loaded/considered stale
+    @testset "Project.toml formatting" Aqua.test_project_toml_formatting(MacroModelling)
+    @testset "Stale dependencies" Aqua.test_stale_deps(MacroModelling)#;ignore = [:Aqua])# fix this to handle CondaPkg not being loaded/considered stale
     @testset "Unbound type parameters" Aqua.test_unbound_args(MacroModelling)
     @testset "Undefined exports" Aqua.test_undefined_exports(MacroModelling)
     @testset "Piracy" Aqua.test_piracy(MacroModelling)
