### 2023-07-25 c3eaea6
pass through tolerance to sylvester solver
