### 2023-07-25 3f77005
fix standalone tests
