### 2023-07-25 83303c4
analytical conditions first attempt
