+++ test/test_Krylov_solver_and_autodiff.jl
@@ -1,0 +1,440 @@
+using MacroModelling
+using SparseArrays
+import LinearAlgebra as ℒ
+import LinearOperators
+import Krylov
+import ForwardDiff as ℱ
+import RecursiveFactorization as RF
+
+include("models/FS2000.jl")
+
+using FiniteDifferences
+SSS(m,derivatives = false)[1]
+pars = copy(m.parameter_values)
+fin_grad = FiniteDifferences.grad(central_fdm(5,1),x->SSS(m,derivatives = false, parameters = x)[1],pars)[1]
+SSS(m)[1,2:end]
+
+SSS(m)
+
+
+include("models/RBC_CME.jl")
+
+
+get_solution(m,algorithm = :pruned_second_order)
+
+parameters = m.parameter_values
+𝓂 = m
+verbose = true
+
+SS_and_pars, solution_error = 𝓂.SS_solve_func(parameters, 𝓂, verbose)
+    
+∇₁ = calculate_jacobian(parameters, SS_and_pars, 𝓂)
+
+𝑺₁, solved = calculate_first_order_solution(∇₁; T = 𝓂.timings)
+
+∇₂ = calculate_hessian(parameters, SS_and_pars, 𝓂)
+
+
+M₂ = 𝓂.solution.perturbation.second_order_auxilliary_matrices
+T = 𝓂.timings
+tol = eps()
+
+
+
+# Indices and number of variables
+i₊ = T.future_not_past_and_mixed_idx;
+i₋ = T.past_not_future_and_mixed_idx;
+
+n₋ = T.nPast_not_future_and_mixed
+n₊ = T.nFuture_not_past_and_mixed
+nₑ = T.nExo
+n  = T.nVars
+nₑ₋ = n₋ + 1 + nₑ
+
+# 1st order solution
+𝐒₁ = @views [𝑺₁[:,1:n₋] zeros(n) 𝑺₁[:,n₋+1:end]] |> sparse
+droptol!(𝐒₁,tol)
+
+𝐒₁₋╱𝟏ₑ = @views [𝐒₁[i₋,:]; zeros(nₑ + 1, n₋) spdiagm(ones(nₑ + 1))[1,:] zeros(nₑ + 1, nₑ)];
+
+⎸𝐒₁𝐒₁₋╱𝟏ₑ⎹╱𝐒₁╱𝟏ₑ₋ = @views [(𝐒₁ * 𝐒₁₋╱𝟏ₑ)[i₊,:]
+                            𝐒₁
+                            spdiagm(ones(nₑ₋))[[range(1,n₋)...,n₋ + 1 .+ range(1,nₑ)...],:]];
+
+𝐒₁₊╱𝟎 = @views [𝐒₁[i₊,:]
+                zeros(n₋ + n + nₑ, nₑ₋)];
+
+
+∇₁₊𝐒₁➕∇₁₀ = @views -∇₁[:,1:n₊] * 𝐒₁[i₊,1:n₋] * ℒ.diagm(ones(n))[i₋,:] - ∇₁[:,range(1,n) .+ n₊]
+
+spinv = sparse(inv(∇₁₊𝐒₁➕∇₁₀))
[…]
