+++ src/MacroModelling.jl
@@ -9,7 +9,7 @@ import PythonCall
 import Symbolics
 import ForwardDiff as ℱ 
 # import Zygote
-import SparseArrays: SparseMatrixCSC#, sparse, spzeros, droptol!, sparsevec, spdiagm, findnz#, sparse!
+import SparseArrays: SparseMatrixCSC, SparseVector#, sparse, spzeros, droptol!, sparsevec, spdiagm, findnz#, sparse!
 import LinearAlgebra as ℒ
 import ComponentArrays as 𝒞
 import BlockTriangularForm
@@ -2983,57 +2983,76 @@ end
 
 
 
-function solve_sylvester_equation_condition(BCX, S)
-    (; B, C, X) = BCX
 
-    X + S - B * S * C
-end
+function solve_sylvester_equation(concat_sparse_vec::SparseVector{Float64}; dims::Vector{Tuple{Int,Int}}, tol::AbstractFloat = eps())
+    lenA = dims[1][1] * dims[1][2]
+    lenB = dims[2][1] * dims[2][2]
+    lenX = dims[3][1] * dims[3][2]
 
+    A = sparse(reshape(concat_sparse_vec[1 : lenA],dims[1]))
+    B = sparse(reshape(concat_sparse_vec[lenA .+ (1 : lenB)],dims[2]))
 
-function solve_sylvester_equation(BCX::AbstractArray{Float64}; tol::AbstractFloat = 1e-10)
-    (; B, C, X) = BCX
-
-    sylvester = LinearOperators.LinearOperator(Float64, length(X), length(X), false, false, 
-    (sol,𝐱) -> begin 
-        𝐗 = sparse(reshape(𝐱, size(X)))
-        sol .= vec(𝐗 - B * 𝐗 * C)
+    sylvester = LinearOperators.LinearOperator(Float64, lenX, lenX, false, false, 
+        (sol,𝐱) -> begin 
+        𝐗 = reshape(𝐱, dims[3])
+        sol .= vec(𝐗 - A * 𝐗 * B)
         return sol
     end)
 
-    S2, info = Krylov.bicgstab(sylvester, vec(-X))
+    X, info = Krylov.bicgstab(sylvester, concat_sparse_vec[lenA + lenB + 1 : end])
 
     if !info.solved
-        S2, info = Krylov.gmres(sylvester, vec(-X))
+        X, info = Krylov.gmres(sylvester, concat_sparse_vec[lenA + lenB + 1 : end])
     end
 
-    𝐒₂ = sparse(reshape(S2,size(X)))
-    droptol!(𝐒₂,tol)
+    X̂ = sparse(reshape(X,dims[3]))
+    droptol!(X̂, tol)
 
-    return 𝐒₂
+    return X̂
 end
 
 
-function solve_sylvester_equation(BCX::AbstractArray{ℱ.Dual{Z,S,N}}; tol::AbstractFloat = 1e-10) where {Z,S,N}
+function solve_sylvester_equation_conditions(concat_sparse_vec::SparseVector, x::SparseMatrixCSC; dims::Vector{Tuple{Int,Int}})
+    lenA = dims[1][1] * dims[1][2]
+    lenB = dims[2][1] * dims[2][2]
+
+    A = sparse(reshape(concat_sparse_vec[1 : lenA],dims[1]))
+    B = sparse(reshape(concat_sparse_vec[lenA .+ (1 : lenB)],dims[2]))
+    X = sparse(reshape(concat_sparse_vec[lenA + lenB + 1 : end],dims[3]))
+
+    collect(X + x - A * x * B)
+end
+
+
[…]
