+++ Project.toml
@@ -7,6 +7,7 @@ version = "0.1.25"
 AxisKeys = "94b1ba4f-4ee9-5380-92f1-94cde586c3c5"
 BlockTriangularForm = "adeb47b7-70bf-415a-bb24-c358563e873a"
 ChainRulesCore = "d360d2e6-b24c-11e9-a2a3-2a2ae2dbcce4"
+Combinatorics = "861a8166-3701-5b0c-9a16-15d98fcdc6aa"
 CommonEq = "3709ef60-1bee-4518-9f2f-acd86f176c50"
 CommonSolve = "38540f10-b2f7-11e9-35d8-d573e4eb0ff2"
 ComponentArrays = "b0b7db55-cfe3-40fc-9ded-d10e2dbeff66"
@@ -14,6 +15,8 @@ CondaPkg = "992eb4ea-22a4-4c89-a5bb-47a3300528ab"
 DataStructures = "864edb3b-99cc-5e75-8d2d-829cb0a9cfe8"
 DocStringExtensions = "ffbed154-4ef7-542d-bbb7-c09d3a79fcae"
 DynarePreprocessor_jll = "23afba7c-24e5-5ee2-bc2c-b42e07f0492a"
+FiniteDiff = "6a86dc24-6348-571c-b903-95158fe2bd41"
+FiniteDifferences = "26cc04aa-876d-5657-8c51-4c34ba976000"
 ForwardDiff = "f6369f11-7733-5829-9624-2563aa707210"
 ImplicitDifferentiation = "57b37032-215b-411a-8a7c-41a003a55207"
 JSON = "682c06a0-de6a-54ab-a142-c8b1cf79cde6"
@@ -34,6 +37,7 @@ SpecialFunctions = "276daf66-3868-5448-9aa4-cd146d93841b"
 SpeedMapping = "f1835b91-879b-4a3f-a438-e4baacf14412"
 Subscripts = "2b7f82d5-8785-4f63-971e-f18ddbeb808e"
 Symbolics = "0c5d862f-8b57-4792-8d23-62f2024744c7"
+TensorOperations = "6aa20fa7-93e2-5fca-9bc0-fbd0db3c71a2"
 
 [weakdeps]
 StatsPlots = "f3b207a7-027a-5e70-b257-86293d7955fd"
