+++ src/MacroModelling.jl
@@ -218,7 +218,7 @@ function levenberg_marquardt(f::Function,
     lower_bounds::Array{T,1}, 
     upper_bounds::Array{T,1}; 
     xtol::T = eps(), 
-    ftol::T = 1e-10, 
+    ftol::T = eps(), 
     iterations::S = 250, 
     ϕ̄::T    =       8.0,
     ϕ̂::T    =       0.904,
@@ -2984,13 +2984,25 @@ end
 
 
 
-function solve_sylvester_equation(concat_sparse_vec::SparseVector{Float64}; dims::Vector{Tuple{Int,Int}}, tol::AbstractFloat = eps())
+function solve_sylvester_equation_conditions(concat_sparse_vec::AbstractVector, x::AbstractArray; dims::Vector{Tuple{Int,Int}})
+    lenA = dims[1][1] * dims[1][2]
+    lenB = dims[2][1] * dims[2][2]
+
+    A = (reshape(concat_sparse_vec[1 : lenA],dims[1]))
+    B = (reshape(concat_sparse_vec[lenA .+ (1 : lenB)],dims[2]))
+    X = (reshape(concat_sparse_vec[lenA + lenB + 1 : end],dims[3]))
+
+    collect(X + x - A * x * B)
+end
+
+
+function solve_sylvester_equation(concat_sparse_vec::AbstractVector{Float64}; dims::Vector{Tuple{Int,Int}}, tol::AbstractFloat = eps())
     lenA = dims[1][1] * dims[1][2]
     lenB = dims[2][1] * dims[2][2]
     lenX = dims[3][1] * dims[3][2]
 
-    A = sparse(reshape(concat_sparse_vec[1 : lenA],dims[1]))
-    B = sparse(reshape(concat_sparse_vec[lenA .+ (1 : lenB)],dims[2]))
+    A = (reshape(concat_sparse_vec[1 : lenA],dims[1]))
+    B = (reshape(concat_sparse_vec[lenA .+ (1 : lenB)],dims[2]))
 
     sylvester = LinearOperators.LinearOperator(Float64, lenX, lenX, false, false, 
         (sol,𝐱) -> begin 
@@ -2999,32 +3011,20 @@ function solve_sylvester_equation(concat_sparse_vec::SparseVector{Float64}; dims
         return sol
     end)
 
-    X, info = Krylov.bicgstab(sylvester, concat_sparse_vec[lenA + lenB + 1 : end])
+    X, info = Krylov.gmres(sylvester, -concat_sparse_vec[lenA + lenB + 1 : end])#, atol = tol)
 
     if !info.solved
-        X, info = Krylov.gmres(sylvester, concat_sparse_vec[lenA + lenB + 1 : end])
+        X, info = Krylov.bicgstab(sylvester, -concat_sparse_vec[lenA + lenB + 1 : end])#, atol = tol)
     end
 
-    X̂ = sparse(reshape(X,dims[3]))
-    droptol!(X̂, tol)
+    X̂ = (reshape(X,dims[3]))
+    # droptol!(X̂, tol)
 
     return X̂
 end
 
 
-function solve_sylvester_equation_conditions(concat_sparse_vec::SparseVector, x::SparseMatrixCSC; dims::Vector{Tuple{Int,Int}})
-    lenA = dims[1][1] * dims[1][2]
-    lenB = dims[2][1] * dims[2][2]
-
-    A = sparse(reshape(concat_sparse_vec[1 : lenA],dims[1]))
-    B = sparse(reshape(concat_sparse_vec[lenA .+ (1 : lenB)],dims[2]))
-    X = sparse(reshape(concat_sparse_vec[lenA + lenB + 1 : end],dims[3]))
-
-    collect(X + x - A * x * B)
-end
-
-
[…]
