### 2023-07-27 d41d57b
use analytical jacobian sylvester equation
