+++ src/MacroModelling.jl
@@ -3031,8 +3031,8 @@ function solve_sylvester_equation(concat_sparse_vec::AbstractVector{ℱ.Dual{Z,S
     lenA = dims[1][1] * dims[1][2]
     lenB = dims[2][1] * dims[2][2]
 
-    A = (reshape(concat_sparse_vec_values[1 : lenA],dims[1]))
-    B = (reshape(concat_sparse_vec_values[lenA .+ (1 : lenB)],dims[2]))
+    A = sparse(reshape(concat_sparse_vec_values[1 : lenA],dims[1]))
+    B = sparse(reshape(concat_sparse_vec_values[lenA .+ (1 : lenB)],dims[2]))
 
     # you can play with the dimension here, sometimes it makes sense to transpose
     ps = mapreduce(ℱ.partials, hcat, concat_sparse_vec)'
@@ -3041,23 +3041,22 @@ function solve_sylvester_equation(concat_sparse_vec::AbstractVector{ℱ.Dual{Z,S
     val = solve_sylvester_equation(concat_sparse_vec_values, dims = dims, tol = tol)
 
     # get J(f, vs) * ps (cheating). Write your custom rule here. This used to be the conditions but here they are analytically derived.
-    b = ℱ.jacobian(x -> solve_sylvester_equation_conditions(x, val, dims = dims), concat_sparse_vec_values)
-    a = ℱ.jacobian(x -> solve_sylvester_equation_conditions(concat_sparse_vec_values, x, dims = dims), val)
-    # b = hcat(ℒ.kron(-val * B, ℒ.I(size(A,1)))', ℒ.kron(ℒ.I(size(B,1)), A * val), ℒ.I(length(val)))
-    # a = reshape(permutedims(reshape(ℒ.I - ℒ.kron(A, B) ,size(B,1), size(A,1), size(A,1), size(B,1)), [2, 3, 4, 1]), size(A,1) * size(B,1), size(A,1) * size(B,1))
+    A¹ = sparse(size(A,1) * (findnz(A)[2].-1) .+ findnz(A)[1],size(A,1) * (findnz(A)[2].-1) .+ findnz(A)[1],1,length(A),length(A))
+    B¹ = sparse(size(B,1) * (findnz(B)[2].-1) .+ findnz(B)[1],size(B,1) * (findnz(B)[2].-1) .+ findnz(B)[1],1,length(B),length(B))
 
-    Â = RF.lu(collect(a), check = false)
+    jacobian_A = A¹ * ℒ.kron(-val * B, ℒ.I(size(A,1)))
+    jacobian_B = ℒ.kron(ℒ.I(size(B,1)), -A * val) * B¹
+
+    b = hcat(jacobian_A', jacobian_B, ℒ.I(length(val)))
+    a = reshape(permutedims(reshape(ℒ.I - ℒ.kron(A, B) ,size(B,1), size(A,1), size(A,1), size(B,1)), [2, 3, 4, 1]), size(A,1) * size(B,1), size(A,1) * size(B,1))
+
+    Â = RF.lu(a, check = false)
 
     if !ℒ.issuccess(Â)
-        Â = ℒ.svd(collect(a))
+        Â = ℒ.svd(a)
     end
     
     jvp = -(Â \ b) * ps
-
-    # lm = LinearMap{Float64}(x -> A * reshape(x, size(B)), length(B))
-
-    # jvp = - sparse(reshape(ℐ.gmres(lm, sparsevec(B)), size(B))) * ps
-    # jvp *= -ps
 
     # pack: SoA -> AoS
     return reshape(map(val, eachrow(jvp)) do v, p
@@ -3788,74 +3787,74 @@ end
 
 
 
-@setup_workload begin
-    # Putting some things in `setup` can reduce the size of the
-    # precompile file and potentially make loading faster.
[…]
