### 2023-07-27 4718965
solve with krylov; separate sparsevec dual
