+++ src/MacroModelling.jl
@@ -2983,20 +2983,7 @@ end
 
 
 
-
-function solve_sylvester_equation_conditions(concat_sparse_vec::AbstractVector, x::AbstractArray; dims::Vector{Tuple{Int,Int}})
-    lenA = dims[1][1] * dims[1][2]
-    lenB = dims[2][1] * dims[2][2]
-
-    A = (reshape(concat_sparse_vec[1 : lenA],dims[1]))
-    B = (reshape(concat_sparse_vec[lenA .+ (1 : lenB)],dims[2]))
-    X = (reshape(concat_sparse_vec[lenA + lenB + 1 : end],dims[3]))
-
-    collect(X + x - A * x * B)
-end
-
-
-function solve_sylvester_equation(concat_sparse_vec::AbstractVector{Float64}; dims::Vector{Tuple{Int,Int}}, tol::AbstractFloat = eps())
+function solve_sylvester_equation(concat_sparse_vec::SparseVector{Float64}; dims::Vector{Tuple{Int,Int}}, tol::AbstractFloat = eps())
     lenA = dims[1][1] * dims[1][2]
     lenB = dims[2][1] * dims[2][2]
     lenX = dims[3][1] * dims[3][2]
@@ -3023,26 +3010,47 @@ function solve_sylvester_equation(concat_sparse_vec::AbstractVector{Float64}; di
     return X̂
 end
 
+function separate_values_and_partials_from_sparsevec_dual(V::SparseVector{ℱ.Dual{Z,S,N}}; tol::AbstractFloat = eps()) where {Z,S,N}
+    nrows = length(V)
+    ncols = length(V.nzval[1].partials)
 
-function solve_sylvester_equation(concat_sparse_vec::AbstractVector{ℱ.Dual{Z,S,N}}; dims::Vector{Tuple{Int,Int}}, tol::AbstractFloat = eps()) where {Z,S,N}
+    rows = Int[]
+    cols = Int[]
+
+    prtls = Float64[]
+
+    for (i,v) in enumerate(V.nzind)
+        for (k,w) in enumerate(V.nzval[i].partials)
+            if abs(w) > tol
+                push!(rows,v)
+                push!(cols,k)
+                push!(prtls,w)
+            end
+        end
+    end
+
+    vvals = sparsevec(V.nzind,[i.value for i in V.nzval],nrows)
+    ps = sparse(rows,cols,prtls,nrows,ncols)
+
+    return vvals, ps
+end
+
+function solve_sylvester_equation(concat_sparse_vec::SparseVector{ℱ.Dual{Z,S,N}}; dims::Vector{Tuple{Int,Int}}, tol::AbstractFloat = eps()) where {Z,S,N}
     # unpack: AoS -> SoA
-    concat_sparse_vec_values = ℱ.value.(concat_sparse_vec)
+    values, partials = separate_values_and_partials_from_sparsevec_dual(concat_sparse_vec)
 
     lenA = dims[1][1] * dims[1][2]
     lenB = dims[2][1] * dims[2][2]
 
-    A = sparse(reshape(concat_sparse_vec_values[1 : lenA],dims[1]))
-    B = sparse(reshape(concat_sparse_vec_values[lenA .+ (1 : lenB)],dims[2]))
-
-    # you can play with the dimension here, sometimes it makes sense to transpose
[…]
