### 2023-07-27 e831fc3
rm ComponentArrays; sparsevec handover
