+++ src/MacroModelling.jl
@@ -11,7 +11,7 @@ import ForwardDiff as ℱ
 # import Zygote
 import SparseArrays: SparseMatrixCSC, SparseVector#, sparse, spzeros, droptol!, sparsevec, spdiagm, findnz#, sparse!
 import LinearAlgebra as ℒ
-import ComponentArrays as 𝒞
+# import ComponentArrays as 𝒞
 import BlockTriangularForm
 import Subscripts: super, sub
 import Krylov
@@ -1570,9 +1570,12 @@ end
 
 
 
-function second_order_stochastic_steady_state_iterative_solution(𝐒₁𝐒₂::AbstractArray{Float64}, 𝓂::ℳ, pruning::Bool;
-    tol::AbstractFloat = 1e-10)
-    (; 𝐒₁, 𝐒₂) = 𝐒₁𝐒₂
+function second_order_stochastic_steady_state_iterative_solution(𝐒₁𝐒₂::SparseVector{Float64}, dims::Vector{Tuple{Int,Int}}, 𝓂::ℳ, pruning::Bool;
+    tol::AbstractFloat = 1e-10)    
+    len𝐒₁ = dims[1][1] * dims[1][2]
+
+    𝐒₁ = reshape(𝐒₁𝐒₂[1 : len𝐒₁],dims[1])
+    𝐒₂ = sparse(reshape(𝐒₁𝐒₂[len𝐒₁ + 1 : end],dims[2]))
 
     state = zeros(𝓂.timings.nVars)
     shock = zeros(𝓂.timings.nExo)
@@ -1609,8 +1612,11 @@ function second_order_stochastic_steady_state_iterative_solution(𝐒₁𝐒₂:
 end
 
 
-function second_order_stochastic_steady_state_iterative_solution_condition(𝐒₁𝐒₂, SSS, 𝓂::ℳ, pruning::Bool)
-    (; 𝐒₁, 𝐒₂) = 𝐒₁𝐒₂
+function second_order_stochastic_steady_state_iterative_solution_condition(𝐒₁𝐒₂::SparseVector, SSS, dims::Vector{Tuple{Int,Int}}, 𝓂::ℳ, pruning::Bool)
+    len𝐒₁ = dims[1][1] * dims[1][2]
+
+    𝐒₁ = reshape(𝐒₁𝐒₂[1 : len𝐒₁],dims[1])
+    𝐒₂ = sparse(reshape(𝐒₁𝐒₂[len𝐒₁ + 1 : end],dims[2]))
 
     shock = zeros(𝓂.timings.nExo)
 
@@ -1630,21 +1636,16 @@ function second_order_stochastic_steady_state_iterative_solution_condition(𝐒�
 end
 
 
-function second_order_stochastic_steady_state_iterative_solution(𝐒₁𝐒₂::AbstractArray{ℱ.Dual{Z,S,N}}, 𝓂::ℳ, pruning::Bool) where {Z,S,N}
-
-    # unpack: AoS -> SoA
-    S₁S₂ = ℱ.value.(𝐒₁𝐒₂)
-
-    # you can play with the dimension here, sometimes it makes sense to transpose
-    ps = mapreduce(ℱ.partials, hcat, 𝐒₁𝐒₂)'
+function second_order_stochastic_steady_state_iterative_solution(𝐒₁𝐒₂::SparseVector{ℱ.Dual{Z,S,N}}, dims::Vector{Tuple{Int,Int}}, 𝓂::ℳ, pruning::Bool) where {Z,S,N}
+    S₁S₂, ps = separate_values_and_partials_from_sparsevec_dual(𝐒₁𝐒₂)
 
     # get f(vs)
[…]
