+++ Project.toml
@@ -10,7 +10,6 @@ ChainRulesCore = "d360d2e6-b24c-11e9-a2a3-2a2ae2dbcce4"
 Combinatorics = "861a8166-3701-5b0c-9a16-15d98fcdc6aa"
 CommonEq = "3709ef60-1bee-4518-9f2f-acd86f176c50"
 CommonSolve = "38540f10-b2f7-11e9-35d8-d573e4eb0ff2"
-ComponentArrays = "b0b7db55-cfe3-40fc-9ded-d10e2dbeff66"
 CondaPkg = "992eb4ea-22a4-4c89-a5bb-47a3300528ab"
 DataStructures = "864edb3b-99cc-5e75-8d2d-829cb0a9cfe8"
 DocStringExtensions = "ffbed154-4ef7-542d-bbb7-c09d3a79fcae"
@@ -49,7 +48,6 @@ BlockTriangularForm = "^0.1"
 ChainRulesCore = "^1"
 CommonEq = "^0.2"
 CommonSolve = "^0.2"
-ComponentArrays = "^0.13, 0.14"
 CondaPkg = "^0.2"
 DataStructures = "^0.18"
 DocStringExtensions = "^0.8, 0.9"
