### 2023-07-28 thorek1/MacroModelling.jl#40 @gdalle
CI tests still failing
This time it seems to be because of our crappy ImplicitDifferentiation interface

I'm working on it today, hopefully @mohamed82008 and I can issue a new release this weekend with a simpler API

https://github.com/openjournals/joss-reviews/issues/5598
