### 2023-07-28 thorek1/MacroModelling.jl#40 @thorek1
The CI works for the main branch.

I was trying in another branch to get ImplicitDifferentiation >= 0.4 to work but that failed.

I'll post a new issue about it.
