### 2023-07-28 thorek1/MacroModelling.jl#40 @gdalle
Honestly don't bother switching from 0.3 to 0.4, wait for 0.5 instead
