### 2023-07-28 thorek1/MacroModelling.jl#40 @gdalle
We're reverting the byproduct api to make it transparent to users who don't need it
