+++ test/test_Krylov_solver_and_autodiff.jl
@@ -6,10 +6,101 @@ import ForwardDiff as ℱ
 import RecursiveFactorization as RF
 
 using MacroModelling
-include("models/FS2000.jl")
+# include("models/FS2000.jl") # 42ms
+include("models/RBC_CME.jl") # 6/7ms vs 5.5 with dense lu
+# include("models/SW03.jl")
+# m = SW03
+# var(m)
+
+# SSS(m)
+SSS(m, algorithm = :pruned_third_order)
 
 
-SSS(m)
+
+𝑺₁̂ = [0.9 8.326672684688674e-17 -1.7383299474834864e-17 0.0 0.0068; 0.022380105873521232 -0.0036490225028059124 0.0012133617511768497 6.740898617649275e-6 0.00016909413326660324; 0.033603762572854703 -0.0054790127669757955 0.0018218644912565338 1.0121469395869782e-5 0.00025389509499489977; 0.2874801013640593 0.049647049058691016 -0.06603393685875958 -0.0003668552047708867 0.0021720718769728947; 0.9934103445024477 0.9513539519423103 -0.12653651724319917 -0.0007029806513511057 0.007505767047351817; 1.280890445866506 0.023601001001001085 -2.4667062706664246e-17 -0.0 0.009677838924324712; 0.0 0.0 0.9 0.005 -0.0]
+
+
+partials = [0.0 0.0 0.0 0.0 0.0 1.0 0.0 0.0 0.0; 0.05560594678824912 -0.6657652463325195 0.6585458406236359 0.06708664830192332 -0.026359877980302714 0.11411497174354064 0.0 -1.3814719098180438e-17 0.0; 0.08349241259496858 -1.033284916988622 0.9888075685039596 0.10073070315604102 -0.01717688798491776 0.1713438014167277 0.0 -2.0742821468739397e-17 0.0; -0.222810340666038 -9.490434885884481 5.948453384557047 0.0 0.0 2.0398230489095726 0.0 5.234286205429912e-16 0.0; 5.157767754557424 19.618410903519067 -16.05618357813237 0.0 0.0 -0.6166114423912324 0.0 -5.234286205429912e-16 0.0; 4.934957413891384 10.127976017634586 -10.107730193575334 0.0 0.0 1.42321160651834 0.0 0.0 0.0; 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0; 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0; 0.03944870826911845 0.3171516302456502 -0.3168194064105162 -0.010938316854933442 0.006662631443726296 3.832339955109047e-17 0.0 -1.1375292606400344e-17 0.0; 0.059232294698376035 0.48168814628173273 -0.47570481443020396 -0.016423899181581745 0.006351275938722254 5.754264196860431e-17 0.0 -1.7080018928529048e-17 0.0; -0.20796878552454498 -2.7083671490090824 2.021800136505072 0.0 0.0 1.6609394112485375e-15 0.0 8.018674143532576e-17 0.0; 0.2079687855245455 1.7063641450040778 -2.0218001365050733 0.0 0.0 -1.6609394112485375e-15 0.0 -8.018674143532576e-17 0.0; 5.398420057174561e-16 -1.002003004
[…]
