### 2023-07-28 thorek1/MacroModelling.jl#40 @thorek1
Ok, I posted two issues, one related to byproducts. Solving those two would be very helpful. Currently I'm mostly working with my own implementations for ForwardDiff Duals.
