### 2023-07-28 c9f9ef0
add Baxter and King 1993 model + tests
