### 2023-07-29 b76318f
take into account byproduct
