### 2023-07-30 bb4e371
precompile kalman
