### 2023-07-30 6648ba8
return 0 for conditions if SS not found
