+++ src/get_functions.jl
@@ -1496,7 +1496,9 @@ function get_variance_decomposition(𝓂::ℳ;
 
     sol, solved = calculate_first_order_solution(∇₁; T = 𝓂.timings)
 
-    variances_by_shock = reduce(hcat,[ℒ.diag(calculate_covariance_forward(sol[:,[1:𝓂.timings.nPast_not_future_and_mixed..., 𝓂.timings.nPast_not_future_and_mixed+i]], T = 𝓂.timings, subset_indices = collect(1:𝓂.timings.nVars))[1]) for i in 1:𝓂.timings.nExo])
+    calculate_covariance = calculate_covariance_AD(sol[:,[1:𝓂.timings.nPast_not_future_and_mixed..., 𝓂.timings.nPast_not_future_and_mixed+i]], T = 𝓂.timings, subset_indices = collect(1:𝓂.timings.nVars))
+
+    variances_by_shock = reduce(hcat,[ℒ.diag(calculate_covariance(sol[:,[1:𝓂.timings.nPast_not_future_and_mixed..., 𝓂.timings.nPast_not_future_and_mixed+i]])[1]) for i in 1:𝓂.timings.nExo])
 
     var_decomp = variances_by_shock ./ sum(variances_by_shock,dims=2)
 
