+++ src/MacroModelling.jl
@@ -1515,65 +1515,65 @@ function block_solver(parameters_and_solved_vars::Vector{Float64},
 end
 
 
-function block_solver(parameters_and_solved_vars::Vector{ℱ.Dual{Z,S,N}}, 
-    n_block::Int, 
-    ss_solve_blocks::Function, 
-    # SS_optimizer, 
-    # f::OptimizationFunction, 
-    guess::Vector{Float64}, 
-    lbs::Vector{Float64}, 
-    ubs::Vector{Float64};
-    tol::AbstractFloat = eps(),
-    # timeout = 120,
-    starting_points::Vector{Float64} = [0.897, 1.2, .9, .75, 1.5, -.5, 2, .25],
-    # fail_fast_solvers_only = true,
-    verbose::Bool = false) where {Z,S,N}
+# function block_solver(parameters_and_solved_vars::Vector{ℱ.Dual{Z,S,N}}, 
+#     n_block::Int, 
+#     ss_solve_blocks::Function, 
+#     # SS_optimizer, 
+#     # f::OptimizationFunction, 
+#     guess::Vector{Float64}, 
+#     lbs::Vector{Float64}, 
+#     ubs::Vector{Float64};
+#     tol::AbstractFloat = eps(),
+#     # timeout = 120,
+#     starting_points::Vector{Float64} = [0.897, 1.2, .9, .75, 1.5, -.5, 2, .25],
+#     # fail_fast_solvers_only = true,
+#     verbose::Bool = false) where {Z,S,N}
 
-    # unpack: AoS -> SoA
-    inp = ℱ.value.(parameters_and_solved_vars)
+#     # unpack: AoS -> SoA
+#     inp = ℱ.value.(parameters_and_solved_vars)
 
-    # you can play with the dimension here, sometimes it makes sense to transpose
-    ps = mapreduce(ℱ.partials, hcat, parameters_and_solved_vars)'
+#     # you can play with the dimension here, sometimes it makes sense to transpose
+#     ps = mapreduce(ℱ.partials, hcat, parameters_and_solved_vars)'
 
-    if verbose println("Solution for derivatives.") end
-    # get f(vs)
-    val, min = block_solver(inp, 
-                        n_block, 
-                        ss_solve_blocks, 
-                        # SS_optimizer, 
-                        # f, 
-                        guess, 
-                        lbs, 
-                        ubs;
-                        tol = tol,
-                        # timeout = timeout,
-                        starting_points = starting_points,
-                        # fail_fast_solvers_only = fail_fast_solvers_only,
-                        verbose = verbose)
+#     if verbose println("Solution for derivatives.") end
+#     # get f(vs)
+#     val, min = block_solver(inp, 
+#                         n_block, 
+#                         ss_solve_blocks, 
+#                         # SS_optimizer, 
+#                         # f, 
+#                         guess, 
+#                         lbs, 
+#                         ubs;
+#                         tol = tol,
+#                         # timeout = timeout,
+#                         starting_points = starting_points,
+#                         # fail_fast_solvers_only = fail_fast_solvers_only,
+#                         verbose = verbose)
 
-    if min > tol
-        jvp = fill(0,length(val),length(inp)) * ps
-    else
[…]
