### 2023-08-01 thorek1/MacroModelling.jl#33 @gdalle
https://github.com/jverzani/SymPyCall.jl

Might be interesting
