### 2023-08-01 thorek1/MacroModelling.jl#33 @thorek1
Yeah, I implemented an adapted version of their code in the latest version of MacroModelling. It works with CondaPkg now and the installation should work out of the box.
I asked the SymPyCall authors to register their package so that I can use that instead my hacky way.
