### 2023-08-01 thorek1/MacroModelling.jl#33 @gdalle
Should have known it was not a coincidence 😎
