+++ .github/workflows/ci.yml
@@ -32,6 +32,11 @@ jobs:
           - version: '^1.10.0-0'
             os: ubuntu-latest
             arch: x64
+            allow_failure: true
+          - version: 'nightly'
+            os: ubuntu-latest
+            arch: x64
+            allow_failure: true
     steps:
       - uses: actions/checkout@v3
       - uses: julia-actions/setup-julia@latest
