+++ .github/workflows/ci.yml
@@ -29,9 +29,9 @@ jobs:
           - version: '1.8'
             os: ubuntu-latest
             arch: x64
-          - version: '1.9'
-            os: ubuntu-latest
-            arch: aarch64
+          #- version: '1.9'
+          #  os: ubuntu-latest
+          #  arch: aarch64
           - version: '^1.10.0-0'
             os: ubuntu-latest
             arch: x64
