+++ .github/workflows/ci.yml
@@ -45,9 +45,13 @@ jobs:
       #- uses: uraimo/run-on-arch-action@v2.5.0
       - name: Set up QEMU
         if: matrix.arch == 'aarch64'
-        uses: docker/setup-qemu-action@latest
-        uses: docker/setup-buildx-action@latest
-        uses: docker/build-push-action@latest
+        uses: docker/setup-qemu-action@v2
+      - name: Set up Docker Buildx
+        if: matrix.arch == 'aarch64'
+        uses: docker/setup-buildx-action@v2
+      - name: Build and push
+        if: matrix.arch == 'aarch64'
+        uses: docker/build-push-action@v3
       - uses: julia-actions/setup-julia@latest
         with:
           version: ${{ matrix.version }}
