+++ Dockerfile
@@ -1,0 +1,4 @@
+# This builds a dockerfile containing a working copy of PySR
+# with all pre-requisites installed.
+
+FROM julia:latest
