+++ Dockerfile
@@ -1,4 +1,0 @@
-# This builds a dockerfile containing a working copy of PySR
-# with all pre-requisites installed.
-
-FROM julia:latest
