+++ .github/workflows/ci.yml
@@ -29,9 +29,9 @@ jobs:
           - version: '1.8'
             os: ubuntu-latest
             arch: x64
-          - version: '1.9'
-            os: ubuntu-latest
-            arch: aarch64
+          # - version: '1.9'
+          #   os: ubuntu-latest
+          #   arch: aarch64
           - version: '^1.10.0-0'
             os: ubuntu-latest
             arch: x64
@@ -43,15 +43,15 @@ jobs:
     steps:
       - uses: actions/checkout@v3
       #- uses: uraimo/run-on-arch-action@v2.5.0
-      - name: Set up QEMU
-        if: matrix.arch == 'aarch64'
-        uses: docker/setup-qemu-action@v2
-      - name: Set up Docker Buildx
-        if: matrix.arch == 'aarch64'
-        uses: docker/setup-buildx-action@v2
-      - name: Build and push
-        if: matrix.arch == 'aarch64'
-        uses: docker/build-push-action@v3
+      # - name: Set up QEMU
+      #   if: matrix.arch == 'aarch64'
+      #   uses: docker/setup-qemu-action@v2
+      # - name: Set up Docker Buildx
+      #   if: matrix.arch == 'aarch64'
+      #   uses: docker/setup-buildx-action@v2
+      # - name: Build and push
+      #   if: matrix.arch == 'aarch64'
+      #   uses: docker/build-push-action@v3
       - uses: julia-actions/setup-julia@latest
         with:
           version: ${{ matrix.version }}
