+++ .github/workflows/ci.yml
@@ -9,6 +9,7 @@ jobs:
       GKSwstype: "nul"
     name: Julia ${{ matrix.version }} - ${{ matrix.os }} - ${{ matrix.arch }} - ${{ github.event_name }}
     runs-on: ${{ matrix.os }}
+    continue-on-error: ${{ matrix.version == 'nightly' || matrix.version == '^1.10.0-0' }}
     strategy:
       fail-fast: false
       matrix:
