### 2023-08-08 9ac3ee5
first kind of working macro
