+++ test/develop_for_loop_parser.jl
@@ -1,0 +1,263 @@
+# using MacroModelling
+import MacroModelling as MM
+import MacroTools: postwalk, @capture, unblock
+
+# TODOs
+# parse lag indices
+# nested indices
+
+function replace_for_loop_indices(exxpr,index_variable,indices,concatenate)
+    calls = []
+    # println(exxpr)
+    # println(typeof(exxpr))
+    # println(index_variable)
+    # println(typeof(index_variable))
+    # println(indices)
+    # println(concatenate)
+    for idx in indices
+        push!(calls, postwalk(x -> begin
+            x isa Expr ?
+                x.head == :ref ?
+                    @capture(x, name_{index_}[time_]) ?
+                        index == index_variable ?
+                            :($(Expr(:ref, Symbol(string(name) * "◖" * string(idx) * "◗"),time))) :
+                        time isa Expr || time isa Symbol ?
+                            index_variable ∈ MM.get_symbols(time) ?
+                                :($(Expr(:ref, Expr(:curly,name,index), Meta.parse(replace(string(time), string(index_variable) => idx))))) :
+                            x :
+                        x :
+                    @capture(x, name_[time_]) ?
+                        time isa Expr || time isa Symbol ?
+                            index_variable ∈ MM.get_symbols(time) ?
+                                :($(Expr(:ref, name, Meta.parse(replace(string(time), string(index_variable) => idx))))) :
+                            x :
+                        x :
+                    x :
+                @capture(x, name_{index_}) ?
+                    index == index_variable ?
+                        :($(Symbol(string(name) * "◖" * string(idx) * "◗"))) :
+                    x :
+                x :
+            x
+        end,
+        exxpr))
+    end
+    # println(calls)
+    if concatenate
+        return :($(Expr(:call, :+, calls...)))
+    else
+        # return :($(Expr(:block,calls...)))
+        return calls
+    end
+end
+
+
+replace_for_loop_indices(:(phi{co} * S{co}[lag-1]),:lag,-4:0,true)
+MM.get_symbols(:(lag))
+
+
+function core_for_loop_removal_loop(arg)
+    postwalk(x -> begin
+                    x = unblock(x)
+                    x isa Expr ?
+                        x.head == :for ?
+                            x.args[2] isa Array ?
+                                length(x.args[2]) >= 1 ?
+                                    [replace_for_loop_indices(X, Symbol(x.args[1].args[1]), eval(x.args[1].args[2]), false) for X in x.args[2]] :
+                                #     begin 
+                                #         xx = for X in x.args[2] 
+                                #             yy = replace_for_loop_indices(X, 
+                                #                 Symbol(x.args[1].args[1]), 
[…]
