### 2023-08-08 75d2de4
switch to external sympy package
