+++ src/MacroModelling.jl
@@ -5,7 +5,7 @@ import DocStringExtensions: FIELDS, SIGNATURES, TYPEDEF, TYPEDSIGNATURES, TYPEDF
 # import StatsFuns: normcdf
 using PrecompileTools
 import SpecialFunctions: erfcinv, erfc
-import PythonCall
+import SymPyPythonCall as SPyPyC
 import Symbolics
 import ForwardDiff as ℱ 
 # import Zygote
@@ -34,8 +34,7 @@ import Reexport
 Reexport.@reexport using AxisKeys
 Reexport.@reexport import SparseArrays: sparse, spzeros, droptol!, sparsevec, spdiagm, findnz
 
-include("SymPyCall/SymPyCall.jl")
-import .SymPyCall as SymPyCall
+
 # Type definitions
 Symbol_input = Union{Symbol,Vector{Symbol},Matrix{Symbol},Tuple{Symbol,Vararg{Symbol}}}
 
@@ -144,7 +143,7 @@ function simplify(ex::Expr)
     ex_ss = convert_to_ss_equation(ex)
 
     for x in get_symbols(ex_ss)
-	    eval(:($x = SymPyCall.symbols($(string(x)), real = true, finite = true)))
+	    eval(:($x = SPyPyC.symbols($(string(x)), real = true, finite = true)))
     end
 
 	parsed = ex_ss |> eval |> string |> Meta.parse
@@ -427,13 +426,13 @@ function create_symbols_eqs!(𝓂::ℳ)
     end
 
     for pos in symbols_pos
-        eval(:($pos = SymPyCall.symbols($(string(pos)), real = true, finite = true, positive = true)))
+        eval(:($pos = SPyPyC.symbols($(string(pos)), real = true, finite = true, positive = true)))
     end
     for neg in symbols_neg
-        eval(:($neg = SymPyCall.symbols($(string(neg)), real = true, finite = true, negative = true)))
+        eval(:($neg = SPyPyC.symbols($(string(neg)), real = true, finite = true, negative = true)))
     end
     for none in symbols_none
-        eval(:($none = SymPyCall.symbols($(string(none)), real = true, finite = true)))
+        eval(:($none = SPyPyC.symbols($(string(none)), real = true, finite = true)))
     end
 
     symbolics(map(x->eval(:($x)),𝓂.ss_aux_equations),
@@ -511,7 +510,7 @@ function remove_redundant_SS_vars!(𝓂::ℳ, Symbolics::symbolics)
 
     for i in redundant_idx
         for var_to_solve in redundant_vars[i]
-            soll = try SymPyCall.solve(ss_equations[i],var_to_solve)
+            soll = try SPyPyC.solve(ss_equations[i],var_to_solve)
             catch
             end
             
@@ -519,9 +518,9 @@ function remove_redundant_SS_vars!(𝓂::ℳ, Symbolics::symbolics)
                 continue
             end
             
-            if length(soll) == 0 || soll == SymPyCall.Sym[0] # take out variable if it is redundant from that euation only
+            if length(soll) == 0 || soll == SPyPyC.Sym[0] # take out variable if it is redundant from that euation only
                 push!(Symbolics.var_redundant_list[i],var_to_solve)
-                ss_equations[i] = ss_equations[i].subs(var_to_solve,1).replace(SymPyCall.Sym(ℯ),exp(1)) # replace euler constant as it is not translated to julia properly
+                ss_equations[i] = ss_equations[i].subs(var_to_solve,1).replace(SPyPyC.Sym(ℯ),exp(1)) # replace euler constant as it is not translated to julia properly
             end
 
         end
@@ -568,7 +567,7 @@ function solve_steady_state!(𝓂::ℳ, symbolic_SS, Symbolics::symbolics; verbo
     
     n = n_blocks
 
[…]

[…]
