### 2023-08-09 b1b6c81
first successfull lop parsing
