+++ test/develop_for_loop_parser.jl
@@ -8,12 +8,6 @@ import MacroTools: postwalk, @capture, unblock
 
 function replace_for_loop_indices(exxpr,index_variable,indices,concatenate)
     calls = []
-    # println(exxpr)
-    # println(typeof(exxpr))
-    # println(index_variable)
-    # println(typeof(index_variable))
-    # println(indices)
-    # println(concatenate)
     for idx in indices
         push!(calls, postwalk(x -> begin
             x isa Expr ?
@@ -42,7 +36,6 @@ function replace_for_loop_indices(exxpr,index_variable,indices,concatenate)
         end,
         exxpr))
     end
-    # println(calls)
     if concatenate
         return :($(Expr(:call, :+, calls...)))
     else
@@ -56,6 +49,9 @@ replace_for_loop_indices(:(phi{co} * S{co}[lag-1]),:lag,-4:0,true)
 MM.get_symbols(:(lag))
 
 
+replace_for_loop_indices(:(y{s}[0] = C{s}[0] + I{s}[0] + alpha{s} * G{s}[0]),:s,[:T,:NT],false)
+
+
 function core_for_loop_removal_loop(arg)
     postwalk(x -> begin
                     x = unblock(x)
@@ -64,29 +60,7 @@ function core_for_loop_removal_loop(arg)
                             x.args[2] isa Array ?
                                 length(x.args[2]) >= 1 ?
                                     [replace_for_loop_indices(X, Symbol(x.args[1].args[1]), eval(x.args[1].args[2]), false) for X in x.args[2]] :
-                                #     begin 
-                                #         xx = for X in x.args[2] 
-                                #             yy = replace_for_loop_indices(X, 
-                                #                 Symbol(x.args[1].args[1]), 
-                                #                 eval(x.args[1].args[2]),
-                                #                 false)
-                                #             return yy
-                                #         end
-                                #         # println(xx)
-                                #         xx
-                                #         # println(Expr(x.args[1],x.args[2]...))
-                                # #     for X in x.args[2]
-                                # #         replace_for_loop_indices(unblock(X), 
-                                # #                             Symbol(x.args[1].args[1]), 
-                                # #                             eval(x.args[1].args[2]),
-                                # #                             false)
-                                # #     end
-                                #     end :
                                 x :
-                            # begin
-                            #     println(dump(x.args[2]))
-                            #     x.args[2].head == :(=) || (x.args[2].head == :block && all([i isa Expr && i.head == :(=) for i in x.args[2].args]))
-                            # end ?# || (x.args[2].head == :block && x.args[2].args[2].head == :call) ? #equations inside for loops
                             x.args[2].head ∉ [:(=), :block] ?
                                 replace_for_loop_indices(unblock(x.args[2]), 
                                                     Symbol(x.args[1].args[1]), 
@@ -106,18 +80,10 @@ end
 function parse_for_loops(equations_block)
     eqs = Expr[]
     for arg in equations_block.args
-        # println(arg)
         if isa(arg,Expr)
[…]
