+++ src/MacroModelling.jl
@@ -176,21 +176,35 @@ function convert_to_ss_equation(eq::Expr)
 end
 
 
-function replace_for_loop_indices(exxpr,index_variable,indices,concatenate)
+function replace_indices_inside_for_loop(exxpr,index_variable,indices,concatenate)
     calls = []
     for idx in indices
         push!(calls, postwalk(x -> begin
             x isa Expr ?
                 x.head == :ref ?
-                @capture(x, name_{index_}[time_]) ?
+                    @capture(x, name_{index_}[time_]) ?
                         index == index_variable ?
                             :($(Expr(:ref, Symbol(string(name) * "◖" * string(idx) * "◗"),time))) :
+                        time isa Expr || time isa Symbol ?
+                            index_variable ∈ MM.get_symbols(time) ?
+                                :($(Expr(:ref, Expr(:curly,name,index), Meta.parse(replace(string(time), string(index_variable) => idx))))) :
+                            x :
+                        x :
+                    @capture(x, name_[time_]) ?
+                        time isa Expr || time isa Symbol ?
+                            index_variable ∈ MM.get_symbols(time) ?
+                                :($(Expr(:ref, name, Meta.parse(replace(string(time), string(index_variable) => idx))))) :
+                            x :
                         x :
                     x :
                 @capture(x, name_{index_}) ?
                     index == index_variable ?
                         :($(Symbol(string(name) * "◖" * string(idx) * "◗"))) :
                     x :
+                x :
+            @capture(x, name_) ?
+                name == index_variable && idx isa Int ?
+                    :($idx) :
                 x :
             x
         end,
@@ -204,39 +218,194 @@ function replace_for_loop_indices(exxpr,index_variable,indices,concatenate)
     end
 end
 
+
+function replace_indices(exxpr::Expr)
+    postwalk(x -> begin
+        @capture(x, name_{index_}) ?
+            :($(Symbol(string(name) * "◖" * string(eval(index)) * "◗"))) :
+        x
+        end,
+    exxpr)
+end
+
+
+function write_out_for_loops(arg::Expr)
+    postwalk(x -> begin
+                    x = unblock(x)
+                    x isa Expr ?
+                        x.head == :for ?
+                            x.args[2] isa Array ?
+                                length(x.args[2]) >= 1 ?
+                                    [replace_indices_inside_for_loop(X, Symbol(x.args[1].args[1]), eval(x.args[1].args[2]), false) for X in x.args[2]] :
+                                x :
+                            x.args[2].head ∉ [:(=), :block] ?
+                                replace_indices_inside_for_loop(unblock(x.args[2]), 
+                                                    Symbol(x.args[1].args[1]), 
+                                                    eval(x.args[1].args[2]),
+                                                    true) : # for loop part of equation
+                            replace_indices_inside_for_loop(unblock(x.args[2]), 
+                                                Symbol(x.args[1].args[1]), 
+                                                eval(x.args[1].args[2]),
+                                                false) : # for loop part across equations
+                        x :
+                    x
+                end,
+    arg)
+end
+
+
 function parse_for_loops(equations_block)
     eqs = Expr[]
     for arg in equations_block.args
         if isa(arg,Expr)
-            parsed_eqs = postwalk(x -> begin
-                    x isa Expr ? 
-                        x.head == :for ?
-                            x.args[2].args[2].head == :(=) ?
-                                replace_for_loop_indices(unblock(x.args[2]), 
-                                                        Symbol(x.args[1].args[1]), 
-                                                        [i.value for i  in x.args[1].args[2].args], 
[…]
