+++ src/macros.jl
@@ -910,10 +910,10 @@ macro parameters(𝓂,ex...)
     calib_equations_no_var = []
     calib_values_no_var = []
     
-    calib_parameters_no_var = []
+    calib_parameters_no_var = Symbol[]
     
-    calib_eq_parameters = []
-    calib_equations_list = []
+    calib_eq_parameters = Symbol[]
+    calib_equations_list = Expr[]
     
     ss_calib_list = []
     par_calib_list = []
@@ -924,8 +924,8 @@ macro parameters(𝓂,ex...)
     ss_no_var_calib_list = []
     par_no_var_calib_list = []
     
-    calib_parameters = []
-    calib_values = []
+    calib_parameters = Symbol[]
+    calib_values = Float64[]
 
     par_defined_more_than_once = Set()
     
@@ -960,6 +960,8 @@ macro parameters(𝓂,ex...)
             x,
         exp)
     end
+
+    parameter_definitions = replace_indices(ex[end])
 
     # parse parameter inputs
     # label all variables parameters and exogenous vairables and timings across all equations
@@ -1001,7 +1003,7 @@ macro parameters(𝓂,ex...)
                 x :
             x :
         x,
-    ex[end])
+    parameter_definitions)
 
 
 
@@ -1044,7 +1046,7 @@ macro parameters(𝓂,ex...)
                 x :
             x :
         x,
-    ex[end])
+    parameter_definitions)
     
     @assert length(par_defined_more_than_once) == 0 "Parameters can only be defined once. This is not the case for: " * repr([par_defined_more_than_once...])
     
@@ -1062,7 +1064,7 @@ macro parameters(𝓂,ex...)
     calib_parameters_no_var = setdiff(calib_parameters_no_var,calib_parameters)
     
     for (i, cal_eq) in enumerate(calib_equations)
-        ss_tmp = Set()
+        ss_tmp = Set{Symbol}()
         par_tmp = Set()
     
         # parse SS variables
@@ -1322,7 +1324,12 @@ macro parameters(𝓂,ex...)
     # println($m)
     return quote
         mod = @__MODULE__
-        @assert length(setdiff(setdiff(setdiff(union(reduce(union,$par_calib_list,init = []),mod.$𝓂.parameters_in_equations),$calib_parameters),$calib_parameters_no_var),$calib_eq_parameters)) == 0 "Undefined parameters: " * repr([setdiff(setdiff(setdiff(union(reduce(union,$par_calib_list,init = []),mod.$𝓂.parameters_in_equations),$calib_parameters),$calib_parameters_no_var),$calib_eq_parameters)...])
+
+        calib_parameters, calib_values = expand_indices($calib_parameters, $calib_values, [mod.$𝓂.parameters_in_equations; mod.$𝓂.var])
+        calib_eq_parameters, calib_equations_list, ss_calib_list, par_calib_list = expand_calibration_equations($calib_eq_parameters, $calib_equations_list, $ss_calib_list, $par_calib_list, [mod.$𝓂.parameters_in_equations; mod.$𝓂.var])
+        calib_parameters_no_var, calib_equations_no_var_list = expand_indices($calib_parameters_no_var, $calib_equations_no_var_list, [mod.$𝓂.parameters_in_equations; mod.$𝓂.var])
+
[…]
