+++ test/develop_for_loop_parser.jl
@@ -1,246 +1,0 @@
-# using MacroModelling
-import MacroModelling as MM
-import MacroTools: postwalk, @capture, unblock
-
-# TODOs
-# parse lag indices
-# nested indices
-
-function replace_for_loop_indices(exxpr,index_variable,indices,concatenate)
-    calls = []
-    for idx in indices
-        push!(calls, postwalk(x -> begin
-            x isa Expr ?
-                x.head == :ref ?
-                    @capture(x, name_{index_}[time_]) ?
-                        index == index_variable ?
-                            :($(Expr(:ref, Symbol(string(name) * "◖" * string(idx) * "◗"),time))) :
-                        time isa Expr || time isa Symbol ?
-                            index_variable ∈ MM.get_symbols(time) ?
-                                :($(Expr(:ref, Expr(:curly,name,index), Meta.parse(replace(string(time), string(index_variable) => idx))))) :
-                            x :
-                        x :
-                    @capture(x, name_[time_]) ?
-                        time isa Expr || time isa Symbol ?
-                            index_variable ∈ MM.get_symbols(time) ?
-                                :($(Expr(:ref, name, Meta.parse(replace(string(time), string(index_variable) => idx))))) :
-                            x :
-                        x :
-                    x :
-                @capture(x, name_{index_}) ?
-                    index == index_variable ?
-                        :($(Symbol(string(name) * "◖" * string(idx) * "◗"))) :
-                    x :
-                x :
-            x
-        end,
-        exxpr))
-    end
-    if concatenate
-        return :($(Expr(:call, :+, calls...)))
-    else
-        # return :($(Expr(:block,calls...)))
-        return calls
-    end
-end
-
-
-replace_for_loop_indices(:(phi{co} * S{co}[lag-1]),:lag,-4:0,true)
-MM.get_symbols(:(lag))
-
-
-replace_for_loop_indices(:(y{s}[0] = C{s}[0] + I{s}[0] + alpha{s} * G{s}[0]),:s,[:T,:NT],false)
-
-
-function core_for_loop_removal_loop(arg)
-    postwalk(x -> begin
-                    x = unblock(x)
-                    x isa Expr ?
-                        x.head == :for ?
-                            x.args[2] isa Array ?
-                                length(x.args[2]) >= 1 ?
-                                    [replace_for_loop_indices(X, Symbol(x.args[1].args[1]), eval(x.args[1].args[2]), false) for X in x.args[2]] :
-                                x :
-                            x.args[2].head ∉ [:(=), :block] ?
[…]
