### 2023-08-10 52eea5f
Delete test_multicountry.jl
