+++ test/test_multicountry.jl
@@ -1,649 +1,0 @@
-using MacroModelling
-import MacroModelling as MM
-import MacroTools: postwalk, @capture, unblock
-
-
-
-@model RBC_multicountry begin
-    for co in countries
-
-        Y{co}[0] = ((LAMBDA{co}*K{co}(-{J})^theta{co}*N{co}[0]^(1-theta{co}))^(-nu{co}) + sigma{co}*Z{co}[-1]^(-nu{co}))^(-1/nu{co});
-        K{co}[0] = (1-delta{co})*K{co}[-1] + S{co}[0];
-        X{co}[0] =
-        for lag in (-J+1):0
-                + phi{co}*S{co}({lag})
-        end
-        ;
-
-        A{co}[0] = (1-eta{co})*A{co}[-1] + N{co}[0];
-        L{co}[0] = 1 - alpha{co}*N{co}[0] - (1-alpha{co})*eta{co}*A{co}[-1];
-
-        # Utility multiplied by gamma
-        # U{co} = (C{co}^mu{co}*L{co}^(1-mu{co}))^gamma{co};
-
-        # FOC with respect to consumption
-        psi{co}*mu{co}/C{co}[0] * U{co}[0] = LGM[0];
-
-        # FOC with respect to labor
-        # NOTE: this condition is only valid for alpha = 1
-        psi{co}*(1-mu{co})/L{co}[0] * U{co}[0] * (-alpha{co}) = - LGM[0] * (1-theta{co})/N{co}[0] * (LAMBDA{co}[0] * K{co}(-{J})^theta{co}*N{co}[0]^(1-theta{co}))^(-nu{co})*Y{co}[0]^(1+nu{co});
-
-        # FOC with respect to capital
-        for lag in 0:(J-1)
-            +beta{co}^{lag}*LGM(+{lag})*phi{co}
-        end
-        for lag in 1:J
-            -beta{co}^{lag}*LGM(+{lag})*phi{co}*(1-delta{co})
-        end
-        = beta{co}^{J}*LGM(+{J})*theta{co}/K{co}*(LAMBDA{co}(+{J})*K{co}^theta{co}*N{co}(+{J})^(1-theta{co}))^(-nu{co})*Y{co}(+{J})^(1+nu{co});
-
-        # FOC with respect to stock of inventories
-        LGM=beta{co}*LGM(+1)*(1+sigma{co}*Z{co}^(-nu{co}-1)*Y{co}(+1)^(1+nu{co}));
-
-        # Shock process
-        if co == countries[1]
-            define alt_co = countries[2]
-        else
-            define alt_co = countries[1]
-        end
-        (LAMBDA{co}-1) = rho{co}{co}*(LAMBDA{co}(-1)-1) + rho{co}{alt_co}*(LAMBDA{alt_co}(-1)-1) + E{co};
-
-
-        NX{co} = (Y{co} - (C{co} + X{co} + Z{co} - Z{co}(-1)))/Y{co};
-
-    end
-
-    # World ressource constraint
-    for co in countries
-    +C_@{co} + X_@{co} + Z_@{co} - Z_@{co}(-1)
-    end
-        =
-    for co in countries
-    +Y_@{co}
-    end
-        ;
-
-end
-
-
-
-
-
-
-
-
-
-countries = [:EA,:US]
-
-exxp = :(@model RBC_baseline begin
[…]
