+++ CondaPkg.toml
@@ -1,2 +1,0 @@
-[deps]
-sympy = ""
