### 2023-08-10 f8c146c
more strings accepted
