### 2023-08-10 c977e0d
go over string acceptance once more
