### 2023-08-11 thorek1/MacroModelling.jl#34 @thorek1
To be checked if this is related:

https://discourse.julialang.org/t/the-improvement-of-memory-management-in-multihtreading-julia-1-10beta-is-amazing/102468
