### 2023-08-11 #5598 @gdalle
> by the way reviewing the dependencies requires another submission and it does not belong to this submission.

I wasn't reviewing the dependencies, rather improving one of them (which I develop) to help the author of the current package @thorek1. This is nearly done now so I'll be able to resume the review shortly
