### 2023-08-11 abd9dcf
print number of threads used
