+++ docs/src/unfinished_docs/todo.md
@@ -5,6 +5,7 @@
 - [ ] write get function for variables, parameters, equations with proper parsing so people can understand what happens when invoking for loops
 - [ ] have a look again at get_statistics function
 - [ ] have all functions accept strings ad write tests for it
+- [ ] test string input
 - [ ] document for loops as a separate thing
 - [ ] implement occasionally binding constraints with shocks
 - [ ] implement moment matching for pruned models
@@ -12,7 +13,7 @@
 - [ ] add balanced growth path handling
 - [ ] add JOSS article (see Makie.jl)
 - [ ] write docs for (non-linear) solution algorithms
-- [ ] for cond forecastind and kalman, get rid of observables input and use axis key of data input
+- [ ] for cond forecasting and kalman, get rid of observables input and use axis key of data input
 - [ ] for cond forecasting allow less shocks than conditions with a warning. should be svd then
 - [ ] have parser accept rss | (r[ss] - 1) * 400 = rss
 - [ ] when doing calibration with optimiser have better return values when he doesnt find a solution (probably NaN)
@@ -30,12 +31,11 @@
 - [ ] kick out unused parameters from m.parameters
 - [ ] higher order solution derivs with Zygote
 - [ ] use cache for gradient calc in estimation (see DifferentiableStateSpaceModels)
-- [ ] use krylov instead of linearsolve and speed up sparse matrix calcs in implicit diff of higher order funcs
+- [ ] speed up sparse matrix calcs in implicit diff of higher order funcs
 - [ ] improve docs: timing in first sentence seems off; have something more general in first sentence; why is the syntax user friendly? give an example; make the former and the latter a footnote
 - [ ] streamline estimation part (dont do string matching... but rely on precomputed indices...)
 - [ ] change docs to reflect that the output of irfs include aux vars and also the model info Base.show includes aux vars
 - [ ] write functions to debug (fix_SS.jl...)
-- [ ] parser model into per equation functions instead of single big functions
 - [ ] model compression (speed up 2nd moment calc (derivatives) for large models; gradient loglikelihood is very slow due to large matmuls) -> model setup as maximisation problem (gEcon) -> HANK models
 - [ ] implement global solution methods
 - [ ] add more models
@@ -53,6 +53,8 @@
 - [ ] Find any SS by optimising over both SS guesses and parameter inputs
 - [ ] weed out SS solver and saved objects
 
+- [x] parser model into per equation functions instead of single big functions
+- [x] use krylov instead of linearsolve
 - [x] implement for loops in model macro (e.g. to setup multi country models)
 - [x] fix ss of pruned solution in plotsolution. seems detached
 - [x] try solve first order with JuMP - doesnt work because JuMP cannot handle matrix constraints/objectives 
