### 2023-08-11 aafa094
test string inputs
