+++ test/functionality_tests.jl
@@ -10,12 +10,22 @@ function functionality_test(m; algorithm = :first_order, plots = true, verbose =
     nsss_select_par_deriv3 = get_steady_state(m, verbose = true, parameter_derivatives = Tuple(m.parameters[1:3]))
     nsss_select_par_deriv4 = get_steady_state(m, verbose = true, parameter_derivatives = reshape(m.parameters[1:3],3,1))
 
+    nsss_select_par_deriv1 = get_steady_state(m, verbose = true, parameter_derivatives = string.(m.parameters[1]))
+    nsss_select_par_deriv2 = get_steady_state(m, verbose = true, parameter_derivatives = string.(m.parameters[1:2]))
+    nsss_select_par_deriv3 = get_steady_state(m, verbose = true, parameter_derivatives = string.(Tuple(m.parameters[1:3])))
+    nsss_select_par_deriv4 = get_steady_state(m, verbose = true, parameter_derivatives = string.(reshape(m.parameters[1:3],3,1)))
+
     old_par_vals = copy(m.parameter_values)
 
     new_nsss1 = get_steady_state(m, verbose = true, parameters = m.parameter_values * 1.0001)
     new_nsss2 = get_steady_state(m, verbose = true, parameters = (m.parameters[1] => m.parameter_values[1] * 1.0001))
     new_nsss3 = get_steady_state(m, verbose = true, parameters = Tuple(m.parameters[1:2] .=> m.parameter_values[1:2] * 1.0001))
     new_nsss4 = get_steady_state(m, verbose = true, parameters = (m.parameters[1:2] .=> m.parameter_values[1:2] / 1.0001))
+    old_nsss = get_steady_state(m, verbose = true, parameters = old_par_vals)
+
+    new_nsss2 = get_steady_state(m, verbose = true, parameters = (string.(m.parameters[1]) => m.parameter_values[1] * 1.0001))
+    new_nsss3 = get_steady_state(m, verbose = true, parameters = Tuple(string.(m.parameters[1:2]) .=> m.parameter_values[1:2] * 1.0001))
+    new_nsss4 = get_steady_state(m, verbose = true, parameters = (string.(m.parameters[1:2]) .=> m.parameter_values[1:2] / 1.0001))
     old_nsss = get_steady_state(m, verbose = true, parameters = old_par_vals)
     nsss = get_non_stochastic_steady_state(m)
     nsss = get_SS(m)
@@ -32,6 +42,13 @@ function functionality_test(m; algorithm = :first_order, plots = true, verbose =
         new_auto_corr3 = autocorr(m, verbose = true, parameters = (m.parameters[1:2] .=> m.parameter_values[1:2] * 1.0002))
         old_auto_corr = get_autocorrelation(m, verbose = true, parameters = old_par_vals)
 
+        new_auto_corr1 = get_autocorrelation(m, verbose = true, parameters = (string.(m.parameters[1]) => m.parameter_values[1] * 1.0001))
+        new_auto_corr2 = get_autocorrelation(m, verbose = true, parameters = Tuple(string.(m.parameters[1:2]) .=> m.parameter_values[1:2] * 1.0001))
+        new_auto_corr3 = get_autocorrelation(m, verbose = true, parameters = (string.(m.parameters[1:2]) .=> m.parameter_values[1:2] / 1.0001))
+        new_auto_corr3 = get_autocorr(m, verbose = true, parameters = (string.(m.parameters[1:2]) .=> m.parameter_values[1:2] / 1.0001))
+        new_auto_corr3 = autocorr(m, verbose = true, parameters = (string.(m.parameters[1:2]) .=> m.parameter_values[1:2] * 1.0002))
+        old_auto_corr = get_autocorrelation(m, verbose = true, parameters = old_par_vals)
+
         corr_nv = get_correlation(m)
[…]
