+++ src/MacroModelling.jl
@@ -187,13 +187,13 @@ function replace_indices_inside_for_loop(exxpr,index_variable,indices,concatenat
                         index == index_variable ?
                             :($(Expr(:ref, Symbol(string(name) * "◖" * string(idx) * "◗"),time))) :
                         time isa Expr || time isa Symbol ?
-                            index_variable ∈ MM.get_symbols(time) ?
+                            index_variable ∈ get_symbols(time) ?
                                 :($(Expr(:ref, Expr(:curly,name,index), Meta.parse(replace(string(time), string(index_variable) => idx))))) :
                             x :
                         x :
                     @capture(x, name_[time_]) ?
                         time isa Expr || time isa Symbol ?
-                            index_variable ∈ MM.get_symbols(time) ?
+                            index_variable ∈ get_symbols(time) ?
                                 :($(Expr(:ref, name, Meta.parse(replace(string(time), string(index_variable) => idx))))) :
                             x :
                         x :
@@ -3180,7 +3180,7 @@ end
 
 
 
-function riccati_forward(∇₁::Matrix{ℱ.Dual{Z,S,N}}; T::timings = T, explosive::Bool = false) where {Z,S,N}
+function riccati_forward(∇₁::Matrix{ℱ.Dual{Z,S,N}}; T::timings, explosive::Bool = false) where {Z,S,N}
     # unpack: AoS -> SoA
     ∇̂₁ = ℱ.value.(∇₁)
     # you can play with the dimension here, sometimes it makes sense to transpose
@@ -3708,23 +3708,23 @@ function parse_variables_input_to_index(variables::Union{Symbol_input,String_inp
     elseif variables == :all_including_auxilliary
         return 1:length(union(T.var,T.aux,T.exo_present))
     elseif variables isa Matrix{Symbol}
-        if !issubset(variables,T.var)
-            return @warn "Following variables are not part of the model: " * string.(setdiff(variables,T.var))
+        if length(setdiff(variables,T.var)) > 0
+            return @warn "Following variables are not part of the model: " * join(string.(setdiff(variables,T.var)),", ")
         end
         return getindex(1:length(T.var),convert(Vector{Bool},vec(sum(variables .== T.var,dims= 2))))
     elseif variables isa Vector{Symbol}
-        if !issubset(variables,T.var)
-            return @warn "Following variables are not part of the model: " * string.(setdiff(variables,T.var))
+        if length(setdiff(variables,T.var)) > 0
+            return @warn "Following variables are not part of the model: " * join(string.(setdiff(variables,T.var)),", ")
         end
         return getindex(1:length(T.var),convert(Vector{Bool},vec(sum(reshape(variables,1,length(variables)) .== T.var,dims= 2))))
     elseif variables isa Tuple{Symbol,Vararg{Symbol}}
-        if !issubset(variables,T.var)
-            return @warn "Following variables are not part of the model: " * string.(setdiff(variables,T.var))
+        if length(setdiff(variables,T.var)) > 0
+            return @warn "Following variables are not part of the model: " * join(string.(setdiff(Symbol.(collect(variables)),T.var)), ", ")
         end
         return getindex(1:length(T.var),convert(Vector{Bool},vec(sum(reshape(collect(variables),1,length(variables)) .== T.var,dims= 2))))
     elseif variables isa Symbol
-        if !issubset([variables],T.var)
-            return @warn "Following variable is not part of the model: " * string(setdiff([variables],T.var)[1])
[…]
