+++ docs/Project.toml
@@ -5,6 +5,7 @@ ComponentArrays = "b0b7db55-cfe3-40fc-9ded-d10e2dbeff66"
 DataFrames = "a93c6f00-e57d-5684-b7b6-d8193f3e46c0"
 Documenter = "e30172f5-a6a5-5a46-863b-614d45cd2de4"
 DynamicPPL = "366bfd00-2699-11ea-058f-f148b4cae6d8"
+JET = "c3a54625-cd67-489e-a8e7-0a5a0ff4e31b"
 LineSearches = "d3d80556-e9d4-5f37-9878-2ab0fcc64255"
 MCMCChains = "c7f686f2-ff18-58e9-bc7b-31028e88f75d"
 MCMCDiagnosticTools = "be115224-59cd-429b-ad48-344e309966f0"
