+++ .github/workflows/ci.yml
@@ -54,6 +54,9 @@ jobs:
         with:
           version: ${{ matrix.version }}
           arch: ${{ matrix.arch }}
+      - name: Set JULIA_NUM_THREADS for Julia 1.9 on Ubuntu
+        if: matrix.version == '1.9' && matrix.os == 'ubuntu-latest'
+        run: echo "JULIA_NUM_THREADS=2" >> $GITHUB_ENV
       - uses: actions/cache@v3
         env:
           cache-name: cache-artifacts
@@ -69,9 +72,6 @@ jobs:
       - uses: julia-actions/julia-runtest@latest
         with:
           prefix: ${{ matrix.prefix }}  # for `xvfb-run`
-        if: matrix.os == 'ubuntu-latest' && matrix.version == '1.9'
-        env: 
-          JULIA_NUM_THREADS: 2
       - uses: julia-actions/julia-processcoverage@latest
       - uses: codecov/codecov-action@v3
         with:
