### 2023-08-11 0b0f9f6
improved docs to mention that strings are accepted
