### 2023-08-11 4a899b8
accept keyedarrays with strings as axis
