### 2023-08-11 07d518e
fix keyed array converter
