### 2023-08-11 76d9dad
test keyedarray with strings as input
