### 2023-08-11 99a7084
output indices properly
