+++ src/plotting.jl
@@ -146,7 +146,7 @@ function plot_model_estimates(𝓂::ℳ,
             push!(pp,begin
                     StatsPlots.plot()
                     StatsPlots.plot!(shocks_to_plot[shock_idx[i - length(var_idx)],:],
-                        title = string(𝓂.timings.exo[shock_idx[i - length(var_idx)]]) * "₍ₓ₎", 
+                        title = replace_indices_in_symbol(𝓂.timings.exo[shock_idx[i - length(var_idx)]]) * "₍ₓ₎", 
                         ylabel = shock_decomposition ? "Absolute Δ" : "Level",label = "", 
                         color = shock_decomposition ? estimate_color : :auto)
                     StatsPlots.hline!([0],
@@ -159,7 +159,7 @@ function plot_model_estimates(𝓂::ℳ,
             if shock_decomposition SS = zero(SS) end
 
             can_dual_axis = gr_back &&  all((variables_to_plot[var_idx[i],:] .+ SS) .> eps(Float32)) && (SS > eps(Float32)) && !shock_decomposition
-            
+
             push!(pp,begin
                     StatsPlots.plot()
                     if shock_decomposition
@@ -170,12 +170,12 @@ function plot_model_estimates(𝓂::ℳ,
                             alpha = transparency)
                     end
                     StatsPlots.plot!(variables_to_plot[var_idx[i],:] .+ SS,
-                        title = string(𝓂.timings.var[var_idx[i]]), 
+                        title = replace_indices_in_symbol(𝓂.timings.var[var_idx[i]]), 
                         ylabel = shock_decomposition ? "Absolute Δ" : "Level",label = "", 
                         color = shock_decomposition ? estimate_color : :auto)
                     if var_idx[i] ∈ obs_idx 
                         StatsPlots.plot!(data_in_deviations[indexin([var_idx[i]],obs_idx),:]' .+ SS,
-                            title = string(𝓂.timings.var[var_idx[i]]),
+                            title = replace_indices_in_symbol(𝓂.timings.var[var_idx[i]]),
                             ylabel = shock_decomposition ? "Absolute Δ" : "Level", 
                             label = "", 
                             color = shock_decomposition ? data_color : :auto) 
@@ -468,7 +468,7 @@ function plot_irf(𝓂::ℳ;
             if !(all(isapprox.(Y[i,:,shock],0,atol = eps(Float32))))
                 push!(pp,begin
                                 StatsPlots.plot(Y[i,:,shock] .+ SS,
-                                                title = string(𝓂.timings.var[var_idx[i]]),
+                                                title = replace_indices_in_symbol(𝓂.timings.var[var_idx[i]]),
                                                 ylabel = "Level",
                                                 label = "")
 
@@ -496,8 +496,8 @@ function plot_irf(𝓂::ℳ;
                         shock_string = ""
                         shock_name = "no_shock"
                     elseif shocks isa Union{Symbol_input,String_input}
-                        shock_string = ": " * string(𝓂.timings.exo[shock_idx[shock]])
-                        shock_name = string(𝓂.timings.exo[shock_idx[shock]])
+                        shock_string = ": " * replace_indices_in_symbol(𝓂.timings.exo[shock_idx[shock]])
+                        shock_name = replace_indices_in_symbol(𝓂.timings.exo[shock_idx[shock]])
                     else
                         shock_string = "Series of shocks"
                         shock_name = "shock_matrix"
@@ -530,14 +530,14 @@ function plot_irf(𝓂::ℳ;
[…]

[…]
