### 2023-08-12 4b38a03
make tests work for for loop model
