### 2023-08-12 700a132
plot_solution legend
