+++ src/plotting.jl
@@ -1042,11 +1042,12 @@ function plot_solution(𝓂::ℳ,
         end
     end
 
+    state = replace_indices_in_symbol(state)
 
     for (i,k) in enumerate(vars_to_plot)
         kk = Symbol(replace(string(k), r"ᴸ⁽⁻?[⁰¹²³⁴⁵⁶⁷⁸⁹]+⁾" => ""))
 
-        # kk = replace_indices_in_symbol(kk)
+        k = replace_indices_in_symbol(k)
 
         if !has_impact_list[i] continue end
 
