+++ docs/src/how-to/loops.md
@@ -1,0 +1,154 @@
+# Programmatic model writing
+
+Programmatic model writing is a powerful tool to write complex models using concise code. More specifically, the `@model` and `@parameters` macros allow for the use of indexed variables and for-loops.
+
+## Model block
+
+### for loops for time indices
+
+In practice this means that you no longer need to write this:
+
+```julia
+Y_annual[0] = Y[0] + Y[-1] + Y[-2] + Y[-3]
+```
+
+but instead you can write this:
+
+```julia
+Y_annual[0] = for lag in -3:0 Y[lag] end
+```
+
+In the background the package expands the `for` loop and adds up the elements for the different values of `lag`.
+
+### for loops for variables / parameter specific indices
+
+Another use-case are models with repetitive equations such as multi-sector or multi-country models.
+
+For example, defining the production function for two countries (home country `H` and foreign country `F`) would  look as follows without the use of programmatic features:
+
+```julia
+y_H[0] = A_H[0] * k_H[-1]^alpha_H
+y_F[0] = A_F[0] * k_F[-1]^alpha_F
+```
+
+and this can be written more conveniently using loops:
+
+```julia
+for co in [H, F] y{co}[0] = A{co}[0] * k{co}[-1]^alpha{co} end
+```
+
+Note that the package internally writes out the for loop and creates two equations; one each for country `H` and `F`. The variables and parameters are indexed using the curly braces `{}`. These can also be used outside loops. When using more than one index it is important to make sure the indices are in the right order.
+
+### Example model block
+
+Putting these these elements together we can write the multi-country model equations of the Backus, Kehoe and Kydland (1992) model like this:
+
+```julia
+@model Backus_Kehoe_Kydland_1992 begin
+    for co in [H, F]
+        Y{co}[0] = ((LAMBDA{co}[0] * K{co}[-4]^theta{co} * N{co}[0]^(1 - theta{co}))^(-nu{co}) + sigma{co} * Z{co}[-1]^(-nu{co}))^(-1 / nu{co})
+
+        K{co}[0] = (1 - delta{co}) * K{co}[-1] + S{co}[0]
+
+        X{co}[0] = for lag in (-4+1):0 phi{co} * S{co}[lag] end
+
+        A{co}[0] = (1 - eta{co}) * A{co}[-1] + N{co}[0]
+
+        L{co}[0] = 1 - alpha{co} * N{co}[0] - (1 - alpha{co}) * eta{co} * A{co}[-1]
+
+        U{co}[0] = (C{co}[0]^mu{co} * L{co}[0]^(1 - mu{co}))^gamma{co}
+
+        psi{co} * mu{co} / C{co}[0] * U{co}[0] = LGM[0]
+
+        psi{co} * (1 - mu{co}) / L{co}[0] * U{co}[0] * (-alpha{co}) = - LGM[0] * (1 - theta{co}) / N{co}[0] * (LAMBDA{co}[0] * K{co}[-4]^theta{co} * N{co}[0]^(1 - theta{co}))^(-nu{co}) * Y{co}[0]^(1 + nu{co})
+
+        for lag in 0:(4-1)  
+            beta{co}^lag * LGM[lag]*phi{co}
+        end +
+        for lag in 1:4
+            -beta{co}^lag * LGM[lag] * phi{co} * (1 - delta{co})
+        end = beta{co}^4 * LGM[+4] * theta{co} / K{co}[0] * (LAMBDA{co}[+4] * K{co}[0]^theta{co} * N{co}[+4]^(1 - theta{co})) ^ (-nu{co}) * Y{co}[+4]^(1 + nu{co})
+
[…]
