### 2023-08-15 6f2ad3d
for loop multiply
