### 2023-08-18 fab6275
add model printing
