+++ src/inspect.jl
@@ -1,0 +1,803 @@
+get_symbols(ex::Symbol) = [ex]
+
+function get_symbols(ex::Expr)
+    par = Set{Symbol}()
+    postwalk(x ->   
+    x isa Expr ? 
+        x.head == :(=) ?
+            for i in x.args
+                i isa Symbol ? 
+                    push!(par,i) :
+                x
+            end :
+        x.head == :call ? 
+            for i in 2:length(x.args)
+                x.args[i] isa Symbol ? 
+                    push!(par,x.args[i]) : 
+                x
+            end : 
+        x : 
+    x, ex)
+    return par
+end
+
+
+"""
+$(SIGNATURES)
+Return the equations of the model. In case programmatic model writing was used this function returns the parsed equations (see loop over shocks in example).
+
+# Arguments
+- $MODEL
+
+# Examples
+```jldoctest
+using MacroModelling
+
+@model RBC begin
+    1  /  c[0] = (β  /  c[1]) * (α * exp(z{TFP}[1]) * k[0]^(α - 1) + (1 - exp(z{δ}[1]) * δ))
+    c[0] + k[0] = (1 - exp(z{δ}[0])δ) * k[-1] + q[0]
+    q[0] = exp(z{TFP}[0]) * k[-1]^α
+    for shock in [TFP, δ]
+        z{shock}[0] = ρ{shock} * z{shock}[-1] + σ{shock} * (eps{shock}[x] + eps_news{shock}[x-1])
+    end
+    Δc_share[0] = log(c[0]/q[0]) - log(c[-1]/q[-1])
+    Δk_4q[0] = log(k[0]) - log(k[-4])
+end
+
+@parameters RBC begin
+    σ = 0.01
+    ρ = 0.2
+    capital_to_output = 1.5
+    k[ss] / (4 * q[ss]) = capital_to_output | δ
+    alpha = .5
+    α = alpha
+    β = 0.95
+end
+
+get_equations(RBC)
+# output
+7-element Vector{String}:
+ "1 / c[0] = (β / c[1]) * (α * ex" ⋯ 25 bytes ⋯ " - 1) + (1 - exp(z{δ}[1]) * δ))"
+ "c[0] + k[0] = (1 - exp(z{δ}[0]) * δ) * k[-1] + q[0]"
+ "q[0] = exp(z{TFP}[0]) * k[-1] ^ α"
+ "z{TFP}[0] = ρ{TFP} * z{TFP}[-1]" ⋯ 18 bytes ⋯ "TFP}[x] + eps_news{TFP}[x - 1])"
+ "z{δ}[0] = ρ{δ} * z{δ}[-1] + σ{δ} * (eps{δ}[x] + eps_news{δ}[x - 1])"
+ "Δc_share[0] = log(c[0] / q[0]) - log(c[-1] / q[-1])"
+ "Δk_4q[0] = log(k[0]) - log(k[-4])"
+```
+"""
+function get_equations(𝓂::ℳ)
+    replace.(string.(𝓂.original_equations), "◖" => "{", "◗" => "}")
+end
+
+
+"""
+$(SIGNATURES)
+Return the non-stochastic steady state (NSSS) equations of the model. The difference to the equations as they were written in the `@model` block is that exogenous shocks are set to `0`, time subscripts are eliminated (e.g. `c[-1]` becomes `c`), trivial simplifications are carried out (e.g. `log(k) - log(k) = 0`), and auxilliary variables are added for expressions that cannot become negative. 
+
[…]
