### 2023-08-18 9979aa1
fix reduce init
