+++ Project.toml
@@ -45,7 +45,7 @@ DataStructures = "^0.18"
 DocStringExtensions = "^0.8, 0.9"
 DynarePreprocessor_jll = "^6.2"
 ForwardDiff = "^0.10"
-ImplicitDifferentiation = "^0.2, 0.3"
+ImplicitDifferentiation = "^0.5"
 JSON = "^0.21"
 Krylov = "^0.9"
 LaTeXStrings = "^1"
