### 2023-08-18 240dc19
change calculate covariance
