### 2023-08-18 92e3994
add dual versions back in for julia1.8
