### 2023-08-19 7230df6
specify linear solver and backend for block_solver
