+++ src/MacroModelling.jl
@@ -17,7 +17,8 @@ import Subscripts: super, sub
 import Krylov
 import LinearOperators
 import DataStructures: CircularBuffer
-using ImplicitDifferentiation
+import ImplicitDifferentiation as ID
+import AbstractDifferentiation as AD
 import SpeedMapping: speedmapping
 # import NLboxsolve: nlboxsolve
 # using NamedArrays
@@ -96,20 +97,20 @@ dnorm(p::Number) = normpdf(p)
 
 
 Base.show(io::IO, 𝓂::ℳ) = println(io, 
-                "Model:       ", 𝓂.model_name, 
+                "Model:        ", 𝓂.model_name, 
                 "\nVariables", 
-                "\n Total:      ", 𝓂.timings.nVars,
-                "\n  Auxiliary: ", length(𝓂.exo_present) + length(𝓂.aux),
-                "\n States:     ", 𝓂.timings.nPast_not_future_and_mixed,
-                "\n  Auxiliary: ",  length(intersect(𝓂.timings.past_not_future_and_mixed, 𝓂.aux_present)),
-                "\n Jumpers:    ", 𝓂.timings.nFuture_not_past_and_mixed, # 𝓂.timings.mixed, 
-                "\n  Auxiliary: ", length(intersect(𝓂.timings.future_not_past_and_mixed, union(𝓂.aux_present, 𝓂.aux_future))),
-                "\nShocks:      ", 𝓂.timings.nExo,
-                "\nParameters:  ", length(𝓂.parameters_in_equations),
+                "\n Total:       ", 𝓂.timings.nVars,
+                "\n  Auxiliary:  ", length(𝓂.exo_present) + length(𝓂.aux),
+                "\n States:      ", 𝓂.timings.nPast_not_future_and_mixed,
+                "\n  Auxiliary:  ",  length(intersect(𝓂.timings.past_not_future_and_mixed, 𝓂.aux_present)),
+                "\n Jumpers:     ", 𝓂.timings.nFuture_not_past_and_mixed, # 𝓂.timings.mixed, 
+                "\n  Auxiliary:  ", length(intersect(𝓂.timings.future_not_past_and_mixed, union(𝓂.aux_present, 𝓂.aux_future))),
+                "\nShocks:       ", 𝓂.timings.nExo,
+                "\nParameters:   ", length(𝓂.parameters_in_equations),
                 if 𝓂.calibration_equations == Expr[]
                     ""
                 else
-                    "\nCalibration equations: " * repr(length(𝓂.calibration_equations))
+                    "\nCalibration\nequations:    " * repr(length(𝓂.calibration_equations))
                 end,
                 # "\n¹: including auxilliary variables"
                 # "\nVariable bounds (upper,lower,any): ",sum(𝓂.upper_bounds .< Inf),", ",sum(𝓂.lower_bounds .> -Inf),", ",length(𝓂.bounds),
@@ -1133,7 +1134,7 @@ function solve_steady_state!(𝓂::ℳ, symbolic_SS, Symbolics::symbolics; verbo
 
                 # push!(SS_solve_func,:(println([$(calib_pars_input...),$(other_vars_input...)])))
 
-                push!(SS_solve_func,:(block_solver_AD = ImplicitFunction(block_solver, 𝓂.ss_solve_blocks[$(n_block)])))
+                push!(SS_solve_func,:(block_solver_AD = ID.ImplicitFunction(block_solver, 𝓂.ss_solve_blocks[$(n_block)]; linear_solver = ID.DirectLinearSolver(), conditions_backend = AD.ForwardDiffBackend())))
 
[…]
