### 2023-08-19 deeb5c2
more implicitdiff with tuned options
