+++ src/MacroModelling.jl
@@ -1802,8 +1802,7 @@ end
 
 
 
-function second_order_stochastic_steady_state_iterative_solution(𝐒₁𝐒₂::AbstractArray{Float64}, 𝓂::ℳ, pruning::Bool;
-    tol::AbstractFloat = eps())
+function second_order_stochastic_steady_state_iterative_solution_forward(𝐒₁𝐒₂::AbstractArray{Float64}; 𝓂::ℳ, pruning::Bool, tol::AbstractFloat = eps())
     (; 𝐒₁, 𝐒₂) = 𝐒₁𝐒₂
 
     state = zeros(𝓂.timings.nVars)
@@ -1841,7 +1840,7 @@ function second_order_stochastic_steady_state_iterative_solution(𝐒₁𝐒₂:
 end
 
 
-function second_order_stochastic_steady_state_iterative_solution_condition(𝐒₁𝐒₂, SSS, 𝓂::ℳ, pruning::Bool)
+function second_order_stochastic_steady_state_iterative_solution_conditions(𝐒₁𝐒₂, SSS, converged::Bool; 𝓂::ℳ, pruning::Bool, tol::AbstractFloat = eps())
     (; 𝐒₁, 𝐒₂) = 𝐒₁𝐒₂
 
     shock = zeros(𝓂.timings.nExo)
@@ -1862,7 +1861,7 @@ function second_order_stochastic_steady_state_iterative_solution_condition(𝐒�
 end
 
 
-function second_order_stochastic_steady_state_iterative_solution(𝐒₁𝐒₂::AbstractArray{ℱ.Dual{Z,S,N}}, 𝓂::ℳ, pruning::Bool) where {Z,S,N}
+function second_order_stochastic_steady_state_iterative_solution_forward(𝐒₁𝐒₂::AbstractArray{ℱ.Dual{Z,S,N}}; 𝓂::ℳ, pruning::Bool, tol::AbstractFloat = eps()) where {Z,S,N}
 
     # unpack: AoS -> SoA
     S₁S₂ = ℱ.value.(𝐒₁𝐒₂)
@@ -1871,12 +1870,12 @@ function second_order_stochastic_steady_state_iterative_solution(𝐒₁𝐒₂:
     ps = mapreduce(ℱ.partials, hcat, 𝐒₁𝐒₂)'
 
     # get f(vs)
-    val, converged = second_order_stochastic_steady_state_iterative_solution(S₁S₂, 𝓂, pruning)
+    val, converged = second_order_stochastic_steady_state_iterative_solution_forward(S₁S₂; 𝓂 = 𝓂, pruning = pruning, tol = tol)
 
     if converged
         # get J(f, vs) * ps (cheating). Write your custom rule here
-        B = ℱ.jacobian(x -> second_order_stochastic_steady_state_iterative_solution_condition(x, val, 𝓂, pruning), S₁S₂)
-        A = ℱ.jacobian(x -> second_order_stochastic_steady_state_iterative_solution_condition(S₁S₂, x, 𝓂, pruning), val)
+        B = ℱ.jacobian(x -> second_order_stochastic_steady_state_iterative_solution_conditions(x, val, converged; 𝓂 = 𝓂, pruning = pruning, tol = tol), S₁S₂)
+        A = ℱ.jacobian(x -> second_order_stochastic_steady_state_iterative_solution_conditions(S₁S₂, x, converged; 𝓂 = 𝓂, pruning = pruning, tol = tol), val)
 
         Â = RF.lu(A, check = false)
 
@@ -1901,6 +1900,11 @@ function second_order_stochastic_steady_state_iterative_solution(𝐒₁𝐒₂:
 end
 
 
[…]
