### 2023-08-19 a8b7ce8
forward 2nd order SSS
