### 2023-08-19 d0744ff
fix kwargs, solved flags and sylvester conditions
