+++ docs/src/unfinished_docs/todo.md
@@ -6,6 +6,7 @@
 - [ ] use other quadratic iteration for diffable first order solve (useful because schur can error in estimation)
 - [ ] use more implicit diff for the other functions as well
 - [ ] add for loop parser in @parameters
+- [ ] compressed higher order derivatives and sparsity of jacobian
 - [ ] implement more multi country models
 - [ ] implement moment matching for pruned models
 - [ ] speed benchmarking (focus on ImplicitDiff part)
