+++ test/prodmom.m
@@ -1,286 +1,0 @@
-function nXI4min = Xi_Gaussian_prodmom4_orderApp2_nx2_nu2(arg)
-SIGe_1_1 = arg(1);
-SIGe_2_1 = arg(2);
-SIGe_2_2 = arg(3);
-E_XF1_1 = arg(4);
-E_XF1_2 = arg(5);
-E_XF2_1 = arg(6);
-E_XF2_2 = arg(7);
-E_XF2_3 = arg(8);
-E_XF3_1 = arg(9);
-E_XF3_2 = arg(10);
-E_XF3_3 = arg(11);
-E_XF3_4 = arg(12);
-E_XF4_1 = arg(13);
-E_XF4_2 = arg(14);
-E_XF4_3 = arg(15);
-E_XF4_4 = arg(16);
-E_XF4_5 = arg(17);
-nXI4min=zeros(495,1);
-nXI4min(1,1) = 3*SIGe_1_1^2;
-nXI4min(2,1) = (3*SIGe_1_1^(3/2)*SIGe_2_1*SIGe_2_2^(1/2))/(SIGe_1_1*SIGe_2_2)^(1/2);
-nXI4min(6,1) = 3*E_XF1_1*SIGe_1_1^2;
-nXI4min(7,1) = (3*E_XF1_1*SIGe_1_1^(3/2)*SIGe_2_1*SIGe_2_2^(1/2))/(SIGe_1_1*SIGe_2_2)^(1/2);
-nXI4min(8,1) = 3*E_XF1_2*SIGe_1_1^2;
-nXI4min(9,1) = (3*E_XF1_2*SIGe_1_1^(3/2)*SIGe_2_1*SIGe_2_2^(1/2))/(SIGe_1_1*SIGe_2_2)^(1/2);
-nXI4min(10,1) = SIGe_1_1*SIGe_2_2 + 2*SIGe_2_1^2;
-nXI4min(14,1) = (3*E_XF1_1*SIGe_1_1^(3/2)*SIGe_2_1*SIGe_2_2^(1/2))/(SIGe_1_1*SIGe_2_2)^(1/2);
-nXI4min(15,1) = E_XF1_1*(SIGe_1_1*SIGe_2_2 + 2*SIGe_2_1^2);
-nXI4min(16,1) = (3*E_XF1_2*SIGe_1_1^(3/2)*SIGe_2_1*SIGe_2_2^(1/2))/(SIGe_1_1*SIGe_2_2)^(1/2);
-nXI4min(17,1) = E_XF1_2*(SIGe_1_1*SIGe_2_2 + 2*SIGe_2_1^2);
-nXI4min(18,1) = 10*SIGe_1_1^3;
-nXI4min(19,1) = (12*SIGe_1_1^(5/2)*SIGe_2_1*SIGe_2_2^(1/2))/(SIGe_1_1*SIGe_2_2)^(1/2) - 2*SIGe_1_1^2*SIGe_2_1;
-nXI4min(20,1) = 10*SIGe_1_1*SIGe_2_1^2;
-nXI4min(25,1) = SIGe_1_1*SIGe_2_1^2 + SIGe_1_1^2*SIGe_2_2*((12*SIGe_2_1^2)/(SIGe_1_1*SIGe_2_2) + 3) - (6*SIGe_1_1^(3/2)*SIGe_2_1^2*SIGe_2_2^(1/2))/(SIGe_1_1*SIGe_2_2)^(1/2);
[…]
