### 2023-08-19 cbcb82f
undo change in main script and get
