### 2023-08-20 b233208
add andreasen et al citation for pruning
