### 2023-08-20 721d7ab
implement pruning according to andreasen 2018
