### 2023-08-20 9a03cad
add tests for pruning
