+++ test/test_standalone_function.jl
@@ -248,33 +248,77 @@ end
         return Tz * aug_state + second_order_solution * kron(aug_state, aug_state) / 2 + third_order_solution * kron(kron(aug_state,aug_state),aug_state) / 6
     end
 
+    unspecified_initial_state = true
 
-    iirrff = irf(first_order_state_update, zeros(T.nVars), zeros(T.nVars), false, T)
+    iirrff = irf(first_order_state_update, zeros(T.nVars), zeros(T.nVars), false, unspecified_initial_state, T)
 
     @test isapprox(iirrff[4,1,:],[ -0.00036685520477089503
     0.0021720718769730014],rtol = eps(Float32))
-    ggiirrff = girf(first_order_state_update, zeros(T.nVars), zeros(T.nVars), false, T)
+    ggiirrff = girf(first_order_state_update, zeros(T.nVars), zeros(T.nVars), false, unspecified_initial_state, T)
     @test isapprox(iirrff[4,1,:],ggiirrff[4,1,:],rtol = eps(Float32))
 
 
     SSS_delta = RBC_CME.solution.non_stochastic_steady_state[1:length(RBC_CME.var)] - RBC_CME.solution.perturbation.second_order.stochastic_steady_state
 
-    ggiirrff2 = girf(second_order_state_update, SSS_delta, zeros(T.nVars), false, T, draws = 1000,warmup_periods = 100)
+    ggiirrff2 = girf(second_order_state_update, SSS_delta, zeros(T.nVars), false, unspecified_initial_state, T, draws = 1000,warmup_periods = 100, algorithm = :second_order)
     @test isapprox(ggiirrff2[4,1,:],[-0.0003668849861768406
     0.0021711333455274096],rtol = 1e-3)
 
-    iirrff2 = irf(second_order_state_update, zeros(T.nVars), zeros(T.nVars), false, T)
+    iirrff2 = irf(second_order_state_update, zeros(T.nVars), zeros(T.nVars), false, unspecified_initial_state, T, algorithm = :second_order)
     @test isapprox(iirrff2[4,1,:],[-0.0004547347878067665, 0.0020831426377533636],rtol = 1e-6)
 
 
     SSS_delta = RBC_CME.solution.non_stochastic_steady_state[1:length(RBC_CME.var)] - RBC_CME.solution.perturbation.third_order.stochastic_steady_state
 
-    ggiirrff3 = girf(third_order_state_update, SSS_delta, zeros(T.nVars), false, T,draws = 1000,warmup_periods = 100)
+    ggiirrff3 = girf(third_order_state_update, SSS_delta, zeros(T.nVars), false, unspecified_initial_state, T,draws = 1000,warmup_periods = 100, algorithm = :third_order)
     @test isapprox(ggiirrff3[4,1,:],[ -0.00036686142588429404
     0.002171120660323429],rtol = 1e-3)
 
-    iirrff3 = irf(third_order_state_update, zeros(T.nVars), zeros(T.nVars), false, T)
+    iirrff3 = irf(third_order_state_update, zeros(T.nVars), zeros(T.nVars), false, unspecified_initial_state, T, algorithm = :third_order)
     @test isapprox(iirrff3[4,1,:],[-0.00045473149068020854, 0.002083198241302615], rtol = 1e-6)
+
+    SSS_delta = RBC_CME.solution.non_stochastic_steady_state[1:length(RBC_CME.var)] - RBC_CME.solution.perturbation.pruned_second_order.stochastic_steady_state
+
+    pruned_second_order_state_update = function(pruned_states::Vector{Vector{Float64}}, shock::Vector{Float64})
[…]
