### 2023-08-20 a2e715e
include pruning in comparison table
