+++ docs/src/index.md
@@ -10,7 +10,7 @@ As of now the package can:
 
 - parse a model written with user friendly syntax (variables are followed by time indices `...[2], [1], [0], [-1], [-2]...`, or `[x]` for shocks)
 - (tries to) solve the model only knowing the model equations and parameter values (no steady state file needed)
-- calculate first, second, and third order perturbation solutions using (forward or reverse-mode) automatic differentiation (AD)
+- calculate first, second, and third order (pruned) perturbation solutions using (forward or reverse-mode) automatic differentiation (AD)
 - calculate (generalised) impulse response functions, simulate the model, or do conditional forecasts
 - calibrate parameters using (non stochastic) steady state relationships
 - match model moments
@@ -54,6 +54,7 @@ The package contains the following models in the `models` folder:
 **Automatic declaration of variables and parameters**|yes|||||||||||||
 **Derivatives (Automatic Differentiation) wrt parameters**|yes|||||||||yes - for all 1st, 2nd order perturbation solution related output *if user supplied steady state equations*|||
 **Perturbation solution order**|1, 2, 3|k|1 to 5|1|1|1|1|1, 2, 3|1, 2, 3|1, 2|1||1 to 5|
+**Pruning**|yes|yes|tbc|||||tbc|tbc|tbc|||tbc|
 **Automatic derivation of first order conditions**|||||||||||yes||
 **Handles occasionally binding constraints**||yes|yes|||yes||yes|yes|||yes||
 **Global solution**||||||||yes|yes|||yes||
