### 2023-08-21 ad5b054
solve for pruned solutions as well
