### 2023-08-22 98b5708
add pruned 2nd order mean + deriv helpers
