### 2023-08-22 09c6f6d
add get_moments/get_mean funcs
