+++ test/models/FS2000.jl
@@ -1,4 +1,5 @@
 # Translated from: https://archives.dynare.org/documentation/examples.html
+# be aware that dynare dynamics differ if c[2] or P[2] (not sure which one) are not declared explicitly as an auxilliary variable (c_lead(0) = c(+1);). The system in dynare has one less variable and the higher order solutions are different for the stochastic vol term.
 
 @model m begin
     dA[0] = exp(gam + z_e_a  *  e_a[x])
@@ -29,9 +30,6 @@
 
     gp_obs[0] = (P[0] / P[-1]) * m[-1] / dA[0]
 
-    log_gy_obs[0] = log(gy_obs[0])
-
-    log_gp_obs[0] = log(gp_obs[0])
 end
 
 
