### 2023-08-22 36d079f
add func and complete mean calc
