### 2023-08-22 56bc0a4
got second order cov to work with diff
