+++ src/MacroModelling.jl
@@ -120,6 +120,135 @@ Base.show(io::IO, 𝓂::ℳ) = println(io,
 
 
 
+
+                # higher order solutions moment helper functions
+
+
+function bivariate_moment(moment::Vector{Int}, rho::Int)::Int
+    if (moment[1] + moment[2]) % 2 == 1
+        return 0
+    end
+
+    result = 1
+    coefficient = 1
+    odd_value = 2 * (moment[1] % 2)
+
+    for j = 1:min(moment[1] ÷ 2, moment[2] ÷ 2)
+        coefficient *= 2 * (moment[1] ÷ 2 + 1 - j) * (moment[2] ÷ 2 + 1 - j) * rho^2 / (j * (2 * j - 1 + odd_value))
+        result += coefficient
+    end
+
+    if odd_value == 2
+        result *= rho
+    end
+
+    result *= prod(1:2:moment[1]) * prod(1:2:moment[2])
+
+    return result
+end
+
+
+function product_moments(V, ii, nu)
+    s = sum(nu)
+
+    if s == 0
+        return 1
+    elseif isodd(s)
+        return 0
+    end
+
+    mask = .!(nu .== 0)
+    nu = nu[mask]
+    ii = ii[mask]
+    V = V[ii, ii]
+
+    m, s2 = length(ii), s / 2
+
+    if m == 1
+        return (V^s2 * prod(1:2:s-1))[1]
+    elseif m == 2
+        if V[1,1]==0 || V[2,2]==0
+            return 0
+        end
+        rho = V[1, 2] / sqrt(V[1, 1] * V[2, 2])
+        return (V[1, 1]^(nu[1] / 2) * V[2, 2]^(nu[2] / 2) * bivariate_moment(nu, Int(rho)))[1]
+    end
+
+    nu, inu = sort(nu, dims=2, rev=true)
+    V = V[inu, inu]
+
+    x = zeros(Int, 1, m)
+    V = V / 2
+    nu2 = nu / 2
+    p = 2
+    q = nu2' * V * nu2
+    y = 0
+
+    for _ in 1:round(Int, prod(nu .+ 1) / 2)
+        y += p * q^s2
+        for j in 1:m
+            if x[j] < nu[j]
+                x[j] += 1
+                p = -round(p * (nu[j] + 1 - x[j]) / x[j])
+                q -= 2 * (nu2 .- x) * V[:, j] + V[j, j]
+                break
+            else
+                x[j] = 0
+                p = isodd(nu[j]) ? -p : p
+                q += 2 * nu[j] * (nu2 .- x) * V[:, j] - nu[j]^2 * V[j, j]
+            end
+        end
+    end
+
+    return (y / prod(1:s2))[1]
+end
+
+
+function multiplicate(p::Int, order::Int)
+    # precompute p powers
+    pⁿ = [p^i for i in 0:order-1]
+
+    DP = spzeros(Bool, p^order, prod(p - 1 .+ (1:order)) ÷ factorial(order))
+
+    binom_p_ord = binomial(p + order - 1, order)
+
+    # Initialize index and binomial arrays
+    indexes = ones(Int, order)  # Vector to hold current indexes
+    binomials = zeros(Int, order)  # Vector to hold binomial values
+
+    # Helper function to handle the nested loops
+    function loop(level::Int)
+        for i=1:p
+            indexes[level] = i
+            binomials[level] = binomial(p + level - 1 - i, level)
+
+            if level < order  # If not at innermost loop yet, continue nesting
+                loop(level + 1)
[…]
