+++ versions.js
@@ -2,5 +2,5 @@ var DOC_VERSIONS = [
   "stable",
   "v0.1",
 ];
-var DOCUMENTER_NEWEST = "v0.1.26";
+var DOCUMENTER_NEWEST = "v0.1.27";
 var DOCUMENTER_STABLE = "stable";
