+++ src/MacroModelling.jl
@@ -2016,7 +2016,7 @@ second_order_stochastic_steady_state_iterative_solution = ID.ImplicitFunction(se
 function calculate_second_order_stochastic_steady_state(parameters::Vector{M}, 𝓂::ℳ; verbose::Bool = false, pruning::Bool = false) where M
     SS_and_pars, solution_error = 𝓂.SS_solve_func(parameters, 𝓂, verbose)
     
-    ∇₁ = calculate_jacobian(parameters, SS_and_pars, 𝓂)
+    ∇₁ = calculate_jacobian(parameters, SS_and_pars, 𝓂) |> Matrix
     
     𝐒₁, solved = calculate_first_order_solution(∇₁; T = 𝓂.timings)
     
@@ -2139,7 +2139,7 @@ end
 function calculate_third_order_stochastic_steady_state(parameters::Vector{M}, 𝓂::ℳ; verbose::Bool = false, pruning::Bool = false, tol::AbstractFloat = eps()) where M
     SS_and_pars, solution_error = 𝓂.SS_solve_func(parameters, 𝓂, verbose)
     
-    ∇₁ = calculate_jacobian(parameters, SS_and_pars, 𝓂)
+    ∇₁ = calculate_jacobian(parameters, SS_and_pars, 𝓂) |> Matrix
     
     𝐒₁, solved = calculate_first_order_solution(∇₁; T = 𝓂.timings)
     
@@ -2220,7 +2220,7 @@ function solve!(𝓂::ℳ;
 
             # @assert solution_error < eps() "Could not find non stochastic steady steady."
             
-            ∇₁ = calculate_jacobian(𝓂.parameter_values, SS_and_pars, 𝓂)
+            ∇₁ = calculate_jacobian(𝓂.parameter_values, SS_and_pars, 𝓂) |> Matrix
             
             sol_mat, solved = calculate_first_order_solution(∇₁; T = 𝓂.timings)
             
@@ -2329,7 +2329,7 @@ function solve!(𝓂::ℳ;
             
             SS_and_pars, solution_error = 𝓂.solution.outdated_NSSS ? 𝓂.SS_solve_func(𝓂.parameter_values, 𝓂, verbose) : (𝓂.solution.non_stochastic_steady_state, eps())
 
-            ∇₁ = calculate_jacobian(𝓂.parameter_values, SS_and_pars, 𝓂)
+            ∇₁ = calculate_jacobian(𝓂.parameter_values, SS_and_pars, 𝓂) |> Matrix
             
             sol_mat, converged = calculate_quadratic_iteration_solution(∇₁; T = 𝓂.timings)
             
@@ -2346,7 +2346,7 @@ function solve!(𝓂::ℳ;
[…]
