### 2023-08-28 f40e78f
add log obs back to FS200 for kalman test
