### 2023-08-29 thorek1/MacroModelling.jl#40 @gdalle
Solved, CI still fails on nightly at this time but no one cares
