### 2023-08-29 thorek1/MacroModelling.jl#36 @gdalle
That's a great addition!
Small typo in the second sentence: these _kindS_ of models
