### 2023-08-29 thorek1/MacroModelling.jl#49 @gdalle
Clarify code example in paper
I think it would be nice (although not essential) to add comments to the RBC code example, saying what the variables and parameters stand for.

A few typos in this section:
- the "non-stochastic steady state" is first mentioned on line 22, so the acronym should go there
- a space is missing after `σ_z`
- on line 25 it is written "non-stochastic steady steady" -> if the acronym is introduced earlier, you can just say NSSS

https://github.com/openjournals/joss-reviews/issues/5598
