### 2023-08-29 thorek1/MacroModelling.jl#38 @gdalle
> by that I mean that they solve all 14 models implemented with MacroModelling out of the box without any starting point guess

This is a good thing to include in the comparison if alternative packages cannot do that
