### 2023-08-29 thorek1/MacroModelling.jl#38 @gdalle
Also now the statement of need has become a bit redundant with the initial summary and the comparison, maybe you can pour it inside the other sections
