### 2023-08-29 thorek1/MacroModelling.jl#37 @gdalle
This is a good addition!

> they are not as general and convenience focused as the MATLAB packages and MacroModelling.jl

This particular claim could use a stronger / more precise justification
