### 2023-08-29 thorek1/MacroModelling.jl#33 @gdalle
Installation is working
