### 2023-08-29 #5598 @gdalle
Review finished on my end. There are a few sentences to add and things to tweak (see the issues I opened on the MacroModelling.jl repo), but after that, the paper and package will be good to go.
