### 2023-08-29 6b0b696
working version of efficient 3rd order cov
