+++ src/MacroModelling.jl
@@ -124,6 +124,90 @@ Base.show(io::IO, 𝓂::ℳ) = println(io,
                 # higher order solutions moment helper functions
 
 
+
+
+
+
+function warshall_algorithm!(R::SparseMatrixCSC{Bool,Int64})
+    # Size of the matrix
+    n, m = size(R)
+    
+    @assert n == m "Warshall algorithm only works for square matrices."
+
+    # The core idea of the Warshall algorithm is to consider each node (in this case, block)
+    # as an intermediate node and check if a path can be created between two nodes by using the
+    # intermediate node.
+    
+    # k is the intermediate node (or block).
+    for k in 1:n
+        # i is the starting node (or block).
+        for i in 1:n
+            # j is the ending node (or block).
+            for j in 1:n
+                # If there is a direct path from i to k AND a direct path from k to j, 
+                # then a path from i to j exists via k.
+                # Thus, set the value of R[i, j] to 1 (true).
+                R[i, j] = R[i, j] || (R[i, k] && R[k, j])
+            end
+        end
+    end
+    
+    # Return the transitive closure matrix.
+    return R
+end
+
+
+function determine_efficient_order(∇₁::SparseMatrixCSC{<: Real}, T::timings, observables::Union{Symbol,Vector{Symbol}}; verbose::Bool = false)
+    if observables == :full_covar
+        return [T.var => T.var]
+    elseif observables == :all
+        observables = T.var
+    end
+
+    expand = [  spdiagm(ones(T.nVars))[T.future_not_past_and_mixed_idx,:],
+                spdiagm(ones(T.nVars))[T.past_not_future_and_mixed_idx,:]] 
+    
+    ∇₊ = ∇₁[:,1:T.nFuture_not_past_and_mixed] * expand[1]
+    ∇₀ = ∇₁[:,T.nFuture_not_past_and_mixed .+ range(1,T.nVars)]
+    ∇₋ = ∇₁[:,T.nFuture_not_past_and_mixed + T.nVars .+ range(1,T.nPast_not_future_and_mixed)] * expand[2]
+
+    incidence = abs.(∇₊) + abs.(∇₀) + abs.(∇₋)
+
+    Q, P, R, nmatch, n_blocks = BlockTriangularForm.order(sparse(incidence))
+    R̂ = []
+    for i in 1:n_blocks
+        [push!(R̂, n_blocks - i + 1) for ii in R[i]:R[i+1] - 1]
+    end
+    push!(R̂,1)
+    
+    vars = hcat(P, R̂)'
+    eqs  = hcat(Q, R̂)'
+    
+    dependency_matrix = incidence[vars[1,:], eqs[1,:]] .!= 0
+    
+    warshall_algorithm!(dependency_matrix)
+
+    permut = sortperm(indexin(observables, T.var[eqs[1,:]]))
+    
+    solve_order = Vector{Symbol}[]
[…]
