+++ src/MacroModelling.jl
@@ -157,11 +157,13 @@ function warshall_algorithm!(R::SparseMatrixCSC{Bool,Int64})
 end
 
 
-function determine_efficient_order(∇₁::SparseMatrixCSC{<: Real}, T::timings, observables::Union{Symbol,Vector{Symbol}}; verbose::Bool = false)
-    if observables == :full_covar
+function determine_efficient_order(∇₁::SparseMatrixCSC{<: Real}, T::timings, 
+    variables::Union{Symbol_input,String_input}; verbose::Bool = false)
+    if variables == :full_covar
         return [T.var => T.var]
-    elseif observables == :all
-        observables = T.var
+    else
+        var_idx = parse_variables_input_to_index(variables, T)
+        observables = T.var[var_idx]
     end
 
     expand = [  spdiagm(ones(T.nVars))[T.future_not_past_and_mixed_idx,:],
@@ -3081,6 +3083,24 @@ end
 
 
 # helper for get functions
+function covariance_parameter_derivatives_third_order(parameters::Vector{ℱ.Dual{Z,S,N}}, variables::Union{Symbol_input,String_input}, parameters_idx, 𝓂::ℳ; verbose::Bool = false) where {Z,S,N}
+    params = copy(𝓂.parameter_values)
+    params = convert(Vector{ℱ.Dual{Z,S,N}},params)
+    params[parameters_idx] = parameters
+    convert(Vector{ℱ.Dual{Z,S,N}},max.(ℒ.diag(calculate_third_order_covariance(params, variables, 𝓂, verbose = verbose)[1]),eps(Float64)))
+end
+
+
+# helper for get functions
+function covariance_parameter_derivatives_third_order(parameters::ℱ.Dual{Z,S,N}, variables::Union{Symbol_input,String_input}, parameters_idx::Int, 𝓂::ℳ; verbose::Bool = false) where {Z,S,N}
+    params = copy(𝓂.parameter_values)
+    params = convert(Vector{ℱ.Dual{Z,S,N}},params)
+    params[parameters_idx] = parameters
+    convert(Vector{ℱ.Dual{Z,S,N}},max.(ℒ.diag(calculate_third_order_covariance(params, variables, 𝓂, verbose = verbose)[1]),eps(Float64)))
+end
+
+
+# helper for get functions
 function mean_parameter_derivatives(parameters::Vector{ℱ.Dual{Z,S,N}}, parameters_idx, 𝓂::ℳ; algorithm::Symbol = :pruned_second_order, verbose::Bool = false) where {Z,S,N}
     params = copy(𝓂.parameter_values)
     params = convert(Vector{ℱ.Dual{Z,S,N}},params)
@@ -4281,7 +4301,7 @@ end
 
 
 
-function calculate_second_order_covariance_forward(AC::AbstractVector{Float64}; dims::Vector{Tuple{Int,Int}})
+function solve_symmetric_sylvester_forward(AC::AbstractVector{Float64}; dims::Vector{Tuple{Int,Int}})
     lenA = dims[1][1] * dims[1][2]
 
     A = reshape(AC[1 : lenA],dims[1])
@@ -4305,7 +4325,7 @@ end
 
 
 
-function calculate_second_order_covariance_conditions(AC::AbstractVector{<: Real}, covar::AbstractMatrix{<: Real}, solved::Bool; dims::Vector{Tuple{Int,Int}})
[…]
