+++ src/get_functions.jl
@@ -1839,6 +1839,7 @@ Return the first and second moments of endogenous variables using the linearised
 - `standard_deviation` [Default: `true`, Type: `Bool`]: switch to return standard deviation of endogenous variables
 - `variance` [Default: `false`, Type: `Bool`]: switch to return variance of endogenous variables
 - `covariance` [Default: `false`, Type: `Bool`]: switch to return covariance matrix of endogenous variables
+- $VARIABLES
 - $DERIVATIVES
 - $PARAMETER_DERIVATIVES
 - $VERBOSE
@@ -1899,6 +1900,7 @@ function get_moments(𝓂::ℳ;
     standard_deviation::Bool = true, 
     variance::Bool = false, 
     covariance::Bool = false, 
+    variables::Union{Symbol_input,String_input} = :all_including_auxilliary, 
     derivatives::Bool = true,
     parameter_derivatives::Union{Symbol_input,String_input} = :all,
     algorithm::Symbol = :first_order,
@@ -1908,6 +1910,8 @@ function get_moments(𝓂::ℳ;
     solve!(𝓂, parameters = parameters, algorithm = algorithm, verbose = verbose, silent = silent)
 
     # write_parameters_input!(𝓂,parameters, verbose = verbose)
+
+    var_idx = parse_variables_input_to_index(variables, 𝓂.timings)
 
     parameter_derivatives = parameter_derivatives isa String_input ? parameter_derivatives .|> Meta.parse .|> replace_indices : parameter_derivatives
 
@@ -1958,7 +1962,7 @@ function get_moments(𝓂::ℳ;
 
     if derivatives
         if non_stochastic_steady_state
-            axis1 = [𝓂.var...,𝓂.calibration_equations_parameters...]
+            axis1 = [𝓂.var[var_idx]...,𝓂.calibration_equations_parameters...]
     
             if any(x -> contains(string(x), "◖"), axis1)
                 axis1_decomposed = decompose_name.(axis1)
@@ -1975,10 +1979,10 @@ function get_moments(𝓂::ℳ;
             dNSSS = ℱ.jacobian(x -> collect(SS_parameter_derivatives(x, param_idx, 𝓂, verbose = verbose)[1]), 𝓂.parameter_values[param_idx])
             
             # dNSSS = ℱ.jacobian(x->𝓂.SS_solve_func(x, 𝓂),𝓂.parameter_values)
-            SS =  KeyedArray(hcat(collect(NSSS),dNSSS);  Variables = axis1, Steady_state_and_∂steady_state∂parameter = axis2)
+            SS =  KeyedArray(hcat(collect(NSSS[var_idx]),dNSSS[var_idx,:]);  Variables = axis1, Steady_state_and_∂steady_state∂parameter = axis2)
         end
         
-        axis1 = 𝓂.var
+        axis1 = 𝓂.var[var_idx]
 
         if any(x -> contains(string(x), "◖"), axis1)
             axis1_decomposed = decompose_name.(axis1)
@@ -1999,7 +2003,15 @@ function get_moments(𝓂::ℳ;
                 dvariance = ℱ.jacobian(x -> covariance_parameter_derivatives_second_order(x, param_idx, 𝓂, verbose = verbose), 𝓂.parameter_values[param_idx])
 
                 if mean
-                    var_means = KeyedArray(state_μ;  Variables = axis1)
+                    var_means = KeyedArray(state_μ[var_idx];  Variables = axis1)
+                end
+            elseif algorithm == :pruned_third_order
+                covar_dcmp, state_μ = calculate_third_order_covariance(𝓂.parameter_values, variables, 𝓂, verbose = verbose)
+
[…]
