### 2023-08-31 1fe01be
reactivate precompiling
