### 2023-08-31 thorek1/MacroModelling.jl#49 @thorek1
changed the NSSS mess and added more details on the model (let me know if it is clear enough to the non-econ reader)
