### 2023-08-31 thorek1/MacroModelling.jl#38 @thorek1
added a few sentences to highlight why this feature is important, better than with other packages, and how the package achieves it
