### 2023-08-31 thorek1/MacroModelling.jl#37 @thorek1
added more details on convenience and recycled statement of need
