+++ paper.bib
@@ -27,3 +27,18 @@
   journal={arXiv preprint arXiv:1209.5145},
   year={2012}
 }
+
+@article{kydland_and_prescott_1982,
+ ISSN = {00129682, 14680262},
+ URL = {http://www.jstor.org/stable/1913386},
+ abstract = {The equilibrium growth model is modified and used to explain the cyclical variances of a set of economic time series, the covariances between real output and the other series, and the autocovariance of output. The model is fitted to quarterly data for the post-war U.S. economy. Crucial features of the model are the assumption that more than one time period is required for the construction of new productive capital, and the non-time-separable utility function that admits greater intertemporal substitution of leisure. The fit is surprisingly good in light of the model's simplicity and the small number of free parameters.},
+ author = {Finn E. Kydland and Edward C. Prescott},
+ journal = {Econometrica},
+ number = {6},
+ pages = {1345--1370},
+ publisher = {[Wiley, Econometric Society]},
+ title = {Time to Build and Aggregate Fluctuations},
+ urldate = {2023-08-31},
+ volume = {50},
+ year = {1982}
+}
