+++ .github/workflows/ci.yml
@@ -30,6 +30,9 @@ jobs:
           - version: '1.8'
             os: ubuntu-latest
             arch: x64
+          - version: '1.8'
+            os: windows-latest
+            arch: x64
           - version: '^1.10.0-0'
             os: ubuntu-latest
             arch: x64
