### 2023-08-31 f42bca5
fix NSSS indexing in get_moments
