### 2023-08-31 53fb31d
move indices out of jac hess creation
