### 2023-08-31 thorek1/MacroModelling.jl#49 @gdalle
There are a few LaTeX display issues with unicode characters, and some sentences are a bit long (could use a comma), but otherwise this is an improvement indeed :)
