### 2023-08-31 thorek1/MacroModelling.jl#38 @gdalle
perfect
