### 2023-08-31 thorek1/MacroModelling.jl#37 @gdalle
looks good to me
