### 2023-08-31 #5598 @jbytecode
👋👋👋 Dear @rickecon 👋👋👋

JOSS publishes articles about open source research software. The submission I'd like you to review is titled: "**MacroModelling.jl: A Julia package for developing and solving dynamic stochastic general equilibrium models**". You can find more information at the top of this Github issue (https://github.com/openjournals/joss-reviews/issues/5598).

We have already first reviewer's report, and we failed to get even a life signal from the second reviewer. I need someone else's help to get this manuscript reviewed.

Thank you in advance!
