### 2023-08-31 847250e
stylistic rewrite
