### 2023-08-31 thorek1/MacroModelling.jl#49 @thorek1
fixed unicode chars and had chatgpt help me with English
