### 2023-09-01 thorek1/MacroModelling.jl#49 @gdalle
All good! Gonna make a PR with the last few typos
