### 2023-09-01 #5598 @gdalle
Opened a PR for some minor formatting / syntax details, and then as far as I'm concerned this is good to go
