### 2023-09-02 9632823
fix compat issue with SymPy 0.1.1
