### 2023-09-02 #5598 @thorek1
> Opened a PR for some minor formatting / syntax details, and then as far as I'm concerned this is good to go

thanks for the suggestions. I merged the commit and fine-tuned them a bit more in the latest commits
