+++ src/MacroModelling.jl
@@ -121,17 +121,14 @@ Base.show(io::IO, 𝓂::ℳ) = println(io,
 
 
 
-                # higher order solutions moment helper functions
 
 
 
-
-
-function reconstruct_sparse_matrix(sp_vector::SparseVector{Float64, Int64}, dims::Tuple{Int64, Int64})
+function reconstruct_sparse_matrix(sp_vector::SparseVector{T, Int}, dims::Tuple{Int, Int}) where T
     # Function to reconstruct the matrix from the vector and dimensions
     
     # Create an empty sparse matrix with original dimensions
-    sp_reconstructed = spzeros(eltype(sp_vector), dims)
+    sp_reconstructed = spzeros(T, dims)
     
     # Fill in the non-zero values
     for (i, val) in zip(findnz(sp_vector)...)
@@ -142,7 +139,9 @@ function reconstruct_sparse_matrix(sp_vector::SparseVector{Float64, Int64}, dims
     
     return sp_reconstructed
 end
-                
+
+
+# higher order solutions moment helper functions
 
 function warshall_algorithm!(R::SparseMatrixCSC{Bool,Int64})
     # Size of the matrix
@@ -3665,6 +3664,8 @@ function calculate_second_order_solution(∇₁::AbstractMatrix{<: Real}, #first
     droptol!(C,tol)
 
     𝐒₂, solved = solve_sylvester_equation_forward([vec(B) ;vec(C) ;vec(X)], dims = [size(B) ;size(C) ;size(X)], tol = tol)
+    # 𝐒₂, solved = solve_symmetric_sylvester_AD([vec(B) ;vec(C) ;vec(X)], dims = [size(B) ;size(C) ;size(X)], sparse_output = true)
+    
 
     𝐒₂ *= M₂.𝐔₂
 
@@ -3755,6 +3756,7 @@ function calculate_third_order_solution(∇₁::AbstractMatrix{<: Real}, #first 
     droptol!(C,tol)
 
     𝐒₃, solved = solve_sylvester_equation_forward([vec(B) ;vec(C) ;vec(X)], dims = [size(B) ;size(C) ;size(X)], tol = tol)
+    # 𝐒₃, solved = solve_symmetric_sylvester_AD([vec(B) ;vec(C) ;vec(X)], dims = [size(B) ;size(C) ;size(X)], sparse_output = true)
     
     𝐒₃ *= M₃.𝐔₃
 
@@ -4168,68 +4170,68 @@ end
 
 
 
-function calculate_covariance_forward(𝑺₁::AbstractMatrix{Float64}; T::timings, subset_indices::Vector{Int64})
-    A = @views 𝑺₁[subset_indices,1:T.nPast_not_future_and_mixed] * ℒ.diagm(ones(length(subset_indices)))[indexin(T.past_not_future_and_mixed_idx,subset_indices),:]
-    C = @views 𝑺₁[subset_indices,T.nPast_not_future_and_mixed+1:end]
+# function calculate_covariance_forward(𝑺₁::AbstractMatrix{Float64}; T::timings, subset_indices::Vector{Int64})
+#     A = @views 𝑺₁[subset_indices,1:T.nPast_not_future_and_mixed] * ℒ.diagm(ones(length(subset_indices)))[indexin(T.past_not_future_and_mixed_idx,subset_indices),:]
+#     C = @views 𝑺₁[subset_indices,T.nPast_not_future_and_mixed+1:end]
     
-    CC = C * C'
+#     CC = C * C'
 
-    sylvester = LinearOperators.LinearOperator(Float64, length(CC), length(CC), false, false, 
[…]
