### 2023-09-04 884db43
fix method of moments (vars as orig intended)
