+++ src/MacroModelling.jl
@@ -3663,9 +3663,9 @@ function calculate_second_order_solution(∇₁::AbstractMatrix{<: Real}, #first
     C = (M₂.𝐔₂ * ℒ.kron(𝐒₁₋╱𝟏ₑ, 𝐒₁₋╱𝟏ₑ) + M₂.𝐔₂ * M₂.𝛔) * M₂.𝐂₂
     droptol!(C,tol)
 
-    𝐒₂, solved = solve_sylvester_equation_forward([vec(B) ;vec(C) ;vec(X)], dims = [size(B) ;size(C) ;size(X)], tol = tol)
+    # 𝐒₂, solved = solve_sylvester_equation_forward([vec(B) ;vec(C) ;vec(X)], dims = [size(B) ;size(C) ;size(X)], tol = tol)
     # 𝐒₂, solved = solve_symmetric_sylvester_AD([vec(B) ;vec(C) ;vec(X)], dims = [size(B) ;size(C) ;size(X)], sparse_output = true)
-    
+    𝐒₂, solved = solve_symmetric_sylvester_forward([vec(B) ;vec(C) ;vec(X)], dims = [size(B) ;size(C) ;size(X)], sparse_output = true)
 
     𝐒₂ *= M₂.𝐔₂
 
@@ -3755,8 +3755,9 @@ function calculate_third_order_solution(∇₁::AbstractMatrix{<: Real}, #first 
     C *= M₃.𝐂₃
     droptol!(C,tol)
 
-    𝐒₃, solved = solve_sylvester_equation_forward([vec(B) ;vec(C) ;vec(X)], dims = [size(B) ;size(C) ;size(X)], tol = tol)
+    # 𝐒₃, solved = solve_sylvester_equation_forward([vec(B) ;vec(C) ;vec(X)], dims = [size(B) ;size(C) ;size(X)], tol = tol)
     # 𝐒₃, solved = solve_symmetric_sylvester_AD([vec(B) ;vec(C) ;vec(X)], dims = [size(B) ;size(C) ;size(X)], sparse_output = true)
+    𝐒₃, solved = solve_symmetric_sylvester_forward([vec(B) ;vec(C) ;vec(X)], dims = [size(B) ;size(C) ;size(X)], sparse_output = true)
     
     𝐒₃ *= M₃.𝐔₃
 
@@ -4250,7 +4251,8 @@ function calculate_covariance(parameters::Vector{<: Real}, 𝓂::ℳ; verbose::B
     CC = C * C'
 
     covar_raw, _ = solve_symmetric_sylvester_AD_direct([vec(A); vec(-CC)], dims = [size(A), size(CC)])
-
+    # covar_raw, _ = solve_symmetric_sylvester_forward([vec(A); vec(-CC)], dims = [size(A), size(CC)])
+    
     return covar_raw, sol , ∇₁, SS_and_pars
 end
 
@@ -4448,6 +4450,113 @@ function solve_symmetric_sylvester_forward(ABC::AbstractVector{ℱ.Dual{Z,S,N}};
     end,size(val)), solved
 end
 
+
+
+function solve_symmetric_sylvester_forward(abc::SparseVector{ℱ.Dual{Z,S,N}}; dims::Vector{Tuple{Int,Int}}, sparse_output::Bool = false) where {Z,S,N}
+
+    # unpack: AoS -> SoA
+    ABC, partials = separate_values_and_partials_from_sparsevec_dual(abc)
+
+    # get f(vs)
+    val, solved = solve_symmetric_sylvester_forward(ABC, dims = dims, sparse_output = sparse_output)
+
+    lenA = dims[1][1] * dims[1][2]
+
+    A = reconstruct_sparse_matrix(ABC[1 : lenA], dims[1])
[…]
