+++ src/MacroModelling.jl
@@ -4250,7 +4250,7 @@ function calculate_covariance(parameters::Vector{<: Real}, 𝓂::ℳ; verbose::B
     
     CC = C * C'
 
-    covar_raw, _ = solve_symmetric_sylvester_AD_direct([vec(A); vec(-CC)], dims = [size(A), size(CC)])
+    covar_raw, _ = solve_symmetric_sylvester_AD_direct([vec(A); vec(-CC)], dims = [size(A), size(CC)], solver = :bicgstab)
     # covar_raw, _ = solve_symmetric_sylvester_forward([vec(A); vec(-CC)], dims = [size(A), size(CC)])
     
     return covar_raw, sol , ∇₁, SS_and_pars
@@ -4326,7 +4326,11 @@ end
 
 
 
-function solve_symmetric_sylvester_forward(ABC::SparseVector{Float64, Int64}; dims::Vector{Tuple{Int,Int}}, sparse_output::Bool = false)
+function solve_symmetric_sylvester_forward(ABC::SparseVector{Float64, Int64}; 
+    dims::Vector{Tuple{Int,Int}}, 
+    sparse_output::Bool = false,
+    solver::Symbol = :gmres)
+
     lenA = dims[1][1] * dims[1][2]
 
     A = reconstruct_sparse_matrix(ABC[1 : lenA], dims[1])
@@ -4340,17 +4344,30 @@ function solve_symmetric_sylvester_forward(ABC::SparseVector{Float64, Int64}; di
         C = reconstruct_sparse_matrix(ABC[lenA + 1 : end], dims[2])
     end
 
-    sylvester = LinearOperators.LinearOperator(Float64, length(C), length(C), true, true, 
-    (sol,𝐱) -> begin 
-        𝐗 = reshape(𝐱, size(C))
-        sol .= vec(A * 𝐗 * B - 𝐗)
-        return sol
-    end)
+    if solver ∈ [:gmres, :bicgstab]   
+        sylvester = LinearOperators.LinearOperator(Float64, length(C), length(C), true, true, 
+        (sol,𝐱) -> begin 
+            𝐗 = reshape(𝐱, size(C))
+            sol .= vec(A * 𝐗 * B - 𝐗)
+            return sol
+        end)
 
-    𝐂, info = Krylov.gmres(sylvester, [vec(C);])
+        if solver == :gmres
+            𝐂, info = Krylov.gmres(sylvester, [vec(C);])
+        elseif solver == :bicgstab
+            𝐂, info = Krylov.bicgstab(sylvester, [vec(C);])
+        end
 
-    if !info.solved
-        𝐂, info = Krylov.bicgstab(sylvester, [vec(C);])
+    elseif solver == :speedmapping
+        soll = speedmapping(collect(-C); m! = (X, x) -> X .= A * x * B - C, stabilize = true)
+
+        𝐂 = soll.minimizer
+
+        info = soll.converged
+    end
+
+    if !info.solved && !(solver == :gmres)
+        𝐂, info = Krylov.gmres(sylvester, [vec(C);])
     end
 
     return sparse_output ? sparse(reshape(𝐂, size(C))) : reshape(𝐂, size(C)), info.solved # return info on convergence
@@ -4370,17 +4387,30 @@ function solve_symmetric_sylvester_forward(ABC::Vector{Float64}; dims::Vector{Tu
         C = reshape(ABC[lenA + 1 : end], dims[2])
     end
 
-    sylvester = LinearOperators.LinearOperator(Float64, length(C), length(C), true, true, 
-    (sol,𝐱) -> begin 
-        𝐗 = reshape(𝐱, size(C))
-        sol .= vec(A * 𝐗 * B - 𝐗)
-        return sol
-    end)
[…]
