+++ src/MacroModelling.jl
@@ -5032,7 +5032,7 @@ function calculate_kalman_filter_loglikelihood(𝓂::ℳ, data::AbstractArray{Fl
     # Gaussian Prior
     CC = sol[observables_and_states, 𝓂.timings.nPast_not_future_and_mixed+1:end] * sol[observables_and_states, 𝓂.timings.nPast_not_future_and_mixed+1:end]'
 
-    P, _ = solve_symmetric_sylvester_AD_direct([vec(A); vec(-CC)], dims = [size(A), size(CC)])
+    P, _ = solve_symmetric_sylvester_AD_direct([vec(A); vec(-CC)], dims = [size(A), size(CC)], solver = :bicgstab)
     # P, _ = solve_symmetric_sylvester_forward([vec(A); vec(-CC)], dims = [size(A), size(CC)])
     # P, _ = calculate_covariance_AD(sol, T = 𝓂.timings, subset_indices = Int64[observables_and_states...])
 
