### 2023-09-05 8d62e02
streamline kalman and more views
