+++ src/MacroModelling.jl
@@ -3499,44 +3499,6 @@ function calculate_first_order_solution(∇₁::Matrix{S}; T::timings, explosive
 end
 
 
-function solve_sylvester_equation_conditions(concat_sparse_vec::SparseVector, S, solved; dims::Vector{Tuple{Int,Int}}, tol::AbstractFloat = eps())
-    lenA = dims[1][1] * dims[1][2]
-    lenB = dims[2][1] * dims[2][2]
-
-    A = sparse(reshape(concat_sparse_vec[1 : lenA], dims[1]))
-    B = sparse(reshape(concat_sparse_vec[lenA .+ (1 : lenB)], dims[2]))
-    X = sparse(reshape(concat_sparse_vec[lenA + lenB + 1 : end], dims[3]))
-
-    X + S - A * S * B
-end
-
-function solve_sylvester_equation_forward(concat_sparse_vec::SparseVector{Float64}; dims::Vector{Tuple{Int,Int}}, tol::AbstractFloat = eps())
-    lenA = dims[1][1] * dims[1][2]
-    lenB = dims[2][1] * dims[2][2]
-    lenX = dims[3][1] * dims[3][2]
-
-    A = sparse(reshape(concat_sparse_vec[1 : lenA],dims[1]))
-    B = sparse(reshape(concat_sparse_vec[lenA .+ (1 : lenB)],dims[2]))
-
-    sylvester = LinearOperators.LinearOperator(Float64, lenX, lenX, false, false, 
-        (sol,𝐱) -> begin 
-        𝐗 = reshape(𝐱, dims[3])
-        sol .= vec(𝐗 - A * 𝐗 * B)
-        return sol
-    end)
-
-    X, info = Krylov.gmres(sylvester, -concat_sparse_vec[lenA + lenB + 1 : end])#, atol = tol)
-
-    if !info.solved
-        X, info = Krylov.bicgstab(sylvester, -concat_sparse_vec[lenA + lenB + 1 : end])#, atol = tol)
-    end
-
-    𝐒₂ = sparse(reshape(X,dims[3]))
-    droptol!(𝐒₂, tol)
-
-    return 𝐒₂, info.solved
-end
-
 function separate_values_and_partials_from_sparsevec_dual(V::SparseVector{ℱ.Dual{Z,S,N}}; tol::AbstractFloat = eps()) where {Z,S,N}
     nrows = length(V)
     ncols = length(V.nzval[1].partials)
@@ -3561,56 +3523,6 @@ function separate_values_and_partials_from_sparsevec_dual(V::SparseVector{ℱ.Du
 
     return vvals, ps
 end
-
-function solve_sylvester_equation_forward(concat_sparse_vec::SparseVector{ℱ.Dual{Z,S,N}}; dims::Vector{Tuple{Int,Int}}, tol::AbstractFloat = eps()) where {Z,S,N}
-    # unpack: AoS -> SoA
-    values, partials = separate_values_and_partials_from_sparsevec_dual(concat_sparse_vec)
-
-    lenA = dims[1][1] * dims[1][2]
-    lenB = dims[2][1] * dims[2][2]
-
-    A = sparse(reshape(values[1 : lenA],dims[1]))
-    B = sparse(reshape(values[lenA .+ (1 : lenB)],dims[2]))
-
-    # get f(vs)
-    val, solved = solve_sylvester_equation_forward(values, dims = dims, tol = tol)
-
-    # get J(f, vs) * ps (cheating). Write your custom rule here. This used to be the conditions but here they are analytically derived.
-    A¹ = sparse((values[1 : lenA]).nzind, (values[1 : lenA]).nzind, 1, lenA, lenA)
[…]
