+++ test/test_standalone_function.jl
@@ -346,8 +346,12 @@ end
 
     # multiple parameter inputs and targets
     sol = Optim.optimize(x -> sum(abs2,get_statistics(RBC_CME, x, parameters = RBC_CME.parameters[[6,1]], standard_deviation = RBC_CME.var[[2,5]])[1] - [.0008,.21]),
-    [0,0], [1,1], [.006,.16], 
-    Optim.Fminbox(Optim.LBFGS(linesearch = LineSearches.BackTracking(order = 2)));
+    [0,0], [1,1], [.5,.16], 
+    Optim.Fminbox(Optim.LBFGS(linesearch = LineSearches.BackTracking(order = 3))),
+    # Optim.Options(show_trace = true,
+    # store_trace = true,
+    # extended_trace = true)
+    ;
     autodiff = :forward)
 
     @test isapprox(get_statistics(RBC_CME, sol.minimizer, parameters = RBC_CME.parameters[[6,1]], standard_deviation = RBC_CME.var[[2,5]])[1], [.0008,.21], atol=1e-6)
