### 2023-09-07 dd7a76e
fix get_statistics docs
