### 2023-09-07 212df3d
fix 3rd order pruned
