+++ test/figure_out_3rd.jl
@@ -1,0 +1,1239 @@
+using MacroModelling
+import MacroModelling: ℳ, calculate_covariance, multiplicate, generateSumVectors, product_moments, calculate_second_order_covariance_AD, determine_efficient_order, calculate_third_order_moments, calculate_second_order_moments
+import LinearAlgebra as ℒ
+include("models/FS2000.jl")
+
+corr(m,algorithm = :pruned_third_order)
+corr(m,algorithm = :pruned_second_order)
+corr(m)
+
+
+𝓂 = m
+parameter_values = m.parameter_values
+parameters = m.parameters
+algorithm = :pruned_third_order
+verbose = true
+
+function get_statistics(𝓂, 
+    parameter_values::Vector{T}; 
+    parameters::Vector{Symbol} = Symbol[], 
+    non_stochastic_steady_state::Vector{Symbol} = Symbol[],
+    mean::Vector{Symbol} = Symbol[],
+    standard_deviation::Vector{Symbol} = Symbol[],
+    variance::Vector{Symbol} = Symbol[],
+    covariance::Vector{Symbol} = Symbol[],
+    autocorrelation::Vector{Symbol} = Symbol[],
+    autocorrelation_periods::U = 1:5,
+    algorithm::Symbol = :first_order,
+    verbose::Bool = false) where {U,T}
+
+
+    @assert algorithm ∈ [:first_order,:linear_time_iteration,:quadratic_iteration,:pruned_second_order,:pruned_third_order] "Statistics can only be provided for first order perturbation or second and third order pruned perturbation solutions."
+
+    @assert !(non_stochastic_steady_state == Symbol[]) || !(standard_deviation == Symbol[]) || !(mean == Symbol[]) || !(variance == Symbol[]) || !(covariance == Symbol[]) || !(autocorrelation == Symbol[]) "Provide variables for at least one output."
+
+    SS_var_idx = indexin(non_stochastic_steady_state, 𝓂.var)
+
+    mean_var_idx = indexin(mean, 𝓂.var)
+
+    std_var_idx = indexin(standard_deviation, 𝓂.var)
+
+    var_var_idx = indexin(variance, 𝓂.var)
+
+    covar_var_idx = indexin(covariance, 𝓂.var)
+
+    autocorr_var_idx = indexin(autocorrelation, 𝓂.var)
+
+    other_parameter_values = 𝓂.parameter_values[indexin(setdiff(𝓂.parameters, parameters), 𝓂.parameters)]
+
+    sort_idx = sortperm(vcat(indexin(setdiff(𝓂.parameters, parameters), 𝓂.parameters), indexin(parameters, 𝓂.parameters)))
+
+    all_parameters = vcat(other_parameter_values, parameter_values)[sort_idx]
+
[…]
