### 2023-09-07 69e588f
optional cov calc for higher order
