+++ test/runtests.jl
@@ -25,7 +25,7 @@ GC.gc()
 
 @testset verbose = true "Static checking (JET.jl)" begin
     if VERSION >= v"1.9"
-        JET.test_package(MacroModelling; target_defined_modules=true, toplevel_logger=nothing)
+        JET.test_package(MacroModelling; target_defined_modules = true, toplevel_logger = nothing)
     end
 end
 
