+++ src/get_functions.jl
@@ -2468,7 +2468,7 @@ Wrapper for [`get_moments`](@ref) with `mean = true`, the default algorithm bein
 get_mean(args...; kwargs...) =  get_moments(args...; algorithm = :pruned_second_order, kwargs..., variance = false, non_stochastic_steady_state = false, standard_deviation = false, covariance = false, mean = true)[1]
 
 
-"""
-Wrapper for [`get_moments`](@ref) with `mean = true`, the default algorithm being `:pruned_second_order`, and `non_stochastic_steady_state = false, variance = false, standard_deviation = false, covariance = false`
-"""
-mean = get_mean
+# """
+# Wrapper for [`get_moments`](@ref) with `mean = true`, the default algorithm being `:pruned_second_order`, and `non_stochastic_steady_state = false, variance = false, standard_deviation = false, covariance = false`
+# """
+# mean = get_mean
