### 2023-09-08 9388422
calc pruned 3rd order autocorrelation
