### 2023-09-08 df4662b
higer order autocorr in get_stats
