+++ src/get_functions.jl
@@ -2379,20 +2379,18 @@ function get_statistics(𝓂,
     if algorithm == :pruned_third_order
 
         if !(autocorrelation == Symbol[])
-            autocorrelation = Symbol[]
-        end
 
-        if !(standard_deviation == Symbol[]) || !(variance == Symbol[]) || !(covariance == Symbol[])
+            covar_dcmp, state_μ, Σᶻ₃, Eᴸᶻ, autocorr_tmp, Σᶻ₂, Δμˢ₂, Σᶻ₁, μˢ₃δμˢ₁, ss_s, s_to_s₁, ŝ_to_ŝ₃, ŝ_to_y₃, ê_to_ŝ₃, ê_to_y₃, SS_and_pars = calculate_third_order_moments(all_parameters, :full_covar, 𝓂, verbose = verbose, autocorrelation = true)
+
+        elseif !(standard_deviation == Symbol[]) || !(variance == Symbol[]) || !(covariance == Symbol[])
+
             covar_dcmp, state_μ, SS_and_pars = calculate_third_order_moments(all_parameters, union(variance,covariance,standard_deviation), 𝓂, verbose = verbose)
+
         end
 
     elseif algorithm == :pruned_second_order
 
-        if !(autocorrelation == Symbol[])
-            autocorrelation = Symbol[]
-        end
-
-        if !(standard_deviation == Symbol[]) || !(variance == Symbol[]) || !(covariance == Symbol[])
+        if !(standard_deviation == Symbol[]) || !(variance == Symbol[]) || !(covariance == Symbol[]) || !(autocorrelation == Symbol[])
             covar_dcmp, Σᶻ₂, state_μ, Δμˢ₂, autocorr_tmp, ŝ_to_ŝ₂, ŝ_to_y₂, Σʸ₁, Σᶻ₁, SS_and_pars, 𝐒₁, ∇₁, 𝐒₂, ∇₂ = calculate_second_order_moments(all_parameters, 𝓂, verbose = verbose)
         else
             state_μ, Δμˢ₂, Σʸ₁, Σᶻ₁, SS_and_pars, 𝐒₁, ∇₁, 𝐒₂, ∇₂ = calculate_second_order_moments(all_parameters, 𝓂, verbose = verbose, covariance = false)
@@ -2410,9 +2408,50 @@ function get_statistics(𝓂,
             st_dev = sqrt.(varrs)
         end
     elseif !(autocorrelation == Symbol[])
-        A = @views sol[:,1:𝓂.timings.nPast_not_future_and_mixed] * ℒ.diagm(ones(𝓂.timings.nVars))[𝓂.timings.past_not_future_and_mixed_idx,:]
+        if algorithm == :pruned_third_order
+            autocorr = zeros(size(covar_dcmp,1),length(autocorrelation_periods))
 
-        autocorr = reduce(hcat,[ℒ.diag(A ^ i * covar_dcmp ./ ℒ.diag(covar_dcmp)) for i in autocorrelation_periods])
+            nˢ = size(s_to_s₁,1)
+            nᵉ = 𝓂.timings.nExo
+
[…]
