### 2023-09-10 2f00b43
3rd order autocorr
