### 2023-09-10 c010aec
type stability and overlaps in calc
