### 2023-09-10 4a3bf4f
test method of moments for pruned 2nd,3rd order
