### 2023-09-10 f6267ca
test pruned 2nd and 3rd order moment output
