+++ test/functionality_tests.jl
@@ -30,133 +30,98 @@ function functionality_test(m; algorithm = :first_order, plots = true, verbose =
     nsss = get_non_stochastic_steady_state(m)
     nsss = get_SS(m)
 
-    if algorithm == :first_order
+    if algorithm ∈ [:first_order, :pruned_second_order, :pruned_third_order]
+        auto_corr_nv = get_autocorrelation(m, algorithm = algorithm)
+        auto_corrr = get_autocorrelation(m, algorithm = algorithm, verbose = true)
+        new_auto_corr = get_autocorrelation(m, algorithm = algorithm, verbose = true, parameters = m.parameter_values * 1.0001)
+        new_auto_corr1 = get_autocorrelation(m, algorithm = algorithm, verbose = true, parameters = (m.parameters[1] => m.parameter_values[1] * 1.0001))
+        new_auto_corr2 = get_autocorrelation(m, algorithm = algorithm, verbose = true, parameters = Tuple(m.parameters[1:2] .=> m.parameter_values[1:2] * 1.0001))
+        new_auto_corr3 = get_autocorrelation(m, algorithm = algorithm, verbose = true, parameters = (m.parameters[1:2] .=> m.parameter_values[1:2] / 1.0001))
+        new_auto_corr3 = get_autocorr(m, algorithm = algorithm, verbose = true, parameters = (m.parameters[1:2] .=> m.parameter_values[1:2] / 1.0001))
+        new_auto_corr3 = autocorr(m, algorithm = algorithm, verbose = true, parameters = (m.parameters[1:2] .=> m.parameter_values[1:2] * 1.0002))
+        old_auto_corr = get_autocorrelation(m, algorithm = algorithm, verbose = true, parameters = old_par_vals)
 
-        auto_corr_nv = get_autocorrelation(m)
-        auto_corrr = get_autocorrelation(m, verbose = true)
-        new_auto_corr = get_autocorrelation(m, verbose = true, parameters = m.parameter_values * 1.0001)
-        new_auto_corr1 = get_autocorrelation(m, verbose = true, parameters = (m.parameters[1] => m.parameter_values[1] * 1.0001))
-        new_auto_corr2 = get_autocorrelation(m, verbose = true, parameters = Tuple(m.parameters[1:2] .=> m.parameter_values[1:2] * 1.0001))
-        new_auto_corr3 = get_autocorrelation(m, verbose = true, parameters = (m.parameters[1:2] .=> m.parameter_values[1:2] / 1.0001))
-        new_auto_corr3 = get_autocorr(m, verbose = true, parameters = (m.parameters[1:2] .=> m.parameter_values[1:2] / 1.0001))
-        new_auto_corr3 = autocorr(m, verbose = true, parameters = (m.parameters[1:2] .=> m.parameter_values[1:2] * 1.0002))
-        old_auto_corr = get_autocorrelation(m, verbose = true, parameters = old_par_vals)
+        new_auto_corr1 = get_autocorrelation(m, algorithm = algorithm, verbose = true, parameters = (string.(m.parameters[1]) => m.parameter_values[1] * 1.0001))
+        new_auto_corr2 = get_autocorrelation(m, algorithm = algorithm, verbose = true, parameters = Tuple(string.(m.parameters[1:2]) .=> m.parameter_values[1:2] * 1.0001))
+        new_auto_corr3 = get_autocorrelation(m, algorithm = algorithm, verbose = true, parameters = (string.(m.parameters[1:2]) .=> m.parameter_values[1:2] / 1.0001))
+        new_auto_corr3 = get_autocorr(m, algorithm = algorithm, verbose = true, parameters = (string.(m.parameters[1:2]) .=> m.parameter_values[1:2] / 1.0001))
[…]
