### 2023-09-10 1ac5b9f
test variables argument in get_moments
