### 2023-09-10 f6a125b
fix autocorr type
