+++ docs/src/tutorials/calibration.md
@@ -1,0 +1,354 @@
+# Calibration / method of moments - Gali (2015)
+
+This tutorial is intended to show the workflow to calibrate a model using method of moments. The tutorial is based on a standard model of monetary policy and will showcase the the use of gradient based optimisers and 2nd and 3rd order pruned solutions.
+
+## Define the model
+
+The first step is always to name the model and write down the equations. For the Gali (2015) model (chapter 3 of the book) this would go as follows:
+
+```@setup tutorial_2
+ENV["GKSwstype"] = "100"
+using Random
+Random.seed!(30)
+```
+
+```@repl tutorial_3
+using MacroModelling
+
+@model Gali_2015 begin
+	W_real[0] = C[0] ^ σ * N[0] ^ φ
+
+	Q[0] = β * (C[1] / C[0]) ^ (-σ) * Z[1] / Z[0] / Pi[1]
+
+	R[0] = 1 / Q[0]
+
+	Y[0] = A[0] * (N[0] / S[0]) ^ (1 - α)
+
+	R[0] = Pi[1] * realinterest[0]
+
+	R[0] = 1 / β * Pi[0] ^ ϕᵖⁱ * (Y[0] / Y[ss]) ^ ϕʸ * exp(nu[0])
+
+	C[0] = Y[0]
+
+	log(A[0]) = ρ_a * log(A[-1]) + std_a * eps_a[x]
+
+	log(Z[0]) = ρ_z * log(Z[-1]) - std_z * eps_z[x]
+
+	nu[0] = ρ_ν * nu[-1] + std_nu * eps_nu[x]
+
+	MC[0] = W_real[0] / (S[0] * Y[0] * (1 - α) / N[0])
+
+	1 = θ * Pi[0] ^ (ϵ - 1) + (1 - θ) * Pi_star[0] ^ (1 - ϵ)
+
+	S[0] = (1 - θ) * Pi_star[0] ^ (( - ϵ) / (1 - α)) + θ * Pi[0] ^ (ϵ / (1 - α)) * S[-1]
+
+	Pi_star[0] ^ (1 + ϵ * α / (1 - α)) = ϵ * x_aux_1[0] / x_aux_2[0] * (1 - τ) / (ϵ - 1)
+
+	x_aux_1[0] = MC[0] * Y[0] * Z[0] * C[0] ^ (-σ) + β * θ * Pi[1] ^ (ϵ + α * ϵ / (1 - α)) * x_aux_1[1]
+
+	x_aux_2[0] = Y[0] * Z[0] * C[0] ^ (-σ) + β * θ * Pi[1] ^ (ϵ - 1) * x_aux_2[1]
+
+	log_y[0] = log(Y[0])
+
+	log_W_real[0] = log(W_real[0])
+
+	log_N[0] = log(N[0])
+
+	pi_ann[0] = 4 * log(Pi[0])
+
+	i_ann[0] = 4 * log(R[0])
+
+	r_real_ann[0] = 4 * log(realinterest[0])
+
+	M_real[0] = Y[0] / R[0] ^ η
+
+end
+```
+
[…]
