+++ docs/src/unfinished_docs/todo.md
@@ -3,9 +3,10 @@
 ## High priority
 
 - [ ] implement occasionally binding constraints with shocks
-- [ ] write tests for variables argument in get_moment and for higher order moments
 - [ ] recheck get function examples and docs
-- [ ] autocorr to statistics output and in general for higher order pruned sols
+- [ ] write method of moments how to
+- [ ] autocorr and covariance with derivatives. return 3d array
+- [ ] Docs: document outputs and associated functions to work with function
 - [ ] redo naming in moments calc and make whole process faster (precalc wrangling matrices)
 - [ ] use ID for sparse output sylvester solvers (filed issue)
 - [ ] handle sparsity in sylvester solver better (hand over indices and nzvals instead of vec)
@@ -68,6 +69,8 @@
 - [ ] Find any SS by optimising over both SS guesses and parameter inputs
 - [ ] weed out SS solver and saved objects
 
+- [x] autocorr to statistics output and in general for higher order pruned sols
+- [x] write tests for variables argument in get_moment and for higher order moments
 - [x] handle KeyedArrays with strings as dimension names as input
 - [x] add mean in output funcs for higher order 
 - [x] recheck results for third order cov
