### 2023-09-10 33e3dd3
mean also for linear sol
