### 2023-09-10 0838aec
calc mean for linear
