### 2023-09-11 2a771ca
more aggressive GC in tests
