+++ test/runtests.jl
@@ -48,6 +48,9 @@ GC.gc()
     include("models/FS2000.jl")
     functionality_test(m, plots = false)
     for algorithm ∈ [:second_order,:pruned_second_order,:third_order,:pruned_third_order]
+        m = nothing
+        GC.gc()
+        include("models/FS2000.jl")
         functionality_test(m, algorithm = algorithm, plots = false)
     end
 end
