### 2023-09-11 1c39d76
refactor runtests spell out loops
