### 2023-09-11 #5598 @jmejia8
@thorek1  - Firstly, thank you and congrats about the package. The software seems useful to model and analyze certain economic problems.

Below are my initial comments regarding my review of the paper and codebase:

1. The installation proceeded without problems. I want to mention that pre-compilation stage took several minutes on my computer (which can be expected); however, importing the package also require some time, in respect to other packages. Could you explain why is this happening?
2. The automated tests seem covering most of the functionalities, but take a long time to complete.
3. I did not find information on the community guidelines for those willing to contribute.
4. The paper does not have the section 'Statement of need'.
5. The package is compared to related software in an informative manner. I suggest citing the papers related to the software (if possible), besides including the link to the software.

Due to I'm not an expert on the field, the following comments come from a [colleague](https://sites.google.com/site/edgarjscarrera/) of mine who is an economist:

> - Most business cycle research focuses on the behavior of the representative agent model and neglects the effects that are caused by the heterogeneity of agents. So how does the software package proposal address this issue?
> - Moreover, the software package proposal seems to work only for time constraints on a single/representative agent. Why not to consider the usual case of a macroeconomic model based on the stochastic Overlapping Generations (OLG) theory where households maximize discounted life-time utility with regard to their intertemporal consumption, capital and money demand, and labor supply !? Would it work?
