### 2023-09-11 3b120ef
fixed product moments
