### 2023-09-11 1745f72
no deepcopy of model
