### 2023-09-11 54dac9e
test case with >2 shocks and more GC
