### 2023-09-12 c4c7a52
fix pruned moments
