### 2023-09-12 e6ce814
fix calc mean
