### 2023-09-12 4a74632
fix calc mean again
