### 2023-09-12 79428a8
fix calc mean again
