### 2023-09-13 #5598 @thorek1
@jmejia8 thank you for your comments

1. the codebase is fairly large with 3800 lines (vs Plots.jl with 8000). furthermore for each model the equations are parsed (including for precompilation). all taken together it takes time and I tested different ways of trying to speed it up and the current amount of precompilation is the best answer i have so far.
2. they take a long time because all 16 currently implemented models are tested on most functionalities. a model is estimated (takes at least 20 minutes). this extesnive testing actually helps a lot in debugging and i am still far away from the 6h limit. its rather memory limits which have the tests fail from time to time.
3. i will add those
4. the paper used to have it (see previous paper compilations further up in the thread) but @gdalle and I agreed to drop it and have the content spread out in other sections of the paper. let me know if you prefer otherwise
5. I will add them where applicable

Regarding the comments from your colleague:
- heterogenous agent models are out of scope for now but on the todo list
- the package works both with deterministic and stochastic models, as well as finitely lived or infinitely lived agents. the user only needs to provide the first order optimality conditions and parameter values. assuming that a non-stochastic steady state exists the package can solve it.
