### 2023-09-17 f23efb8
speed gains (threading, sparsity/handover)
