+++ src/MacroModelling.jl
@@ -3,6 +3,7 @@ module MacroModelling
 
 import DocStringExtensions: FIELDS, SIGNATURES, TYPEDEF, TYPEDSIGNATURES, TYPEDFIELDS
 # import StatsFuns: normcdf
+import ThreadedSparseArrays
 using PrecompileTools
 import SpecialFunctions: erfcinv, erfc
 import SymPyPythonCall as SPyPyC
@@ -171,57 +172,120 @@ function warshall_algorithm!(R::SparseMatrixCSC{Bool,Int64})
     return R
 end
 
+function combine_pairs(v::Vector{Pair{Vector{Symbol}, Vector{Symbol}}})
+    i = 1
+    while i <= length(v)
+        subset_found = false
+        for j in i+1:length(v)
+            # Check if v[i].second is subset of v[j].second or vice versa
+            if all(elem -> elem in v[j].second, v[i].second) || all(elem -> elem in v[i].second, v[j].second)
+                # Combine the first elements and assign to the one with the larger second element
+                if length(v[i].second) > length(v[j].second)
+                    v[i] = v[i].first ∪ v[j].first => v[i].second
+                else
+                    v[j] = v[i].first ∪ v[j].first => v[j].second
+                end
+                # Remove the one with the smaller second element
+                deleteat!(v, length(v[i].second) > length(v[j].second) ? j : i)
+                subset_found = true
+                break
+            end
+        end
+        # If no subset was found for v[i], move to the next element
+        if !subset_found
+            i += 1
+        end
+    end
+    return v
+end
 
-function determine_efficient_order(∇₁::SparseMatrixCSC{<: Real}, 
-                                    T::timings, 
-                                    variables::Union{Symbol_input,String_input})
+function determine_efficient_order(𝐒₁::Matrix{<: Real}, 
+    T::timings, 
+    variables::Union{Symbol_input,String_input};
+    tol::AbstractFloat = eps())
 
+    orders = Pair{Vector{Symbol}, Vector{Symbol}}[]
+
+    nˢ = T.nPast_not_future_and_mixed
+    
     if variables == :full_covar
-        return [T.var => T.var]
+        return [T.var => T.past_not_future_and_mixed]
     else
-        var_idx = parse_variables_input_to_index(variables, T)
+        var_idx = MacroModelling.parse_variables_input_to_index(variables, T)
         observables = T.var[var_idx]
     end
 
-    expand = [  spdiagm(ones(T.nVars))[T.future_not_past_and_mixed_idx,:],
-                spdiagm(ones(T.nVars))[T.past_not_future_and_mixed_idx,:]] 
-    
-    ∇₊ = ∇₁[:,1:T.nFuture_not_past_and_mixed] * expand[1]
-    ∇₀ = ∇₁[:,T.nFuture_not_past_and_mixed .+ range(1,T.nVars)]
-    ∇₋ = ∇₁[:,T.nFuture_not_past_and_mixed + T.nVars .+ range(1,T.nPast_not_future_and_mixed)] * expand[2]
+    for obs in observables
+        obs_in_var_idx = indexin([obs],T.var)
+        dependencies_in_states = vec(sum(abs, 𝐒₁[obs_in_var_idx,1:nˢ], dims=1) .> tol) .> 0
 
-    incidence = abs.(∇₊) + abs.(∇₀) + abs.(∇₋)
-
[…]
