### 2023-09-18 e973237
rework sylvester; ForwDiff works
