+++ src/MacroModelling.jl
@@ -3660,7 +3660,7 @@ function calculate_second_order_solution(∇₁::AbstractMatrix{<: Real}, #first
     push!(dimensions,size(C))
     push!(dimensions,size(X))
 
-    𝐒₂, solved = solve_sylvester_equation_forward(values, coordinates, dimensions, solver = :gmres, sparse_output = true)
+    𝐒₂, solved = solve_sylvester_equation_forward(values, coords = coordinates, dims = dimensions, solver = :gmres, sparse_output = true)
     # 𝐒₂, solved = solve_sylvester_equation_forward([vec(B) ;vec(C) ;vec(X)], dims = [size(B) ;size(C) ;size(X)], tol = tol)
     # 𝐒₂, solved = solve_sylvester_equation_AD([vec(B) ;vec(C) ;vec(X)], dims = [size(B) ;size(C) ;size(X)], sparse_output = true)
     # 𝐒₂, solved = solve_sylvester_equation_forward([vec(B) ;vec(C) ;vec(X)], dims = [size(B) ;size(C) ;size(X)], sparse_output = true)
@@ -3770,7 +3770,7 @@ function calculate_third_order_solution(∇₁::AbstractMatrix{<: Real}, #first 
     push!(dimensions,size(X))
     
 
-    𝐒₃, solved = solve_sylvester_equation_forward(values, coordinates, dimensions, solver = :gmres, sparse_output = true)
+    𝐒₃, solved = solve_sylvester_equation_forward(values, coords = coordinates, dims = dimensions, solver = :gmres, sparse_output = true)
     # 𝐒₃, solved = solve_sylvester_equation_forward([vec(B) ;vec(C) ;vec(X)], dims = [size(B) ;size(C) ;size(X)], tol = tol)
     # 𝐒₃, solved = solve_sylvester_equation_AD([vec(B) ;vec(C) ;vec(X)], dims = [size(B) ;size(C) ;size(X)], sparse_output = true)
     # 𝐒₃, solved = solve_sylvester_equation_forward([vec(B) ;vec(C) ;vec(X)], dims = [size(B) ;size(C) ;size(X)], sparse_output = true)
@@ -4202,7 +4202,16 @@ function calculate_covariance(parameters::Vector{<: Real}, 𝓂::ℳ; verbose::B
     
     CC = C * C'
 
-    covar_raw, _ = solve_sylvester_equation_AD_direct([vec(A); vec(-CC)], dims = [size(A), size(CC)], solver = :bicgstab)
+    coordinates = Tuple{Vector{Int}, Vector{Int}}[]
+    
+    dimensions = Tuple{Int, Int}[]
+    push!(dimensions,size(A))
+    push!(dimensions,size(CC))
+    
+    values = vcat(vec(A), vec(collect(-CC)))
+
+    covar_raw, _ = solve_sylvester_equation_AD_direct(values, coords = coordinates, dims = dimensions, solver = :doubling)
+    # covar_raw, _ = solve_sylvester_equation_AD_direct([vec(A); vec(-CC)], dims = [size(A), size(CC)], solver = :bicgstab)
     # covar_raw, _ = solve_sylvester_equation_forward([vec(A); vec(-CC)], dims = [size(A), size(CC)])
     
     return covar_raw, sol , ∇₁, SS_and_pars
@@ -4283,77 +4292,77 @@ end
 
 
 
-function solve_sylvester_equation_forward(ABC::SparseVector{Float64, Int64}; 
-    dims::Vector{Tuple{Int,Int}}, 
-    sparse_output::Bool = false,
-    solver::Symbol = :gmres)
[…]
