### 2023-09-18 #5598 @jbytecode
@jmejia8 - Thank you for your suggestions. It seems our author has changed and corrected something and these changes are ready for your further review. Thank you in advance and sorry for bothering and being to much verbose.
