### 2023-09-18 d71f923
test pruned SSS,mean, and std
