### 2023-09-18 0e63a20
no dense array in higher order sol
