### 2023-09-19 d1e41a5
add other packages literature
