+++ paper.bib
@@ -42,3 +42,86 @@
  volume = {50},
  year = {1982}
 }
+
+@TechReport{dynare,
+  author      = {Adjemian, St\'ephane and Bastani, Houtan and
+                 Juillard, Michel and Karam\'e, Fr\'ederic and
+                 Mihoubi, Ferhat and Mutschler, Willi
+                 and Pfeifer, Johannes and Ratto, Marco and
+                 Rion, Normann and Villemot, S\'ebastien},
+  title       = {Dynare: Reference Manual Version 5},
+  year        = {2022},
+  institution = {CEPREMAP},
+  type        = {Dynare Working Papers},
+  number      = {72},
+}
+
+@TechReport{rise,
+  author={Junior Maih},
+  title={{Efficient perturbation methods for solving regime-switching DSGE models}},
+  year=2015,
+  month=Jan,
+  institution={Norges Bank},
+  type={Working Paper},
+  url={https://ideas.repec.org/p/bno/worpap/2015_01.html},
+  number={2015/01},
+  abstract={In an environment where economic structures break, variances change, distributions shift, conventional policies weaken and past events tend to reoccur, economic agents have to form expectations over different regimes. This makes the regime-switching dynamic stochastic general equilibrium (RS-DSGE) model the natural framework for analyzing the dynamics of macroeconomic variables. We present effcient solution methods for solving this class of models, allowing for the transition probabilities to be endogenous and for agents to react to anticipated events. The solution algorithms derived use a perturbation strategy which, unlike what has been proposed in the literature, does not rely on the partitioning of the switching parameters. These algorithms are all implemented in RISE, a exible object-oriented toolbox that can easily integrate alternative solution methods. We show that our algorithms replicate various examples found in the literature. Among those is a switching RBC model for which we present a third-order perturbation solution.},
+  keywords={DSGE; Markov switching; Sylvester equation; Newton algorithm; Pertubation; Matrix polynominal},
+  doi={},
+}
+
+@Article{taylorprojection,
+  author={Oren Levintal},
+  title={{Taylor Projection: A New Solution Method For Dynamic General Equilibrium Models}},
+  journal={International Economic Review},
+  year=2018,
+  volume={59},
+  number={3},
+  pages={1345-1373},
+  month={August},
+  keywords={},
+  doi={10.1111/iere.12306},
+  abstract={This article presents a new solution method for dynamic equilibrium models. The solution is approximated by polynomials that zero the residual function and its derivatives at a given point x0. The algorithm is essentially a type of projection but is significantly faster, since the problem is highly sparse and can be easily solved by a Newton solver. The obtained solution is accurate locally in the neighborhood of x0. Importantly, a local solution can be obtained at any point of the state space. This makes it possible to solve models at points that are further away from the steady state.},
+  url={https://ideas.repec.org/a/wly/iecrev/v59y2018i3p1345-1373.html}
+}
+
+@TechReport{gecon,
+  author={Klima, Grzegorz and Podemski, Karol and Retkiewicz-Wijtiwiak, Kaja and Sowińska, Anna E.},
+  title={{Smets-Wouters '03 model revisited - an implementation in gEcon}},
+  year=2015,
+  month=Feb,
+  institution={University Library of Munich, Germany},
+  type={MPRA Paper},
+  url={https://ideas.repec.org/p/pra/mprapa/64440.html},
+  number={64440},
[…]
