### 2023-09-19 #5598 @thorek1
I added community guidelines at the end of the readme and cited papers relating to mentioned packages in the paper if applicable.
