+++ paper.bib
@@ -63,7 +63,6 @@
   month=Jan,
   institution={Norges Bank},
   type={Working Paper},
-  url={https://ideas.repec.org/p/bno/worpap/2015_01.html},
   number={2015/01},
   abstract={In an environment where economic structures break, variances change, distributions shift, conventional policies weaken and past events tend to reoccur, economic agents have to form expectations over different regimes. This makes the regime-switching dynamic stochastic general equilibrium (RS-DSGE) model the natural framework for analyzing the dynamics of macroeconomic variables. We present effcient solution methods for solving this class of models, allowing for the transition probabilities to be endogenous and for agents to react to anticipated events. The solution algorithms derived use a perturbation strategy which, unlike what has been proposed in the literature, does not rely on the partitioning of the switching parameters. These algorithms are all implemented in RISE, a exible object-oriented toolbox that can easily integrate alternative solution methods. We show that our algorithms replicate various examples found in the literature. Among those is a switching RBC model for which we present a third-order perturbation solution.},
   keywords={DSGE; Markov switching; Sylvester equation; Newton algorithm; Pertubation; Matrix polynominal},
@@ -81,8 +80,7 @@
   month={August},
   keywords={},
   doi={10.1111/iere.12306},
-  abstract={This article presents a new solution method for dynamic equilibrium models. The solution is approximated by polynomials that zero the residual function and its derivatives at a given point x0. The algorithm is essentially a type of projection but is significantly faster, since the problem is highly sparse and can be easily solved by a Newton solver. The obtained solution is accurate locally in the neighborhood of x0. Importantly, a local solution can be obtained at any point of the state space. This makes it possible to solve models at points that are further away from the steady state.},
-  url={https://ideas.repec.org/a/wly/iecrev/v59y2018i3p1345-1373.html}
+  abstract={This article presents a new solution method for dynamic equilibrium models. The solution is approximated by polynomials that zero the residual function and its derivatives at a given point x0. The algorithm is essentially a type of projection but is significantly faster, since the problem is highly sparse and can be easily solved by a Newton solver. The obtained solution is accurate locally in the neighborhood of x0. Importantly, a local solution can be obtained at any point of the state space. This makes it possible to solve models at points that are further away from the steady state.}
 }
 
 @TechReport{gecon,
@@ -92,7 +90,6 @@
   month=Feb,
   institution={University Library of Munich, Germany},
   type={MPRA Paper},
-  url={https://ideas.repec.org/p/pra/mprapa/64440.html},
   number={64440},
   abstract={This paper presents an implementation of the well-known Smets-Wouters 2003 model for Euro Area using the gEcon package - what we call the ``third generation'' DSGE modelling toolbox. Our exercise serves three goals. First, we show how gEcon can be used to implement an important - from both applications and historical perspective - model. Second, through rigorous exposition enforced by the gEcon’s block-agent paradigm we analyse all the Smets-Wouters model’s building blocks. Last, but not least, the implementation presented here serves as a natural starting point for important from applications point of view extensions, like opening the economy, introducing non-lump-sum taxes, or adding sectors to the model economy. Full model implementation is attached.},
   keywords={DSGE; monetary policy; staggered prices; staggered wages},
@@ -119,7 +116,6 @@ abstract = {We introduce our GDSGE framework and a novel global solution method,
   month=Oct,
   institution={National Bureau of Economic Research, Inc},
   type={NBER Working Papers},
-  url={https://ideas.repec.org/p/nbr/nberwo/30573.html},
   number={30573},
[…]
