+++ CITATION.bib
@@ -2,7 +2,7 @@
 	author  = {Thore Kockerols},
 	title   = {MacroModelling.jl},
 	url     = {https://github.com/thorek1/MacroModelling.jl},
-	version = {v0.1.26},
+	version = {v0.1.28},
 	year    = {2023},
 	month   = {8}
 }
