### 2023-09-22 #5598 @jmejia8
@thorek1 - Thank you for the answers, all clear now. Below are my last comments on the paper:

1. Regarding "Statement of need", it is not clear for me who the target audience is (economists, students, scientist, practitioners?) after reading the paper.  It would be better to include this information explicitly, maybe in the first section.
2. It could be better to cite the peer-reviewed paper on Julia (https://doi.org/10.1137/141000671) instead of the technical report.

Although the package has performance issues, it seems useful and works for what it was designed to do, which is the most important IMO. Documentation has many examples, the API is detailed, and installation proceeds as it should. Therefore, I do not have further comments on the software.
