### 2023-09-22 #5598 @thorek1
I added the target audience to the paper and changed the reference to the peer-reviewed Julia paper.

Thank you @jmejia8 for your review.
