+++ src/MacroModelling.jl
@@ -122,7 +122,7 @@ Base.show(io::IO, 𝓂::ℳ) = println(io,
 
 
 
-function jacobian_wrt_values(A::AbstractSparseArray{T}, B::AbstractSparseArray{T}) where T
+function jacobian_wrt_values(A, B)
     # does this without creating dense arrays: reshape(permutedims(reshape(ℒ.I - ℒ.kron(A, B) ,size(B,1), size(A,1), size(A,1), size(B,1)), [2, 3, 4, 1]), size(A,1) * size(B,1), size(A,1) * size(B,1))
 
     # Compute the Kronecker product and subtract from identity
@@ -156,7 +156,7 @@ end
 
 
 
-function jacobian_wrt_A(A::AbstractSparseArray{T}, X::Matrix{T}) where T
+function jacobian_wrt_A(A, X)
     # does this without creating dense arrays: reshape(permutedims(reshape(ℒ.I - ℒ.kron(A, B) ,size(B,1), size(A,1), size(A,1), size(B,1)), [2, 3, 4, 1]), size(A,1) * size(B,1), size(A,1) * size(B,1))
 
     # Compute the Kronecker product and subtract from identity
@@ -849,6 +849,7 @@ function levenberg_marquardt(f::Function,
     transformation_level::S = 3,
     backtracking_order::S = 2,
     ) where {T <: AbstractFloat, S <: Integer}
+    # issues with optimization: https://www.gurobi.com/documentation/8.1/refman/numerics_gurobi_guidelines.html
 
     @assert size(lower_bounds) == size(upper_bounds) == size(initial_guess)
     @assert lower_bounds < upper_bounds
@@ -4385,9 +4386,9 @@ function solve_sylvester_equation_forward(ABC::Vector{Float64};
         vB = ABC[lengthA .+ (1:lengthB)]
         vC = ABC[lengthA + lengthB + 1:end]
 
-        A = sparse(coords[1]...,vA,dims[1]...)# |> ThreadedSparseArrays.ThreadedSparseMatrixCSC
-        B = sparse(coords[2]...,vB,dims[2]...)# |> ThreadedSparseArrays.ThreadedSparseMatrixCSC
-        C = sparse(coords[3]...,vC,dims[3]...)# |> ThreadedSparseArrays.ThreadedSparseMatrixCSC
+        A = sparse(coords[1]...,vA,dims[1]...) |> ThreadedSparseArrays.ThreadedSparseMatrixCSC
+        B = sparse(coords[2]...,vB,dims[2]...) |> ThreadedSparseArrays.ThreadedSparseMatrixCSC
+        C = sparse(coords[3]...,vC,dims[3]...) |> ThreadedSparseArrays.ThreadedSparseMatrixCSC
     else
         lengthA = dims[1][1] * dims[1][2]
         A = reshape(ABC[1:lengthA],dims[1]...)
@@ -4416,8 +4417,8 @@ function solve_sylvester_equation_forward(ABC::Vector{Float64};
     elseif solver == :iterative
         iter = 1
         change = 1
-        𝐂  = copy(C)
-        𝐂¹ = copy(C)
+        𝐂  = C
+        𝐂¹ = C
         while change > eps(Float32) && iter < 10000
             𝐂¹ = A * 𝐂 * B - C
             if !(A isa DenseMatrix)
@@ -4514,7 +4515,10 @@ function solve_sylvester_equation_forward(abc::Vector{ℱ.Dual{Z,S,N}};
     ABC = ℱ.value.(abc)
 
     # you can play with the dimension here, sometimes it makes sense to transpose
-    partial_values = mapreduce(ℱ.partials, hcat, abc)'
[…]
