### 2023-09-23 3a44e78
eliminate abs(NSSS) < 1e-12
