+++ docs/src/unfinished_docs/todo.md
@@ -70,7 +70,7 @@
 - [ ] Find any SS by optimising over both SS guesses and parameter inputs
 - [ ] weed out SS solver and saved objects
 
-- [x] set to 0 SS values < 1e-12
+- [x] set to 0 SS values < 1e-12 - doesnt work with Zygote
 - [x] sylvester with analytical derivatives (much faster if sparse) instead of implicit diff - yes but there are still way too large matrices being realised. implicitdiff is better here
 - [x] autocorr to statistics output and in general for higher order pruned sols
 - [x] fix product moments and test for cases with more than 2 shocks
