### 2023-09-23 f7f89bf
first go at how to guide
