+++ docs/src/tutorials/calibration.md
@@ -16,51 +16,51 @@ Random.seed!(30)
 using MacroModelling
 
 @model Gali_2015 begin
-	W_real[0] = C[0] ^ σ * N[0] ^ φ
+    W_real[0] = C[0] ^ σ * N[0] ^ φ
 
-	Q[0] = β * (C[1] / C[0]) ^ (-σ) * Z[1] / Z[0] / Pi[1]
+    Q[0] = β * (C[1] / C[0]) ^ (-σ) * Z[1] / Z[0] / Pi[1]
 
-	R[0] = 1 / Q[0]
+    R[0] = 1 / Q[0]
 
-	Y[0] = A[0] * (N[0] / S[0]) ^ (1 - α)
+    Y[0] = A[0] * (N[0] / S[0]) ^ (1 - α)
 
-	R[0] = Pi[1] * realinterest[0]
+    R[0] = Pi[1] * realinterest[0]
 
-	R[0] = 1 / β * Pi[0] ^ ϕᵖⁱ * (Y[0] / Y[ss]) ^ ϕʸ * exp(nu[0])
+    R[0] = 1 / β * Pi[0] ^ ϕᵖⁱ * (Y[0] / Y[ss]) ^ ϕʸ * exp(nu[0])
 
-	C[0] = Y[0]
+    C[0] = Y[0]
 
-	log(A[0]) = ρ_a * log(A[-1]) + std_a * eps_a[x]
+    log(A[0]) = ρ_a * log(A[-1]) + std_a * eps_a[x]
 
-	log(Z[0]) = ρ_z * log(Z[-1]) - std_z * eps_z[x]
+    log(Z[0]) = ρ_z * log(Z[-1]) - std_z * eps_z[x]
 
-	nu[0] = ρ_ν * nu[-1] + std_nu * eps_nu[x]
+    nu[0] = ρ_ν * nu[-1] + std_nu * eps_nu[x]
 
-	MC[0] = W_real[0] / (S[0] * Y[0] * (1 - α) / N[0])
+    MC[0] = W_real[0] / (S[0] * Y[0] * (1 - α) / N[0])
 
-	1 = θ * Pi[0] ^ (ϵ - 1) + (1 - θ) * Pi_star[0] ^ (1 - ϵ)
+    1 = θ * Pi[0] ^ (ϵ - 1) + (1 - θ) * Pi_star[0] ^ (1 - ϵ)
 
-	S[0] = (1 - θ) * Pi_star[0] ^ (( - ϵ) / (1 - α)) + θ * Pi[0] ^ (ϵ / (1 - α)) * S[-1]
+    S[0] = (1 - θ) * Pi_star[0] ^ (( - ϵ) / (1 - α)) + θ * Pi[0] ^ (ϵ / (1 - α)) * S[-1]
 
-	Pi_star[0] ^ (1 + ϵ * α / (1 - α)) = ϵ * x_aux_1[0] / x_aux_2[0] * (1 - τ) / (ϵ - 1)
+    Pi_star[0] ^ (1 + ϵ * α / (1 - α)) = ϵ * x_aux_1[0] / x_aux_2[0] * (1 - τ) / (ϵ - 1)
 
-	x_aux_1[0] = MC[0] * Y[0] * Z[0] * C[0] ^ (-σ) + β * θ * Pi[1] ^ (ϵ + α * ϵ / (1 - α)) * x_aux_1[1]
+    x_aux_1[0] = MC[0] * Y[0] * Z[0] * C[0] ^ (-σ) + β * θ * Pi[1] ^ (ϵ + α * ϵ / (1 - α)) * x_aux_1[1]
 
[…]
