### 2023-09-23 ccb9532
Delete test/figure_out_3rd.jl
