### 2023-09-23 c0794c9
Delete test/test_higher_moments.jl
