+++ test/test_higher_moments.jl
@@ -1,2554 +1,0 @@
-using MacroModelling
-import IterativeSolvers as ℐ
-using LinearAlgebra, LinearMaps
-
-function intersecting_elements(arr1::Union{Symbol,Vector}, arr2::Union{Symbol,Vector})
-    common = []
-    
-    if arr1 isa Symbol
-        arr1 = [arr1]
-    end
-    if arr2 isa Symbol
-        arr2 = [arr2]
-    else
-        arr2 = copy(arr2)
-    end
-
-    for i in arr1
-        if i in arr2
-            push!(common, i)
-            deleteat!(arr2, findfirst(==(i), arr2))  # remove the matched element from arr2
-        end
-    end
-    return common
-end
-
-function position_in_symmetric_matrix(position::Int, length::Int)
-    # Check if the vector length is a triangular number
-    n = round(Int, (-1 + sqrt(1 + 8*length)) / 2)
-
-    @assert n*(n+1)/2 == length "The length of the input vector is not valid to form a square symmetric matrix."
-
-    @assert position >= 1 && position <= length "Invalid position in the vector."
-
-    # Initialize the matrix position variables
-    row = 1
-    col = 1
-
-    # Iterate until we reach the desired position
-    for i in 1:length
-        if i == position
-            break
-        end
-        if col == n
-            row += 1
-            col = row
-        else
-            col += 1
-        end
-    end
-
-    # If it's a diagonal element, return one position
-    if row == col
-        return (row, col)
-    # If it's an off-diagonal element, return two positions
-    else
-        return (row, col), (col, row)
-    end
-end
-
-
-
-function position_in_full_vector(position::Int, length::Int)
-    # Check if the vector length is a triangular number
-    n = round(Int, (-1 + sqrt(1 + 8*length)) / 2)
-
-    @assert n*(n+1)/2 == length "The length of the input vector is not valid to form a square symmetric matrix."
-
-    @assert position >= 1 && position <= length "Invalid position in the vector."
-
-    # Initialize the matrix position variables
-    row = 1
-    col = 1
-
-    # Iterate until we reach the desired position
-    for i in 1:length
-        if i == position
-            break
-        end
-        if col == n
-            row += 1
-            col = row
-        else
-            col += 1
-        end
-    end
-
-    # Calculate the corresponding position(s) in the vector
-    vector_position = Int(n * (row - 1) + col)
-
-    if row == col
-        return vector_position
-    else
-        return vector_position, Int(n * (col - 1) + row)
-    end
-end
-
-# position_in_full_vector(6,6)
-
-# position_in_symmetric_matrix(2,6)
-
-# AA = rand(4,4)
[…]
