### 2023-09-23 0acebf6
get (calib) parameter names and values
