+++ .github/workflows/ci.yml
@@ -42,7 +42,7 @@ jobs:
             arch: x64
             allow_failure: true
     steps:
-      - uses: actions/checkout@v3
+      - uses: actions/checkout@v4
       #- uses: uraimo/run-on-arch-action@v2.5.0
       # - name: Set up QEMU
       #   if: matrix.arch == 'aarch64'
@@ -53,7 +53,7 @@ jobs:
       # - name: Build and push
       #   if: matrix.arch == 'aarch64'
       #   uses: docker/build-push-action@v3
-      - uses: julia-actions/setup-julia@latest
+      - uses: julia-actions/setup-julia@v1
         with:
           version: ${{ matrix.version }}
           arch: ${{ matrix.arch }}
@@ -71,11 +71,11 @@ jobs:
             ${{ runner.os }}-test-
             ${{ runner.os }}-
       # - run: pip3 install sympy
-      - uses: julia-actions/julia-buildpkg@latest
-      - uses: julia-actions/julia-runtest@latest
+      - uses: julia-actions/julia-buildpkg@v1
+      - uses: julia-actions/julia-runtest@v1
         with:
           prefix: ${{ matrix.prefix }}  # for `xvfb-run`
-      - uses: julia-actions/julia-processcoverage@latest
+      - uses: julia-actions/julia-processcoverage@v1
       - uses: codecov/codecov-action@v3
         with:
           token: ${{ secrets.CODECOV_TOKEN }}
