+++ .github/workflows/ci.yml
@@ -14,25 +14,19 @@ jobs:
       fail-fast: false
       matrix:
         version:
-          # - '1.8'
           - '1.9' #^1-0 and ^1.9-0 are not recognised
-          # - 'nightly' # fails for zygote
         os:
           - ubuntu-latest
-          - macOS-latest #slow
-          - windows-latest #slow
+          - macOS-latest
+          - windows-latest
         arch:
           - x64
-          # - aarch64 #not implemented in Github CI
         include:
           - os: ubuntu-latest
             prefix: xvfb-run
           - version: '1.8'
             os: ubuntu-latest
             arch: x64
-          #- version: '1.8'
-          #  os: windows-latest
-          #  arch: x64
           - version: '^1.10.0-0'
             os: ubuntu-latest
             arch: x64
@@ -43,16 +37,6 @@ jobs:
             allow_failure: true
     steps:
       - uses: actions/checkout@v4
-      #- uses: uraimo/run-on-arch-action@v2.5.0
-      # - name: Set up QEMU
-      #   if: matrix.arch == 'aarch64'
-      #   uses: docker/setup-qemu-action@v2
-      # - name: Set up Docker Buildx
-      #   if: matrix.arch == 'aarch64'
-      #   uses: docker/setup-buildx-action@v2
-      # - name: Build and push
-      #   if: matrix.arch == 'aarch64'
-      #   uses: docker/build-push-action@v3
       - uses: julia-actions/setup-julia@v1
         with:
           version: ${{ matrix.version }}
@@ -70,7 +54,6 @@ jobs:
             ${{ runner.os }}-test-${{ env.cache-name }}-
             ${{ runner.os }}-test-
             ${{ runner.os }}-
-      # - run: pip3 install sympy
       - uses: julia-actions/julia-buildpkg@v1
       - uses: julia-actions/julia-runtest@v1
         with:
