### 2023-09-24 1af4f08
test inspection function add. features
