+++ test/functionality_tests.jl
@@ -475,7 +475,9 @@ function functionality_test(m; algorithm = :first_order, plots = true, verbose =
     get_dynamic_equations(m) 
     get_calibration_equations(m) 
     get_parameters(m) 
+    get_parameters(m, values = true) 
     get_calibrated_parameters(m) 
+    get_calibrated_parameters(m, values = true) 
     get_parameters_in_equations(m) 
     get_parameters_defined_by_parameters(m) 
     get_parameters_defining_parameters(m) 
