+++ src/get_functions.jl
@@ -906,27 +906,42 @@ end
 """
 See [`get_irf`](@ref)
 """
-get_irfs = get_irf
+get_irfs(𝓂::ℳ; kwargs...) = get_irf(𝓂; kwargs...)
 
 """
 See [`get_irf`](@ref)
 """
-get_IRF = get_irf
+get_IRF(𝓂::ℳ; kwargs...) = get_irf(𝓂; kwargs...)
+
+"""
+See [`get_irf`](@ref)
+"""
+irfs(𝓂::ℳ; kwargs...) = get_irf(𝓂; kwargs...)
+
+"""
+See [`get_irf`](@ref)
+"""
+irf(𝓂::ℳ; kwargs...) = get_irf(𝓂; kwargs...)
+
+"""
+See [`get_irf`](@ref)
+"""
+IRF(𝓂::ℳ; kwargs...) = get_irf(𝓂; kwargs...)
 
 """
 Wrapper for [`get_irf`](@ref) with `shocks = :simulate`. Function returns values in levels by default.
 """
-simulate(args...; kwargs...) =  get_irf(args...; levels = true, kwargs..., shocks = :simulate)#[:,:,1]
+simulate(𝓂::ℳ; kwargs...) =  get_irf(𝓂; levels = true, kwargs..., shocks = :simulate)#[:,:,1]
 
 """
 Wrapper for [`get_irf`](@ref) with `shocks = :simulate`. Function returns values in levels by default.
 """
-get_simulation(args...; kwargs...) =  get_irf(args...; levels = true, kwargs..., shocks = :simulate)#[:,:,1]
+get_simulation(𝓂::ℳ; kwargs...) =  get_irf(𝓂; levels = true, kwargs..., shocks = :simulate)#[:,:,1]
 
 """
 Wrapper for [`get_irf`](@ref) with `shocks = :simulate`.
 """
-get_girf(args...; kwargs...) =  get_irf(args...; kwargs..., generalised_irf = true)
+get_girf(𝓂::ℳ; kwargs...) =  get_irf(𝓂; kwargs..., generalised_irf = true)
 
 
 
@@ -1130,47 +1145,58 @@ end
 """
 Wrapper for [`get_steady_state`](@ref) with `stochastic = false`.
 """
-get_non_stochastic_steady_state(args...; kwargs...) = get_steady_state(args...; kwargs..., stochastic = false)
+get_non_stochastic_steady_state(𝓂::ℳ; kwargs...) = get_steady_state(𝓂; kwargs..., stochastic = false)
 
 
 """
 Wrapper for [`get_steady_state`](@ref) with `stochastic = true`.
 """
-get_stochastic_steady_state(args...; kwargs...) = get_steady_state(args...; kwargs..., stochastic = true)
+get_stochastic_steady_state(𝓂::ℳ; kwargs...) = get_steady_state(𝓂; kwargs..., stochastic = true)
 
 
 """
 Wrapper for [`get_steady_state`](@ref) with `stochastic = true`.
 """
-get_SSS(args...; kwargs...) = get_steady_state(args...; kwargs..., stochastic = true)
+get_SSS(𝓂::ℳ; kwargs...) = get_steady_state(𝓂; kwargs..., stochastic = true)
 
 
 """
 Wrapper for [`get_steady_state`](@ref) with `stochastic = true`.
 """
-SSS(args...; kwargs...) = get_steady_state(args...; kwargs..., stochastic = true)
+SSS(𝓂::ℳ; kwargs...) = get_steady_state(𝓂; kwargs..., stochastic = true)
+
+
+"""
[…]
