### 2023-09-24 ae820e7
update bibliography
