+++ paper.bib
@@ -2,7 +2,8 @@
   title={Time Series Analysis by State Space Methods, 2nd edn},
   author={Durbin, J and Koopman, S. J.},
   year={2012},
-  publisher={Oxford University Press}
+  publisher={Oxford University Press},
+  doi={10.1093/acprof:oso/9780199641178.001.0001}
 }
 
 @article{julia,
@@ -14,23 +15,9 @@ number = {1},
 pages = {65-98},
 year = {2017},
 doi = {10.1137/141000671},
-
-URL = { 
-    
-        https://doi.org/10.1137/141000671
-    
-    
-
-},
-eprint = { 
-    
-        https://doi.org/10.1137/141000671
-    
-    
-
-}
-,
-    abstract = { Bridging cultures that have often been distant, Julia combines expertise from the diverse fields of computer science and computational science to create a new approach to numerical computing. Julia is designed to be easy and fast and questions notions generally held to be “laws of nature" by practitioners of numerical computing: \beginlist \item High-level dynamic programs have to be slow. \item One must prototype in one language and then rewrite in another language for speed or deployment. \item There are parts of a system appropriate for the programmer, and other parts that are best left untouched as they have been built by the experts. \endlist We introduce the Julia programming language and its design---a dance between specialization and abstraction. Specialization allows for custom treatment. Multiple dispatch, a technique from computer science, picks the right algorithm for the right circumstance. Abstraction, which is what good computation is really about, recognizes what remains the same after differences are stripped away. Abstractions in mathematics are captured as code through another technique from computer science, generic programming. Julia shows that one can achieve machine performance without sacrificing human convenience. }
+URL = {https://doi.org/10.1137/141000671},
+eprint = {https://doi.org/10.1137/141000671},
+abstract = { Bridging cultures that have often been distant, Julia combines expertise from the diverse fields of computer science and computational science to create a new approach to numerical computing. Julia is designed to be easy and fast and questions notions generally held to be “laws of nature" by practitioners of numerical computing: \beginlist \item High-level dynamic programs have to be slow. \item One must prototype in one language and then rewrite in another language for speed or deployment. \item There are parts of a system appropriate for the programmer, and other parts that are best left untouched as they have been built by the experts. \endlist We introduce the Julia programming language and its design---a dance between specialization and abstraction. Specialization allows for custom treatment. Multiple dispatch, a technique from computer science, picks the right algorithm for the right circumstance. Abstraction, which is what good computation is really about, recognizes what remains the same after differences are stripped away. Abstractions in mathematics are captured as code through another technique from computer science, generic programming. Julia shows that one can achieve machine performance without sacrificing human convenience. }
 }
 
 
@@ -42,7 +29,8 @@ eprint = {
   journal={Journal of Economic Dynamics and Control},
   volume={80},
   pages={1--16},
-  year={2017}
+  year={2017},
+  doi={10.2139/ssrn.2364989}
 }
 
 @techreport{villemot2011solving,
@@ -53,10 +41,15 @@ eprint = {
 }
 
 @article{Bezanson:2012,
-  title={Julia: A fast dynamic language for technical computing},
-  author={Bezanson, Jeff and Karpinski, Stefan and Shah, Viral B and Edelman, Alan},
-  journal={arXiv preprint arXiv:1209.5145},
-  year={2012}
+  title     = {Julia: A fresh approach to numerical computing},
+  doi       = {10.1137/141000671},
+  author    = {Bezanson, Jeff and Edelman, Alan and Karpinski, Stefan and Shah, Viral B},
+  journal   = {SIAM review},
[…]
