### 2023-09-24 33980d4
update bibliography
