+++ paper.bib
@@ -40,17 +40,6 @@ abstract = { Bridging cultures that have often been distant, Julia combines expe
   institution={Dynare Working Papers 2, CEPREMAP}
 }
 
-@article{Bezanson:2012,
-  title     = {Julia: A fresh approach to numerical computing},
-  doi       = {10.1137/141000671},
-  author    = {Bezanson, Jeff and Edelman, Alan and Karpinski, Stefan and Shah, Viral B},
-  journal   = {SIAM review},
-  volume    = {59},
-  number    = {1},
-  pages     = {65--98},
-  year      = {2017},
-  publisher = {SIAM}
-}
 
 @article{kydland_and_prescott_1982,
  ISSN = {00129682, 14680262},
