### 2023-09-24 #5598 @jbytecode
@thorek1 - I've just sent a pull request  (https://github.com/thorek1/MacroModelling.jl/pull/51) of some minor changes in the bibliography file. Please review the changes and apply. Thank you in advance.
