+++ src/get_functions.jl
@@ -906,42 +906,42 @@ end
 """
 See [`get_irf`](@ref)
 """
-get_irfs(𝓂::ℳ; kwargs...) = get_irf(𝓂; kwargs...)
+get_irfs = get_irf
 
 """
 See [`get_irf`](@ref)
 """
-get_IRF(𝓂::ℳ; kwargs...) = get_irf(𝓂; kwargs...)
+get_IRF = get_irf
 
 """
 See [`get_irf`](@ref)
 """
-irfs(𝓂::ℳ; kwargs...) = get_irf(𝓂; kwargs...)
+irfs = get_irf
 
 """
 See [`get_irf`](@ref)
 """
-irf(𝓂::ℳ; kwargs...) = get_irf(𝓂; kwargs...)
+irf = get_irf
 
 """
 See [`get_irf`](@ref)
 """
-IRF(𝓂::ℳ; kwargs...) = get_irf(𝓂; kwargs...)
+IRF = get_irf
 
 """
 Wrapper for [`get_irf`](@ref) with `shocks = :simulate`. Function returns values in levels by default.
 """
-simulate(𝓂::ℳ; kwargs...) =  get_irf(𝓂; levels = true, kwargs..., shocks = :simulate)#[:,:,1]
+simulate(args...; kwargs...) =  get_irf(args...; levels = true, kwargs..., shocks = :simulate)#[:,:,1]
 
 """
 Wrapper for [`get_irf`](@ref) with `shocks = :simulate`. Function returns values in levels by default.
 """
-get_simulation(𝓂::ℳ; kwargs...) =  get_irf(𝓂; levels = true, kwargs..., shocks = :simulate)#[:,:,1]
+get_simulation(args...; kwargs...) =  get_irf(args...; levels = true, kwargs..., shocks = :simulate)#[:,:,1]
 
 """
 Wrapper for [`get_irf`](@ref) with `shocks = :simulate`.
 """
-get_girf(𝓂::ℳ; kwargs...) =  get_irf(𝓂; kwargs..., generalised_irf = true)
+get_girf(args...; kwargs...) =  get_irf(args...; kwargs..., generalised_irf = true)
 
 
 
@@ -1145,58 +1145,58 @@ end
 """
 Wrapper for [`get_steady_state`](@ref) with `stochastic = false`.
 """
-get_non_stochastic_steady_state(𝓂::ℳ; kwargs...) = get_steady_state(𝓂; kwargs..., stochastic = false)
+get_non_stochastic_steady_state(args...; kwargs...) = get_steady_state(args...; kwargs..., stochastic = false)
 
 
 """
 Wrapper for [`get_steady_state`](@ref) with `stochastic = true`.
 """
-get_stochastic_steady_state(𝓂::ℳ; kwargs...) = get_steady_state(𝓂; kwargs..., stochastic = true)
+get_stochastic_steady_state(args...; kwargs...) = get_steady_state(args...; kwargs..., stochastic = true)
 
 
 """
 Wrapper for [`get_steady_state`](@ref) with `stochastic = true`.
 """
-get_SSS(𝓂::ℳ; kwargs...) = get_steady_state(𝓂; kwargs..., stochastic = true)
+get_SSS(args...; kwargs...) = get_steady_state(args...; kwargs..., stochastic = true)
 
 
 """
 Wrapper for [`get_steady_state`](@ref) with `stochastic = true`.
 """
-SSS(𝓂::ℳ; kwargs...) = get_steady_state(𝓂; kwargs..., stochastic = true)
+SSS(args...; kwargs...) = get_steady_state(args...; kwargs..., stochastic = true)
 
 
 """
[…]
