### 2023-09-24 e01b0ce
no direct solver for riccati
