### 2023-09-24 a1ce876
fix docs and plot_solution for pruning
