+++ src/plotting.jl
@@ -37,11 +37,12 @@ In case `shock_decomposition = true`, then the plot shows the variables, shocks,
 - `data_in_levels` [Default: `true`, Type: `Bool`]: indicator whether the data is provided in levels. If `true` the input to the data argument will have the non stochastic steady state substracted.
 - `shock_decomposition` [Default: `false`, Type: `Bool`]: whether to show the contribution of the shocks to the deviations from NSSS for each variable. If `false`, the plot shows the values of the selected variables, data, and shocks
 - `smooth` [Default: `true`, Type: `Bool`]: whether to return smoothed (`true`) or filtered (`false`) values for the variables, shocks, and decomposition.
-- `plots_per_page` [Default: `9`, Type: `Int`]: how many plots to show per page
-- `save_plots` [Default: `false`, Type: `Bool`]: switch to save plots using path and extension from `save_plots_path` and `save_plots_format`. Separate files per shocks and variables depending on number of variables and `plots_per_page`
-- `save_plots_path` [Default: `pwd()`, Type: `String`]: path where to save plots
-- `save_plots_format` [Default: `:pdf`, Type: `Symbol`]: output format of saved plots. See [input formats compatible with GR](https://docs.juliaplots.org/latest/output/#Supported-output-file-formats) for valid formats.
 - `show_plots` [Default: `true`, Type: `Bool`]: show plots. Separate plots per shocks and varibles depending on number of variables and `plots_per_page`.
+- `save_plots` [Default: `false`, Type: `Bool`]: switch to save plots using path and extension from `save_plots_path` and `save_plots_format`. Separate files per shocks and variables depending on number of variables and `plots_per_page`
+- `save_plots_format` [Default: `:pdf`, Type: `Symbol`]: output format of saved plots. See [input formats compatible with GR](https://docs.juliaplots.org/latest/output/#Supported-output-file-formats) for valid formats.
+- `save_plots_path` [Default: `pwd()`, Type: `String`]: path where to save plots
+- `plots_per_page` [Default: `9`, Type: `Int`]: how many plots to show per page
+- `transparency` [Default: `0.6`, Type: `Float64`]: transparency of bars
 - $VERBOSE
 
 # Examples
@@ -311,16 +312,16 @@ The left axis shows the level, and the right the deviation from the reference st
 # Arguments
 - $MODEL
 # Keyword Arguments
-- `plots_per_page` [Default: `9`, Type: `Int`]: how many plots to show per page
-- `save_plots` [Default: `false`, Type: `Bool`]: switch to save plots using path and extension from `save_plots_path` and `save_plots_format`. Separate files per shocks and variables depending on number of variables and `plots_per_page`
-- `save_plots_path` [Default: `pwd()`, Type: `String`]: path where to save plots
-- `save_plots_format` [Default: `:pdf`, Type: `Symbol`]: output format of saved plots. See [input formats compatible with GR](https://docs.juliaplots.org/latest/output/#Supported-output-file-formats) for valid formats.
-- `show_plots` [Default: `true`, Type: `Bool`]: show plots. Separate plots per shocks and varibles depending on number of variables and `plots_per_page`.
 - $PERIODS
-- $ALGORITHM
-- $PARAMETERS
-- $VARIABLES
 - $SHOCKS
+- $VARIABLES
+- $PARAMETERS
+- `show_plots` [Default: `true`, Type: `Bool`]: show plots. Separate plots per shocks and varibles depending on number of variables and `plots_per_page`.
[…]
