+++ paper.bib
@@ -6,6 +6,23 @@
   doi={10.1093/acprof:oso/9780199641178.001.0001}
 }
 
+@article{andreasen2018pruning,
+    author = {Andreasen, Martin M and Fernández-Villaverde, Jesús and Rubio-Ramírez, Juan F},
+    title = "{The Pruned State-Space System for Non-Linear DSGE Models: Theory and Empirical Applications}",
+    journal = {The Review of Economic Studies},
+    volume = {85},
+    number = {1},
+    pages = {1-49},
+    year = {2017},
+    month = {06},
+    abstract = "{This article studies the pruned state-space system for higher-order perturbation approximations to dynamic stochastic general equilibrium (DSGE) models. We show the stability of the pruned approximation up to third order and provide closed-form expressions for first and second unconditional moments and impulse response functions. Our results introduce generalized method of moments (GMM) estimation and impulse-response matching for DSGE models approximated up to third order and provide a foundation for indirect inference and simulated method of moments (SMM). As an application,we consider a New Keynesian model with Epstein–Zin preferences and two novel feedback effects from long-term bonds to the real economy, allowing us to match the level and variability of the \\$10\\$-year term premium in the U.S. with a low relative risk aversion of \\$5\\$.}",
+    issn = {0034-6527},
+    doi = {10.1093/restud/rdx037},
+    url = {https://doi.org/10.1093/restud/rdx037},
+    eprint = {https://academic.oup.com/restud/article-pdf/85/1/1/23033725/rdx037.pdf},
+}
+
+
 @article{julia,
 author = {Bezanson, Jeff and Edelman, Alan and Karpinski, Stefan and Shah, Viral B.},
 title = {Julia: A Fresh Approach to Numerical Computing},
