+++ paper.bib
@@ -47,7 +47,8 @@ abstract = { Bridging cultures that have often been distant, Julia combines expe
   volume={80},
   pages={1--16},
   year={2017},
-  doi={10.2139/ssrn.2364989}
+  doi={10.1016/j.jedc.2017.04.007},
+  URL={https://doi.org/10.1016/j.jedc.2017.04.007}
 }
 
 @techreport{villemot2011solving,
