### 2023-09-24 #5598 @thorek1
> @thorek1 - The JOSS documentation located in URL https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain states that your paper should include a `Statement of need` section that clearly illustrates the research purpose of the software and places it in the context of related work. However, the manuscript in its current form doesn't have a section labeled with that. Could you please correct this issue? (possibly a refactoring would solve this).

I resurrected the Statement of need section and merged your PR (many thanks).
