+++ versions.js
@@ -2,5 +2,5 @@ var DOC_VERSIONS = [
   "stable",
   "v0.1",
 ];
-var DOCUMENTER_NEWEST = "v0.1.28";
+var DOCUMENTER_NEWEST = "v0.1.29";
 var DOCUMENTER_STABLE = "stable";
