### 2023-09-25 f3a2fbc
remove dash between url and citations
