### 2023-09-25 #5598 @thorek1
- I've just sent a PR (https://github.com/thorek1/MacroModelling.jl/pull/52). This PR removes dash between URLs and citations. Please remove them if I've forgotten any. **merged**
- The title, author list, ORCIDs in Zenodo archive should match the info given in the manuscript. Please correct. **changed title to be the same**
- After merging the PR, please perform a full read and correct any typos if exist. **done**
- Report me the latest release tag and Zenodo URL. **https://zenodo.org/record/8374466**
