### 2023-09-25 #5598 @jbytecode
@thorek1 - In line 50 of the manuscript, the row is rendered as

`steady states symbolically, a feature shared only with [gEcon](http://gecon.r-forge.r-project.org`

Could you please correct and ping me again?
