### 2023-09-25 #5598 @jbytecode
@thorek1 - I failed to install and test the package. When I dive into the CI in the software repository, two of the tasks fail with the error message

`
ERROR: The following 2 direct dependencies failed to precompile:

MacroModelling [687ffad2-3618-405e-ac50-e0f7b9c75e44]

Failed to precompile MacroModelling [687ffad2-3618-405e-ac50-e0f7b9c75e44] to "/home/runner/.julia/compiled/v1.10/MacroModelling/jl_UcoYbb".
ERROR: LoadError: UndefVarError: `AdjOrTransDenseMatrix` not defined
`

which is similar to the error message that I got in my local Ubuntu desktop machine.

Could you please clarify this issue?
