### 2023-09-25 #5598 @thorek1
ThreadedSparseArrays.jl (a package I rely on) currently does not support the Julia 1.10 beta. I filed a [PR](https://github.com/jagot/ThreadedSparseArrays.jl/pull/11) to fix this. Until then only Julia 1.9 and 1.8 is supported.
