### 2023-09-25 #5598 @jbytecode
@thorek1 - thank you, that makes sense.

Could you please perform a full read and report me if that's okay in it is current form? (please don't be quick and be careful because once we accept the paper it will be permanent.)
