### 2023-09-25 #5598 @thorek1
I had a full read and indeed spotted some last issues. Fixed now. Ready to go from my side.
