### 2023-09-25 #5598 @jbytecode
@thorek1 - The software, manuscript and the bibliography look okay to me. I am now recommending an acceptance. Our track editor will make the final decision. Thank you for sending your work to JOSS.

@gdalle, @jmejia8 - Thank you so much for assisting in reviewing this paper and helping to me. We couldn't have accomplished this without you.
