### 2023-09-25 #5598 @editorialbot
:wave: @openjournals/sbcs-eics, this paper is ready to be accepted and published.
