### 2023-09-25 #5598 @jbytecode
@openjournals/sbcs-eics - The compiler emits some warning messages in https://github.com/openjournals/joss-reviews/issues/5598#issuecomment-1733941098 but the pdf is rendered properly.
